ISOAbstractCatalogue

ISOAbstractCatalogue

ISOAbstractCatalogue

Format

R6Class object.

Value

Object of R6Class for modelling an ISOAbstracCatalogue

References

ISO 19139:2007 Metadata - XML schema implementation

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOAbstractCatalogue

Public fields

name

name [1..1]: character

scope

scope [1..*]: character

fieldOfApplication

fieldOfApplication [0.*]: character

versionNumber

versionNumber [1..1]: character

versionDate

versionDate [1..1]: Date/Posix

language

language [0..1]: character

characterSet

character set [0..1]: character

locale

locale [0..*]: ISOLocale

subCatalogue

subCatalogue [0..*]: ISOAbstractCatalogue

Methods

Public methods

ISOAbstractCatalogue$new()

ISOAbstractCatalogue$setName()

ISOAbstractCatalogue$addScope()

ISOAbstractCatalogue$delScope()

ISOAbstractCatalogue$addFieldOfApplication()

ISOAbstractCatalogue$delFieldOfApplication()

ISOAbstractCatalogue$setVersionNumber()

ISOAbstractCatalogue$setVersionDate()

ISOAbstractCatalogue$setLanguage()

ISOAbstractCatalogue$setCharacterSet()

ISOAbstractCatalogue$addLocale()

ISOAbstractCatalogue$delLocale()

ISOAbstractCatalogue$addSubCatalogue()

ISOAbstractCatalogue$delSubCatalogue()

ISOAbstractCatalogue$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOAbstractCatalogue$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Sets the name. Locale names can be specified as

Usage

ISOAbstractCatalogue$setName(name, locales = NULL)

Arguments

name

name

locales

locales, object of class list

Method addScope()

Adds scope

Usage

ISOAbstractCatalogue$addScope(scope, locales = NULL)

Arguments

scope

scope

locales

locales, object of class list

Returns

TRUE if added, FALSE otherwise

Method delScope()

Deletes scope

Usage

ISOAbstractCatalogue$delScope(scope, locales = NULL)

Arguments

scope

scope

locales

locales, object of class list

Returns

TRUE if deleted, FALSE otherwise

Method addFieldOfApplication()

Adds field of application

Usage

ISOAbstractCatalogue$addFieldOfApplication(fieldOfApplication, locales = NULL)

Arguments

fieldOfApplication

field of application

locales

locales, object of class list

Returns

TRUE if added, FALSE otherwise

Method delFieldOfApplication()

Deletes field of application

Usage

ISOAbstractCatalogue$delFieldOfApplication(fieldOfApplication)

Arguments

fieldOfApplication

field of application

locales

locales, object of class list

Returns

TRUE if deleted, FALSE otherwise

Method setVersionNumber()

Set version number

Usage

ISOAbstractCatalogue$setVersionNumber(versionNumber)

Arguments

versionNumber

version number

Method setVersionDate()

Set version date

Usage

ISOAbstractCatalogue$setVersionDate(versionDate)

Arguments

versionDate

version date

Method setLanguage()

Set language

Usage

ISOAbstractCatalogue$setLanguage(locale)

Arguments

locale

object of class ISOLanguage or any character from values returned by ISOLanguages$values()

Method setCharacterSet()

Set charset

Usage

ISOAbstractCatalogue$setCharacterSet(charset)

Arguments

charset

object of class ISOCharacterSet or any character from values returned by ISOCharacterSet$values()

Method addLocale()

Adds locale

Usage

ISOAbstractCatalogue$addLocale(locale)

Arguments

locale

object of class ISOLocale

Returns

TRUE if added, FALSE otherwise

Method delLocale()

Deletes locale

Usage

ISOAbstractCatalogue$delLocale(locale)

Arguments

locale

object of class ISOLocale

Returns

TRUE if deleted, FALSE otherwise

Method addSubCatalogue()

Add sub catalogue

Usage

ISOAbstractCatalogue$addSubCatalogue(subCatalogue)

Arguments

subCatalogue

object of class ISOAbstractCatalogue

Returns

TRUE if added, FALSE otherwise

Method delSubCatalogue()

Deletes sub catalogue

Usage

ISOAbstractCatalogue$delSubCatalogue(subCatalogue)

Arguments

subCatalogue

object of class ISOAbstractCatalogue

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOAbstractCatalogue$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.