ISOAggregateInformation

ISOAggregateInformation

ISOAggregateInformation

Format

R6Class object.

Value

Object of R6Class for modelling a ISO AggregateInformation

References

ISO 19115:2003 - Geographic information – Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOAggregateInformation

Public fields

aggregateDataSetName

aggregate dataset name

aggregateDataSetIdentifier

aggregate dataset identifier

associationType

association type

initiativeType

initiative type

Methods

Public methods

ISOAggregateInformation$new()

ISOAggregateInformation$setAggregateDataSetName()

ISOAggregateInformation$setAggregateDataSetIdentifier()

ISOAggregateInformation$setAssociationType()

ISOAggregateInformation$setInitiativeType()

ISOAggregateInformation$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOAggregateInformation$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setAggregateDataSetName()

Set aggregate dataset name

Usage

ISOAggregateInformation$setAggregateDataSetName(datasetName)

Arguments

datasetName

object of class ISOCitation

Method setAggregateDataSetIdentifier()

Set aggregate dataset identifier

Usage

ISOAggregateInformation$setAggregateDataSetIdentifier(datasetIdentifier)

Arguments

datasetIdentifier

object of class ISOMetaIdentifier

Method setAssociationType()

Set association type

Usage

ISOAggregateInformation$setAssociationType(associationType)

Arguments

associationType

object of class ISOAssociationType or character value among values from ISOAssociationType$values()

Method setInitiativeType()

Set association type

Usage

ISOAggregateInformation$setInitiativeType(initiativeType)

Arguments

initiativeType

object of class ISOInitiativeType or character value among values from ISOInitiativeType$values()

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOAggregateInformation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  #encoding
  md <- ISOAggregateInformation$new()
  
  #adding a point of contact
  rp <- ISOResponsibleParty$new()
  rp$setIndividualName("someone")
  rp$setOrganisationName("somewhere")
  rp$setPositionName("someposition")
  rp$setRole("pointOfContact")
  contact <- ISOContact$new()
  phone <- ISOTelephone$new()
  phone$setVoice("myphonenumber")
  phone$setFacsimile("myfacsimile")
  contact$setPhone(phone)
  address <- ISOAddress$new()
  address$setDeliveryPoint("theaddress")
  address$setCity("thecity")
  address$setPostalCode("111")
  address$setCountry("France")
  address$setEmail("someone@theorg.org")
  contact$setAddress(address)
  res <- ISOOnlineResource$new()
  res$setLinkage("http://www.somewhereovertheweb.org")
  res$setName("somename")
  contact$setOnlineResource(res)
  rp$setContactInfo(contact)
  #citation
  ct <- ISOCitation$new()
  ct$setTitle("sometitle")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  ct$addDate(d)
  ct$setEdition("1.0")
  ct$setEditionDate(ISOdate(2015,1,1))
  ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
  ct$addPresentationForm("mapDigital")
#> [1] TRUE
  ct$addCitedResponsibleParty(rp)
#> [1] TRUE
  md$setAggregateDataSetName(ct)
  
  md$setAssociationType("source")
  md$setInitiativeType("investigation")
  
  xml <- md$encode()
#> NULL
#> <ISOCitation>
#> ....|-- title: sometitle
#> ....|-- date <ISODate>
#> ........|-- date: 2015-01-01T01:00:00
#> ........|-- dateType  [codeList=https://standards.iso.org/iso/19139/resources/gmxCodelists.xml#CI_DateTypeCode,codeListValue=publication,codeSpace=NA] <ISODateType>: publication {date identifies when the resource was issued}
#> ............|-- value: date identifies when the resource was issued
#> ....|-- edition: 1.0
#> ....|-- editionDate: 2015-01-01T12:00:00
#> ....|-- identifier <ISOMetaIdentifier>
#> ........|-- code: identifier
#> ....|-- citedResponsibleParty <ISOResponsibleParty>
#> ........|-- individualName: someone
#> ........|-- organisationName: somewhere
#> ........|-- positionName: someposition
#> ........|-- contactInfo <ISOContact>
#> ............|-- phone <ISOTelephone>
#> ................|-- voice: myphonenumber
#> ................|-- facsimile: myfacsimile
#> ............|-- address <ISOAddress>
#> ................|-- deliveryPoint: theaddress
#> ................|-- city: thecity
#> ................|-- postalCode: 111
#> ................|-- country: France
#> ................|-- electronicMailAddress: someone@theorg.org
#> ............|-- onlineResource <ISOOnlineResource>
#> ................|-- linkage <ISOURL>
#> ....................|-- value: http://www.somewhereovertheweb.org
#> ................|-- name: somename
#> ........|-- role  [codeList=https://standards.iso.org/iso/19139/resources/gmxCodelists.xml#CI_RoleCode,codeListValue=pointOfContact,codeSpace=NA] <ISORole>: pointOfContact {party who can be contacted for acquiring knowledge about or acquisition of the resource}
#> ............|-- value: party who can be contacted for acquiring knowledge about or acquisition of the resource
#> ....|-- presentationForm  [codeList=https://standards.iso.org/iso/19139/resources/gmxCodelists.xml#CI_PresentationFormCode,codeListValue=mapDigital,codeSpace=NA] <ISOPresentationForm>: mapDigital {map represented in raster or vector form}
#> ........|-- value: map represented in raster or vector form: mapDigital {}NULL
#> [1] "sometitle"
#> [1] "sometitle"
#> list()
#> [[1]]
#> <ISODate>
#> ....|-- date: 2015-01-01T01:00:00
#> ....|-- dateType  [codeList=https://standards.iso.org/iso/19139/resources/gmxCodelists.xml#CI_DateTypeCode,codeListValue=publication,codeSpace=NA] <ISODateType>: publication {date identifies when the resource was issued}
#> ........|-- value: date identifies when the resource was issued
#> NULL
#> [1] "2015-01-01 01:00:00 GMT"
#> [1] "2015-01-01 01:00:00 GMT"
#> <ISODateType>: publication {date identifies when the resource was issued}
#> ....|-- value: date identifies when the resource was issued[1] "date identifies when the resource was issued"
#> [1] "1.0"
#> [1] "1.0"
#> [1] "2015-01-01 12:00:00 GMT"
#> [1] "2015-01-01 12:00:00 GMT"
#> [[1]]
#> <ISOMetaIdentifier>
#> ....|-- code: identifier
#> NULL
#> NULL
#> [1] "identifier"
#> [1] "identifier"
#> NULL
#> NULL
#> NULL
#> [[1]]
#> <ISOResponsibleParty>
#> ....|-- individualName: someone
#> ....|-- organisationName: somewhere
#> ....|-- positionName: someposition
#> ....|-- contactInfo <ISOContact>
#> ........|-- phone <ISOTelephone>
#> ............|-- voice: myphonenumber
#> ............|-- facsimile: myfacsimile
#> ........|-- address <ISOAddress>
#> ............|-- deliveryPoint: theaddress
#> ............|-- city: thecity
#> ............|-- postalCode: 111
#> ............|-- country: France
#> ............|-- electronicMailAddress: someone@theorg.org
#> ........|-- onlineResource <ISOOnlineResource>
#> ............|-- linkage <ISOURL>
#> ................|-- value: http://www.somewhereovertheweb.org
#> ............|-- name: somename
#> ....|-- role  [codeList=https://standards.iso.org/iso/19139/resources/gmxCodelists.xml#CI_RoleCode,codeListValue=pointOfContact,codeSpace=NA] <ISORole>: pointOfContact {party who can be contacted for acquiring knowledge about or acquisition of the resource}
#> ........|-- value: party who can be contacted for acquiring knowledge about or acquisition of the resource
#> NULL
#> [1] "someone"
#> [1] "someone"
#> [1] "somewhere"
#> [1] "somewhere"
#> [1] "someposition"
#> [1] "someposition"
#> <ISOContact>
#> ....|-- phone <ISOTelephone>
#> ........|-- voice: myphonenumber
#> ........|-- facsimile: myfacsimile
#> ....|-- address <ISOAddress>
#> ........|-- deliveryPoint: theaddress
#> ........|-- city: thecity
#> ........|-- postalCode: 111
#> ........|-- country: France
#> ........|-- electronicMailAddress: someone@theorg.org
#> ....|-- onlineResource <ISOOnlineResource>
#> ........|-- linkage <ISOURL>
#> ............|-- value: http://www.somewhereovertheweb.org
#> ........|-- name: somenameNULL
#> <ISOTelephone>
#> ....|-- voice: myphonenumber
#> ....|-- facsimile: myfacsimileNULL
#> [1] "myphonenumber"
#> [1] "myphonenumber"
#> [1] "myfacsimile"
#> [1] "myfacsimile"
#> NULL
#> NULL
#> <ISOAddress>
#> ....|-- deliveryPoint: theaddress
#> ....|-- city: thecity
#> ....|-- postalCode: 111
#> ....|-- country: France
#> ....|-- electronicMailAddress: someone@theorg.orgNULL
#> [1] "theaddress"
#> [1] "theaddress"
#> [1] "thecity"
#> [1] "thecity"
#> [1] "111"
#> [1] "111"
#> [1] "France"
#> [1] "France"
#> [1] "someone@theorg.org"
#> [1] "someone@theorg.org"
#> <ISOOnlineResource>
#> ....|-- linkage <ISOURL>
#> ........|-- value: http://www.somewhereovertheweb.org
#> ....|-- name: somenameNULL
#> <ISOURL>
#> ....|-- value: http://www.somewhereovertheweb.org[1] "http://www.somewhereovertheweb.org"
#> NULL
#> NULL
#> [1] "somename"
#> [1] "somename"
#> NULL
#> NULL
#> NULL
#> list()
#> NULL
#> NULL
#> <ISORole>: pointOfContact {party who can be contacted for acquiring knowledge about or acquisition of the resource}
#> ....|-- value: party who can be contacted for acquiring knowledge about or acquisition of the resource[1] "party who can be contacted for acquiring knowledge about or acquisition of the resource"
#> [[1]]
#> <ISOPresentationForm>: mapDigital {map represented in raster or vector form}
#> ....|-- value: map represented in raster or vector form
#> [1] "map represented in raster or vector form"
#> NULL
#> NULL
#> NULL
#> NULL
#> NULL
#> list()
#> list()
#> NULL
#> <ISOAssociationType>: source {mapping and charting information from which the dataset content originates}
#> ....|-- value: mapping and charting information from which the dataset content originates[1] "mapping and charting information from which the dataset content originates"
#> <ISOInitiativeType>: investigation {search or systematic inquiry}
#> ....|-- value: search or systematic inquiry[1] "search or systematic inquiry"
#> [geometa][INFO] Object 'ISOAggregateInformation' is VALID according to ISO 19139 XML schemas!