ISOAssociationRole

ISOAssociationRole

ISOAssociationRole

Format

R6Class object.

Value

Object of R6Class for modelling an ISOAssociationRole

References

ISO 19110:2005 Methodology for Feature cataloguing

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOAbstractCarrierOfCharacteristics -> geometa::ISOAbstractPropertyType -> geometa::ISOPropertyType -> ISOAssociationRole

Public fields

type

type: ISORoleType

isOrdered

isOrdered: logical

isNavigable

isNavigable: logical

relation

relation: ISOAssociationRole

rolePlayer

rolePlayer: ISOFeatureType

Methods

Public methods

ISOAssociationRole$new()

ISOAssociationRole$setRoleType()

ISOAssociationRole$setIsOrdered()

ISOAssociationRole$setIsNavigable()

ISOAssociationRole$setRelation()

ISOAssociationRole$addRolePlayer()

ISOAssociationRole$delRolePlayer()

ISOAssociationRole$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

geometa::ISOAbstractCarrierOfCharacteristics$addConstraint()

geometa::ISOAbstractCarrierOfCharacteristics$delConstraint()

geometa::ISOAbstractCarrierOfCharacteristics$setFeatureType()

geometa::ISOAbstractPropertyType$setCardinality()

geometa::ISOAbstractPropertyType$setDefinition()

geometa::ISOAbstractPropertyType$setDefinitionReference()

geometa::ISOAbstractPropertyType$setFeatureCatalogue()

geometa::ISOAbstractPropertyType$setMemberName()

Method new()

Initializes object

Usage

ISOAssociationRole$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setRoleType()

Set role type

Usage

ISOAssociationRole$setRoleType(roleType)

Arguments

roleType

role type, object of class ISORoleType or any character among values returned by ISORoleType$values()

Method setIsOrdered()

Set is ordered

Usage

ISOAssociationRole$setIsOrdered(isOrdered)

Arguments

isOrdered

object of class logical

Method setIsNavigable()

Set is navigable

Usage

ISOAssociationRole$setIsNavigable(isNavigable)

Arguments

isNavigable

object of class logical

Method setRelation()

Set relation

Usage

ISOAssociationRole$setRelation(relation)

Arguments

relation

relation

Method addRolePlayer()

Adds role player

Usage

ISOAssociationRole$addRolePlayer(rolePlayer)

Arguments

rolePlayer

object of class ISOFeatureType

Returns

TRUE if added, FALSE otherwise

Method delRolePlayer()

Deletes role player

Usage

ISOAssociationRole$delRolePlayer(rolePlayer)

Arguments

rolePlayer

object of class ISOFeatureType

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOAssociationRole$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.