ISOCitation

ISOCitation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Citation

Fields

title

[character] title

alternateTitle

[list of character] alternateTitle

date

[ISODate] the citation date with date type

edition

[character] citation edition

editionDate

[Date|POSIXt] date or date/time of edition

identifier

[list of ISOMetaIdentifier] identifier

citedResponsibleParty

[list of ISOResponsibleParty] responsible party

presentationForm

[list of ISOPresentationForm] presentation form

series

[link{ISOCitationSeries}] series

otherCitationDetails

[character] other citation details

collectiveTitle

[character] collective title

ISBN

[character] ISBN

ISSN

[character] ISSN

Methods

new(xml)

This method is used to instantiate an ISOCitation

setTitle(title, locales)

Sets the title. Locale names can be specified as list with the locales argument.

setAlternateTitle(alternateTitle, locales)

Sets an alternate title. Locale names can be specified as list with the locales argument.

addAlternateTitle(alternateTitle, locales)

Adds an alternate title. Locale names can be specified as list with the locales argument.

delAlternateTitle(alternateTitle, locales)

Adds an alternate title. Locale names can be specified as list with the locales argument.

addDate(date)

Adds the date (ISODate object containing date and dateType)

setEdition(edition)

Sets the edition

setEditionDate(editionDate)

Sets the edition date, either an ISODate object containing date and dateType or a simple R date "POSIXct"/"POSIXt" object. For thesaurus citations, an ISODate should be used while for the general citation of ISODataIdentification, a simple R date should be used.

setIdentifier(identifier)

Sets the identifier as object of class ISOMetaIdentifier

addIdentifier(identifier)

Adds an identifier as object of class ISOMetaIdentifier

delIdentifier(identifier)

Deletes an identifier as object of class ISOMetaIdentifier

seCitedResponsibleParty(rp)

Sets the cited responsiblep party, object of class ISOResponsibleParty

setPresentationForm(presentationForm)

Sets the presentation form, object of class ISOPresentationForm

addPresentationForm(presentationForm)

Adds a presentation form, object of class ISOPresentationForm

delPresentationForm(presentationForm)

Deletes a presentation form, object of class ISOPresentationForm

setSeries(series)

Set series, object of class ISOCitationSeries

setOtherCitationDetails(otherCitationDetails, locales)

Set other citation details. Locale names can be specified as list with the locales argument.

setCollectiveTitle(collectiveTitle, locales)

Set collective title. Locale names can be specified as list with the locales argument.

setISBN(isbn)

Set the ISBN

setISSN(issn)

Set the ISSN

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

 #create ISOCitation
 md <- ISOCitation$new()
 md$setTitle("sometitle")
 md$setEdition("1.0")
 md$setEditionDate(ISOdate(2015,1,1))
 md$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
 md$addPresentationForm("mapDigital")
#> [1] TRUE
 
 #add a cited responsible party
 rp <- ISOResponsibleParty$new()
 rp$setIndividualName("someone")
 rp$setOrganisationName("somewhere")
 rp$setPositionName("someposition")
 rp$setRole("pointOfContact")
 contact <- ISOContact$new()
 phone <- ISOTelephone$new()
 phone$setVoice("myphonenumber")
 phone$setFacsimile("myfacsimile")
 contact$setPhone(phone)
 address <- ISOAddress$new()
 address$setDeliveryPoint("theaddress")
 address$setCity("thecity")
 address$setPostalCode("111")
 address$setCountry("France")
 address$setEmail("someone@theorg.org")
 contact$setAddress(address)
 res <- ISOOnlineResource$new()
 res$setLinkage("http://www.somewhereovertheweb.org")
 res$setName("somename")
 contact$setOnlineResource(res)
 rp$setContactInfo(contact)
 md$addCitedResponsibleParty(rp)
#> [1] TRUE
 xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmd}edition': This element is not expected. Expected is one of ( {http://www.isotc211.org/2005/gmd}alternateTitle, {http://www.isotc211.org/2005/gmd}date ) at line 5. 
#> [geometa][WARN] Object 'ISOCitation' is INVALID according to ISO 19139 XML schemas!