ISODataIdentification

ISODataIdentification

ISODataIdentification

Format

R6Class object.

Value

Object of R6Class for modelling an ISO DataIdentification

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOIdentification -> ISODataIdentification

Public fields

spatialRepresentationType

spatialRepresentationType [0..*]: ISOSpatialRepresentationType

spatialResolution

spatialResolution [0..*]: ISOResolution

language

language [1..*]: character

characterSet

characterSet [0..*]: ISOCharacterSet

topicCategory

topicCategory [0..*]: ISOTopicCategory

extent

extent [0..*]: ISOExtent

supplementalInformation

supplementalInformation

Methods

Public methods

ISODataIdentification$new()

ISODataIdentification$addSpatialRepresentationType()

ISODataIdentification$setSpatialRepresentationType()

ISODataIdentification$delSpatialRepresentationType()

ISODataIdentification$addSpatialResolution()

ISODataIdentification$delSpatialResolution()

ISODataIdentification$addLanguage()

ISODataIdentification$setLanguage()

ISODataIdentification$delLanguage()

ISODataIdentification$addCharacterSet()

ISODataIdentification$setCharacterSet()

ISODataIdentification$delCharacterSet()

ISODataIdentification$addTopicCategory()

ISODataIdentification$setTopicCategory()

ISODataIdentification$delTopicCategory()

ISODataIdentification$addExtent()

ISODataIdentification$setExtent()

ISODataIdentification$delExtent()

ISODataIdentification$setSupplementalInformation()

ISODataIdentification$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

geometa::ISOIdentification$addAggregateInformation()

geometa::ISOIdentification$addCredit()

geometa::ISOIdentification$addFormat()

geometa::ISOIdentification$addGraphicOverview()

geometa::ISOIdentification$addKeywords()

geometa::ISOIdentification$addPointOfContact()

geometa::ISOIdentification$addResourceConstraints()

geometa::ISOIdentification$addResourceMaintenance()

geometa::ISOIdentification$addStatus()

geometa::ISOIdentification$delAggregateInformation()

geometa::ISOIdentification$delCredit()

geometa::ISOIdentification$delFormat()

geometa::ISOIdentification$delGraphicOverview()

geometa::ISOIdentification$delKeywords()

geometa::ISOIdentification$delPointOfContact()

geometa::ISOIdentification$delResourceConstraints()

geometa::ISOIdentification$delResourceMaintenance()

geometa::ISOIdentification$delStatus()

geometa::ISOIdentification$setAbstract()

geometa::ISOIdentification$setCitation()

geometa::ISOIdentification$setGraphicOverview()

geometa::ISOIdentification$setKeywords()

geometa::ISOIdentification$setPurpose()

geometa::ISOIdentification$setResourceConstraints()

geometa::ISOIdentification$setResourceMaintenance()

Method new()

Initializes object

Usage

ISODataIdentification$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method addSpatialRepresentationType()

Adds spatial representation type

Usage

ISODataIdentification$addSpatialRepresentationType(spatialRepresentationType)

Arguments

spatialRepresentationType

object of class ISOSpatialRepresentationType or any character among values returned by ISOSpatialRepresentationType$values()

Returns

TRUE if added, FALSE otherwise

Method setSpatialRepresentationType()

Sets spatial representation type

Usage

ISODataIdentification$setSpatialRepresentationType(spatialRepresentationType)

Arguments

spatialRepresentationType

object of class ISOSpatialRepresentationType or any character among values returned by ISOSpatialRepresentationType$values()

Returns

TRUE if added, FALSE otherwise

Method delSpatialRepresentationType()

Deletes spatial representation type

Usage

ISODataIdentification$delSpatialRepresentationType(spatialRepresentationType)

Arguments

spatialRepresentationType

object of class ISOSpatialRepresentationType or any character among values returned by ISOSpatialRepresentationType$values()

Returns

TRUE if deleted, FALSE otherwise

Method addSpatialResolution()

Adds spatial resolution

Usage

ISODataIdentification$addSpatialResolution(resolution)

Arguments

resolution

object of class ISOResolution

Returns

TRUE if added, FALSE otherwise

Method delSpatialResolution()

Deletes spatial resolution

Usage

ISODataIdentification$delSpatialResolution(resolution)

Arguments

resolution

object of class ISOResolution

Returns

TRUE if deleted, FALSE otherwise

Method addLanguage()

Adds language

Usage

ISODataIdentification$addLanguage(locale)

Arguments

locale

object of class ISOLanguage or any character value among those returned by ISOLanguage$values()

Returns

TRUE if added, FALSE otherwise

Method setLanguage()

Sets language

Usage

ISODataIdentification$setLanguage(locale)

Arguments

locale

object of class ISOLanguage or any character value among those returned by ISOLanguage$values()

Returns

TRUE if added, FALSE otherwise

Method delLanguage()

Deletes language

Usage

ISODataIdentification$delLanguage(locale)

Arguments

locale

object of class ISOLanguage or any character value among those returned by ISOLanguage$values()

Returns

TRUE if deleted, FALSE otherwise

Method addCharacterSet()

Adds character set

Usage

ISODataIdentification$addCharacterSet(charset)

Arguments

charset

object of class ISOCharacterSet or any character value among those returned by ISOCharacterSet$values()

Returns

TRUE if added, FALSE otherwise

Method setCharacterSet()

Sets character set

Usage

ISODataIdentification$setCharacterSet(charset)

Arguments

charset

object of class ISOCharacterSet or any character value among those returned by ISOCharacterSet$values()

Returns

TRUE if added, FALSE otherwise

Method delCharacterSet()

Deletes character set

Usage

ISODataIdentification$delCharacterSet(charset)

Arguments

charset

object of class ISOCharacterSet or any character value among those returned by ISOCharacterSet$values()

Returns

TRUE if deleted, FALSE otherwise

Method addTopicCategory()

Adds topic category

Usage

ISODataIdentification$addTopicCategory(topicCategory)

Arguments

topicCategory

object of class ISOTopicCategory or any character value among those returned by ISOTopicCategory$values()

Returns

TRUE if added, FALSE otherwise

Method setTopicCategory()

Sets topic category

Usage

ISODataIdentification$setTopicCategory(topicCategory)

Arguments

topicCategory

object of class ISOTopicCategory or any character value topicCategory those returned by ISOTopicCategory$values()

Returns

TRUE if added, FALSE otherwise

Method delTopicCategory()

Deletes topic category

Usage

ISODataIdentification$delTopicCategory(topicCategory)

Arguments

topicCategory

object of class ISOTopicCategory or any character value among those returned by ISOTopicCategory$values()

Returns

TRUE if deleted, FALSE otherwise

Method addExtent()

Adds extent

Usage

ISODataIdentification$addExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if added, FALSE otherwise

Method setExtent()

Sets extent

Usage

ISODataIdentification$setExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if added, FALSE otherwise

Method delExtent()

Deletes extent

Usage

ISODataIdentification$delExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if deleted, FALSE otherwise

Method setSupplementalInformation()

Set supplemental information

Usage

ISODataIdentification$setSupplementalInformation(
  supplementalInformation,
  locales = NULL
)

Arguments

supplementalInformation

supplemental information

locales

a list of localized information. Default is NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

ISODataIdentification$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   #create dataIdentification
   md <- ISODataIdentification$new()
   md$setAbstract("abstract")
   md$setPurpose("purpose")
   md$addLanguage("eng")
#> [1] TRUE
   md$addCharacterSet("utf8")
#> [1] TRUE
   md$addTopicCategory("biota")
#> [1] TRUE
   md$addTopicCategory("oceans")
#> [1] TRUE
   
   #adding a point of contact
   rp <- ISOResponsibleParty$new()
   rp$setIndividualName("someone")
   rp$setOrganisationName("somewhere")
   rp$setPositionName("someposition")
   rp$setRole("pointOfContact")
   contact <- ISOContact$new()
   phone <- ISOTelephone$new()
   phone$setVoice("myphonenumber")
   phone$setFacsimile("myfacsimile")
   contact$setPhone(phone)
   address <- ISOAddress$new()
   address$setDeliveryPoint("theaddress")
   address$setCity("thecity")
   address$setPostalCode("111")
   address$setCountry("France")
   address$setEmail("someone@theorg.org")
   contact$setAddress(address)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact$setOnlineResource(res)
   rp$setContactInfo(contact)
   md$addPointOfContact(rp)
#> [1] TRUE
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015, 1, 1, 1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
   ct$addPresentationForm("mapDigital")
#> [1] TRUE
   ct$addCitedResponsibleParty(rp)
#> [1] TRUE
   md$setCitation(ct)
   
   #graphic overview
   go <- ISOBrowseGraphic$new(
     fileName = "http://wwww.somefile.org/png",
     fileDescription = "Map Overview",
     fileType = "image/png"
   )
   md$addGraphicOverview(go)
#> [1] TRUE
   
   #maintenance information
   mi <- ISOMaintenanceInformation$new()
   mi$setMaintenanceFrequency("daily")
   md$addResourceMaintenance(mi)
#> [1] TRUE
   
   #adding legal constraints
   lc <- ISOLegalConstraints$new()
   lc$addUseLimitation("limitation1")
#> [1] TRUE
   lc$addUseLimitation("limitation2")
#> [1] TRUE
   lc$addUseLimitation("limitation3")
#> [1] TRUE
   lc$addAccessConstraint("copyright")
#> [1] TRUE
   lc$addAccessConstraint("license")
#> [1] TRUE
   lc$addUseConstraint("copyright")
#> [1] TRUE
   lc$addUseConstraint("license")
#> [1] TRUE
   md$addResourceConstraints(lc)
#> [1] TRUE
   
   #adding extent
   extent <- ISOExtent$new()
   bbox <- ISOGeographicBoundingBox$new(minx = -180, miny = -90, maxx = 180, maxy = 90)
   extent$addGeographicElement(bbox)
#> [1] TRUE
   md$addExtent(extent)
#> [1] TRUE
   
   #add keywords
   kwds <- ISOKeywords$new()
   kwds$addKeyword("keyword1")
#> [1] TRUE
   kwds$addKeyword("keyword2")
#> [1] TRUE
   kwds$setKeywordType("theme")
   th <- ISOCitation$new()
   th$setTitle("General")
   th$addDate(d)
   kwds$setThesaurusName(th)
   md$addKeywords(kwds)
#> [1] TRUE
   
   #supplementalInformation
   md$setSupplementalInformation("some additional information")
   
   xml <- md$encode()
#> [geometa][INFO] Object 'ISODataIdentification' is VALID according to ISO 19139 XML schemas!