ISODataQualityAbstractElement

ISODataQualityAbstractElement

Format

R6Class object.

Value

Object of R6Class for modelling an ISODataQualityAbstractElement

Fields

nameOfMeasure

[character] measure name

measureIdentification

[ISOMetaIdentifier] measure identification

measureDescription

[character] measure description

evaluationMethodType

[ISOEvaluationMethodType] measure evaluation method type

evaluationMethodDescription

[character] measure evaluation method description

evaluationProcedure

[character] evaluation procedure

dateTime

[POSIXt] dateTime

result

[ISOConformanceResult] conformance result(s)

Methods

new(xml)

This method is used to instantiate an ISODataQuality

addNameOfMeasure(name, locales)

Add name. Locale names can be specified as list with the locales argument.

delNameOfMeasure(name, locales)

Deletes name. Locale names can be specified as list with the locales argument. Local names should match those of the name to be deleted, otherwise nothing will be deleted.

setMeasureIdentification(identification)

Sets measure identifier, an object of class ISOMetaIdentifier

setMeasureDescription(description, locales)

Sets measure description. Locale names can be specified as list with the locales argument.

setEvaluationMethodType(type)

Sets the type of evaluation method, an object of class ISOEvaluationMethodType or any character value among codeISOEvaluationMethodType$values()

setEvaluationMethodDescription(description, loales)

Sets the description of evaluation method. Locale names can be specified as list with the locales argument.

setEvaluationMethodProcedure(procedure)

Sets the procedure as an object of class ISOCitation

setDateTime(dateTime)

Sets datetime, object class 'POSIXct'/'POSIXt'

addResult(result)

Sets conformance result, object of class ISOConformanceResult

delResult(result)

Deletes conformance result, object of class ISOConformanceResult

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>