ISOExtendedElementInformation

ISOExtendedElementInformation

ISOExtendedElementInformation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO ExtendedElementInformation

References

ISO 19115:2003 - Geographic information – Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOExtendedElementInformation

Public fields

name

name [1..1]: character

shortName

shortName [0..1]: character

domainCode

domainCode [0..1]: integer

definition

definition [1..1]: character

obligation

obligation [0..1]: ISOObligation

condition

condition [0..1]: character

dataType

dataType [1..1]: ISODatatype

maximumOccurrence

maximumOccurrence [0..1]: character

domainValue

domainValue [0..1]: character

parentEntity

parentEntity [1..*]: character

rule

rule [1..1]: character

rationale

rationale [0..*]: character

source

source [1..*]: ISOResponsibleParty

Methods

Public methods

ISOExtendedElementInformation$new()

ISOExtendedElementInformation$setName()

ISOExtendedElementInformation$setShortName()

ISOExtendedElementInformation$setDomainCode()

ISOExtendedElementInformation$setDefinition()

ISOExtendedElementInformation$setObligation()

ISOExtendedElementInformation$setCondition()

ISOExtendedElementInformation$setDatatype()

ISOExtendedElementInformation$setMaximumOccurrence()

ISOExtendedElementInformation$setDomainValue()

ISOExtendedElementInformation$addParentEntity()

ISOExtendedElementInformation$delParentEntity()

ISOExtendedElementInformation$setRule()

ISOExtendedElementInformation$addRationale()

ISOExtendedElementInformation$delRationale()

ISOExtendedElementInformation$addSource()

ISOExtendedElementInformation$delSource()

ISOExtendedElementInformation$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOExtendedElementInformation$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setName()

Set name

Usage

ISOExtendedElementInformation$setName(name, locales = NULL)

Arguments

name

name

locales

list of localized names. Default is NULL

Method setShortName()

Set short name

Usage

ISOExtendedElementInformation$setShortName(shortName, locales = NULL)

Arguments

shortName

short name

locales

list of localized short names. Default is NULL

Method setDomainCode()

Set domain code

Usage

ISOExtendedElementInformation$setDomainCode(domainCode)

Arguments

domainCode

domain code, object of class integer

Method setDefinition()

Set definition

Usage

ISOExtendedElementInformation$setDefinition(definition, locales = NULL)

Arguments

definition

definition

locales

list of localized definitions. Default is NULL

Method setObligation()

Set obligation

Usage

ISOExtendedElementInformation$setObligation(obligation)

Arguments

obligation

obligation, object of class ISOObligation or any character value among those returned by ISOObligation$values()

Method setCondition()

Set condition

Usage

ISOExtendedElementInformation$setCondition(condition, locales = NULL)

Arguments

condition

condition

locales

list of localized conditions. Default is NULL

Method setDatatype()

Set data type

Usage

ISOExtendedElementInformation$setDatatype(dataType)

Arguments

dataType

data type, object of class ISODatatype or any character value among those returned by ISODatatype$values()

Method setMaximumOccurrence()

Set maximum occurrence

Usage

ISOExtendedElementInformation$setMaximumOccurrence(maximumOccurrence)

Arguments

maximumOccurrence

max occurrence

Method setDomainValue()

Set domain value

Usage

ISOExtendedElementInformation$setDomainValue(domainValue)

Arguments

domainValue

domain value

Method addParentEntity()

Adds parent entity

Usage

ISOExtendedElementInformation$addParentEntity(entity)

Arguments

entity

parent entity

Returns

TRUE if added, FALSE otherwise

Method delParentEntity()

Deletes parent entity

Usage

ISOExtendedElementInformation$delParentEntity(entity)

Arguments

entity

parent entity

Returns

TRUE if deleted, FALSE otherwise

Method setRule()

Set rule

Usage

ISOExtendedElementInformation$setRule(rule, locales = NULL)

Arguments

rule

rule

locales

list of localized rules. Default is NULL

Method addRationale()

Adds rationale

Usage

ISOExtendedElementInformation$addRationale(rationale, locales = NULL)

Arguments

rationale

rationale

locales

list of localized rationales. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delRationale()

Deletes rationale

Usage

ISOExtendedElementInformation$delRationale(rationale, locales = NULL)

Arguments

rationale

rationale

locales

list of localized rationales. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method addSource()

Adds source

Usage

ISOExtendedElementInformation$addSource(source)

Arguments

source

source, object of class ISOResponsibleParty

Returns

TRUE if added, FALSE otherwise

Method delSource()

Deletes source

Usage

ISOExtendedElementInformation$delSource(source)

Arguments

source

source, object of class ISOResponsibleParty

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOExtendedElementInformation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  md <- ISOExtendedElementInformation$new()
  md$setName("name")
  md$setShortName("shortName")
  md$setDomainCode(1L)
  md$setDefinition("some definition")
  md$setObligation("mandatory")
#> Error in if (cl$id == id) {    codelist <<- cl}: argument is of length zero
  md$setCondition("no condition")
  md$setDatatype("characterString")
#> Error in if (cl$id == id) {    codelist <<- cl}: argument is of length zero
  md$setMaximumOccurrence("string")
  md$setDomainValue("value")
  md$addParentEntity("none")
#> [1] TRUE
  md$setRule("rule")
  md$addRationale("rationale")
#> [1] TRUE
  
  #adding a source
  rp <- ISOResponsibleParty$new()
  rp$setIndividualName("someone")
  rp$setOrganisationName("somewhere")
  rp$setPositionName("someposition")
  rp$setRole("pointOfContact")
#> Error in if (cl$id == id) {    codelist <<- cl}: argument is of length zero
  contact <- ISOContact$new()
  phone <- ISOTelephone$new()
  phone$setVoice("myphonenumber")
  phone$setFacsimile("myfacsimile")
  contact$setPhone(phone)
  address <- ISOAddress$new()
  address$setDeliveryPoint("theaddress")
  address$setCity("thecity")
  address$setPostalCode("111")
  address$setCountry("France")
  address$setEmail("someone@theorg.org")
  contact$setAddress(address)
  res <- ISOOnlineResource$new()
  res$setLinkage("http://www.somewhereovertheweb.org")
  res$setName("somename")
  contact$setOnlineResource(res)
  rp$setContactInfo(contact)
  
  md$addSource(rp)
#> [1] TRUE
  
  xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmd}CI_ResponsibleParty': Missing child element(s). Expected is ( {http://www.isotc211.org/2005/gmd}role ) at line 34. 
#> [geometa][WARN] Object 'ISOExtendedElementInformation' is INVALID according to ISO 19139 XML schemas!