ISOFeatureType19139

ISOFeatureType19139

ISOFeatureType19139

Format

R6Class object.

Value

Object of R6Class for modelling an ISO FeatureType in ISO 19139

References

- ISO 19110 - GFC 1.0 https://schemas.isotc211.org/19110/-/gfc/1.0/gfc/#element_FC_FeatureType

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOFeatureType -> ISOFeatureType19139

Public fields

typeName

typeName [1..1]: ISOLocalName or character

definition

definition [0..1]: character

code

code [0..1]: character

isAbstract

isAbstract [1..1]: logical

aliases

aliases [0..*]: ISOLocalName or character

inheritsFrom

inheritsFrom [0..*]: ISOInheritanceRelation

inheritsTo

inheritsTo [0..*]: ISOInheritanceRelation

featureCatalogue

featureCatalogue: ISOFeatureCatalogue

constrainedBy

constrainedBy [0..*]: ISOConstraint

definitionReference

definitionReference [0..*]: ISODefinitionReference

carrierOfCharacteristics

carrierOfCharacteristics [0..*]: ISOCarrierOfCharacteristics

Methods

Public methods

ISOFeatureType19139$new()

ISOFeatureType19139$setTypeName()

ISOFeatureType19139$setDefinition()

ISOFeatureType19139$setCode()

ISOFeatureType19139$setIsAbstract()

ISOFeatureType19139$addAlias()

ISOFeatureType19139$delAlias()

ISOFeatureType19139$addInheritsFrom()

ISOFeatureType19139$delInheritsFrom()

ISOFeatureType19139$addInheritsTo()

ISOFeatureType19139$delInheritsTo()

ISOFeatureType19139$setFeatureCatalogue()

ISOFeatureType19139$addConstraint()

ISOFeatureType19139$delConstraint()

ISOFeatureType19139$setDefinitionReference()

ISOFeatureType19139$addCharacteristic()

ISOFeatureType19139$delCharacteristic()

ISOFeatureType19139$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOFeatureType19139$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setTypeName()

Set type name

Usage

ISOFeatureType19139$setTypeName(typeName)

Arguments

typeName

type name, object of class ISOLocalName or character

Method setDefinition()

Set definition

Usage

ISOFeatureType19139$setDefinition(definition, locales = NULL)

Arguments

definition

definition

locales

list of localized definitions. Default is NULL

Method setCode()

Set code

Usage

ISOFeatureType19139$setCode(code, locales = NULL)

Arguments

code

definition

locales

list of localized codes. Default is NULL

Method setIsAbstract()

Set whether feature type is abstract

Usage

ISOFeatureType19139$setIsAbstract(isAbstract)

Arguments

isAbstract

object of class logical

Method addAlias()

Adds alias

Usage

ISOFeatureType19139$addAlias(alias)

Arguments

alias

object of class ISOLocalName or character

Returns

TRUE if added, FALSE otherwise

Method delAlias()

Usage

ISOFeatureType19139$delAlias(alias)

Arguments

alias

object of class ISOLocalName or character

Returns

TRUE if deleted, FALSE otherwise

Method addInheritsFrom()

Adds 'inheritsFrom' relation

Usage

ISOFeatureType19139$addInheritsFrom(rel)

Arguments

rel

rel, object of class ISOInheritanceRelation

Returns

TRUE if added, FALSE otherwise

Method delInheritsFrom()

Deletes 'inheritsFrom' relation

Usage

ISOFeatureType19139$delInheritsFrom(rel)

Arguments

rel

rel, object of class ISOInheritanceRelation

Returns

TRUE if deleted, FALSE otherwise

Method addInheritsTo()

Adds 'inheritsTo' relation

Usage

ISOFeatureType19139$addInheritsTo(rel)

Arguments

rel

rel, object of class ISOInheritanceRelation

Returns

TRUE if added, FALSE otherwise

Method delInheritsTo()

Deletes 'inheritsTo' relation

Usage

ISOFeatureType19139$delInheritsTo(rel)

Arguments

rel

rel, object of class ISOInheritanceRelation

Returns

TRUE if deleted, FALSE otherwise

Method setFeatureCatalogue()

Set feature catalogue

Usage

ISOFeatureType19139$setFeatureCatalogue(fc)

Arguments

fc

object of class ISOFeatureCatalogue

Method addConstraint()

Adds constraint

Usage

ISOFeatureType19139$addConstraint(constraint)

Arguments

constraint

constraint, object of class ISOConstraint

Returns

TRUE if added, FALSE otherwise

Method delConstraint()

Deletes constraint

Usage

ISOFeatureType19139$delConstraint(constraint)

Arguments

constraint

constraint, object of class ISOConstraint

Returns

TRUE if deleted, FALSE otherwise

Method setDefinitionReference()

Set definition reference

Usage

ISOFeatureType19139$setDefinitionReference(definitionReference)

Arguments

definitionReference

object of class ISODefinitionReference

Method addCharacteristic()

Adds characteristic

Usage

ISOFeatureType19139$addCharacteristic(characteristic)

Arguments

characteristic

characteristic, object inheriting class ISOAbstractCarrierOfCharacteristics

Returns

TRUE if added, FALSE otherwise

Method delCharacteristic()

Deletes characteristic

Usage

ISOFeatureType19139$delCharacteristic(characteristic)

Arguments

characteristic

characteristic, object inheriting class ISOAbstractCarrierOfCharacteristics

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOFeatureType19139$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 #featuretype
 md <- ISOFeatureType$new()
 md$setTypeName("typeName")
 md$setDefinition("definition")
 md$setCode("code")
 md$setIsAbstract(FALSE)
 md$addAlias("alias1")
#> [1] TRUE
 md$addAlias("alias2")
#> [1] TRUE
 
 #add feature attributes
 for(i in 1:3){
   #create attribute
   fat <- ISOFeatureAttribute$new()
   fat$setMemberName(sprintf("name %s",i))
   fat$setDefinition(sprintf("definition %s",i))
   fat$setCardinality(lower=1,upper=1)
   fat$setCode(sprintf("code %s",i))
   
   #add measurement unit
   gml <- GMLBaseUnit$new(id = "ID%")
   gml$setDescriptionReference("someref")
   gml$setIdentifier("identifier", "codespace")
   gml$addName("name1", "codespace")
   gml$addName("name2", "codespace")
   gml$setQuantityTypeReference("someref")
   gml$setCatalogSymbol("symbol")
   gml$setUnitsSystem("somelink")
   fat$setValueMeasurementUnit(gml)
   
   #add listed values
   val1 <- ISOListedValue$new()
   val1$setCode("code1")
   val1$setLabel("label1")
   val1$setDefinition("definition1")
   fat$addListedValue(val1)
   val2 <- ISOListedValue$new()
   val2$setCode("code2")
   val2$setLabel("label2")
   val2$setDefinition("definition2")
   fat$addListedValue(val2)
   fat$setValueType("typeName")
   
   #add feature attribute as carrierOfCharacteristic
   md$addCharacteristic(fat)
 }
 xml <- md$encode()
#> [geometa][WARN] Element '{http://www.opengis.net/gml/3.2}BaseUnit', attribute '{http://www.opengis.net/gml/3.2}id': 'ID%' is not a valid value of the atomic type 'xs:ID' at line 47. 
#> [geometa][WARN] Element '{http://www.opengis.net/gml/3.2}BaseUnit', attribute '{http://www.opengis.net/gml/3.2}id': 'ID%' is not a valid value of the atomic type 'xs:ID' at line 118. 
#> [geometa][WARN] Element '{http://www.opengis.net/gml/3.2}BaseUnit', attribute '{http://www.opengis.net/gml/3.2}id': 'ID%' is not a valid value of the atomic type 'xs:ID' at line 189. 
#> [geometa][WARN] Object 'ISOFeatureType19139' is INVALID according to ISO 19139 XML schemas!