ISOGeorectified

ISOGeorectified

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Georectified

Fields

checkPointAvailability

[logical]

checkPointDescription

[character]

cornerPoints

[GMLPoint]

centerPoint

[GMLPoint]

pointInPixel

[ISOPixelOrientation]

transformationDimensionDescription

[character]

transformationDimensionMapping

[character]

Inherited methods from ISOGridSpatialRepresentation

setNumberOfDimensions

Sets the number of dimensions (value of class integer)

addDimension(dimension)

Adds a dimension. Object of class ISODimension

delDimension(dimension)

Deletes a dimension;

setCellGeometry(cellGeometry)

Sets the cell geometry. Object of class ISOCellGeometry or any value from ISOCellGeometry$values()

setTransformationParameterAvailability(availability)

Sets the transformation parameter availability

setPixelOrientation(pixelOrientation)

Sets the point in pixel orientation, object of class 'character' or ISOPixelOrientation

Methods

new(xml,value)

This method is used to instantiate an ISOGeorectified

setCheckPointAvailability(availability)

Set checkpoint availability, object of class 'logical' (TRUE/FALSE)

setCheckPointDescription(description, locales)

Set checkpoint description

addCornerPoint(sfg,m)

Adds a corner point, either an object of class 'sfg' (from sf) or a 'matrix'

delCornerPoint(sfg,m)

Deletes a corner point, either an object of class 'sfg' (from sf) or a 'matrix'

setCenterPoint(sfg,m)

Sets a center point, either an object of class 'sfg' (from sf) or a 'matrix'

setTransformationDimensionDescription(description, locales)

Sets the transformation dimension description.

addTransformationDimensionMapping(mapping)

Adds a transformation dimension mapping

delTransformationDimensionMapping(mapping)

Deletes a transformation dimension mapping

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>