ISOIdentification

ISOIdentification

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Identification

Fields

citation

[ISOCitation] citation

abstract

[character] abstract

purpose

[character] purpose

credit

[character] credit

status

[ISOStatus] status

pointOfContact

[ISOResponsibleParty] point(s) of contact

resourceMaintenance

[ISOMaintenanceInformation] maintenance information

graphicOverview

[ISOBrowseGraphic] graphic overview(s)

resourceFormat

[ISOFormat] format(s)

descriptiveKeywords

[ISOKeywords] keyword(s)

resourceConstraints

[ISOConstraints] constraint(s)

resourceSpecificUsage

[ISOUsage] usage(s)

aggregationInfo

[ISOAggregateInformation] aggregate information

Methods

new(xml, element, namespace, defaults)

This method is used to instantiate an ISOIdentification

setCitation(citation)

Sets an object of class ISOCitation

setAbstract(abstract, locales)

Sets an abstract (object of class "character"). Locale names can be specified as list with the locales argument.

setPurpose(purpose, locales)

Sets a purpose (object of class "character"). Locale names can be specified as list with the locales argument.

addCredit(credit, locales)

Adds a credit (object of class "character"). Locale names can be specified as list with the locales argument.

delCredit(credit, locales)

Deletes a credit (object of class "character"). Locale names can be specified as list with the locales argument.

addStatus(status)

Adds a status, as object of class "character" or class ISOStatus. If an object of class "character" is specified, it must match the accepted progress status values ISOStatus$values().

delStatus(status)

Deletes a status, as object of class "character" or class ISOStatus. If an object of class "character" is specified, it must match the accepted progress status values ISOStatus$values().

addPointOfContact(pointOfContact)

Adds an object of class ISOResponsibleParty

delPointOfContact(pointOfContact)

Deletes an object of class ISOResponsibleParty

addResourceMaintenance(resourceMaintenance)

Adds a resource maintenance information as object of class ISOMaintenanceInformation.

setResourceMaintenance(resourceMaintenance)

Sets a resource maintenance information as object of class ISOMaintenanceInformation.

delResourceMaintenance(resourceMaintenance)

Deletes a resource maintenance information as object of class ISOMaintenanceInformation.

addGraphicOverview(graphicOverview)

Adds an object of class ISOBrowseGraphic

setGraphicOverview(graphicOverview)

Sets an object of class ISOBrowseGraphic

delGraphicOverview(graphicOverview)

Deletes an object of class ISOBrowseGraphic

addFormat(format)

Adds a resource format, object of class ISOFormat

delFormat(format)

Deletes a resource format, object of class ISOFormat

addKeywords(keywords)

Adds a set of keywords as object of class ISOKeywords

setKeywords(keywords)

Sets a set of keywords as object of class ISOKeywords

delKeywords(keywords)

Deletes a set of keywords as object of class ISOKeywords

addResourceConstraints(resourceConstraints)

Adds an object extending ISOConstraints, either an object of class ISOLegalConstraints or ISOSecurityConstraints

setResourceConstraints(resourceConstraints)

Sets an object extending ISOConstraints, either an object of class ISOLegalConstraints or ISOSecurityConstraints

delResourceConstraints(resourceConstraints)

Deletes an object extending ISOConstraints, either an object of class ISOLegalConstraints or ISOSecurityConstraints

addAggregateInformation(aggregateInfo)

Adds an object of class ISOAggregateInformation

delAggregateInformation(aggregateInfo)

Deletes an object of class ISOAggregateInformation

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>