ISOIdentification19115_3

ISOIdentification19115_3

ISOIdentification19115_3

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Identification in ISO 19115-3

References

- ISO 19115-3 https://schemas.isotc211.org/19115/-3/mri/1.0/mri/#element_AbstractMD_Identification

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOIdentification -> ISOIdentification19115_3

Public fields

citation

citation: ISOCitation

abstract

abstract: character

purpose

purpose [0..1]: character

credit

credit [0..*]: character

status

status [0..*]: ISOProgress

pointOfContact

pointOfContact [0..*]: ISOResponsibility

spatialRepresentationType

spatialRepresentationType [0..*]: ISOSpatialRepresentationType

spatialResolution

spatialResolution [0..*]: ISOResolution

temporalResolution

[0..*]: ISOPeriodDuration

topicCategory

topicCategory [0..*]: ISOTopicCategory

extent

extent [0..*]: ISOExtent (ISO 19139)

additionalDocumentation

additional documentation [0..*]: ISOCitation

processingLevel

processing level [0..1]: ISOMetaIdentifier

resourceMaintenance

resourceMaintenance [0..*]: ISOMaintenanceInformation

graphicOverview

graphicOverview [0..*]: ISOBrowseGraphic

resourceFormat

resourceFormat [0..*]: ISOFormat

descriptiveKeywords

descriptiveKeywords [0..*]: ISOKeywords (ISO 19139)

resourceSpecificUsage

resourceSpecificUsage [0..*]: ISOUsage (ISO 19139)

resourceConstraints

resourceConstraints [0..*]: ISOLegalConstraints

aggregationInfo

aggregationInfo [0..*]: ISOAggregateInformation

Methods

Public methods

ISOIdentification19115_3$new()

ISOIdentification19115_3$setCitation()

ISOIdentification19115_3$setAbstract()

ISOIdentification19115_3$setPurpose()

ISOIdentification19115_3$addCredit()

ISOIdentification19115_3$delCredit()

ISOIdentification19115_3$addStatus()

ISOIdentification19115_3$delStatus()

ISOIdentification19115_3$addPointOfContact()

ISOIdentification19115_3$delPointOfContact()

ISOIdentification19115_3$addSpatialRepresentationType()

ISOIdentification19115_3$delSpatialRepresentationType()

ISOIdentification19115_3$addSpatialResolution()

ISOIdentification19115_3$delSpatialResolution()

ISOIdentification19115_3$addTemporalResolution()

ISOIdentification19115_3$delTemporalResolution()

ISOIdentification19115_3$addTopicCategory()

ISOIdentification19115_3$delTopicCategory()

ISOIdentification19115_3$addExtent()

ISOIdentification19115_3$delExtent()

ISOIdentification19115_3$addAdditionalDocumentation()

ISOIdentification19115_3$delAdditionalDocumentation()

ISOIdentification19115_3$setProcessingLevel()

ISOIdentification19115_3$addResourceMaintenance()

ISOIdentification19115_3$delResourceMaintenance()

ISOIdentification19115_3$addGraphicOverview()

ISOIdentification19115_3$delGraphicOverview()

ISOIdentification19115_3$addFormat()

ISOIdentification19115_3$delFormat()

ISOIdentification19115_3$addKeywords()

ISOIdentification19115_3$delKeywords()

ISOIdentification19115_3$addResourceSpecificUsage()

ISOIdentification19115_3$delResourceSpecificUsage()

ISOIdentification19115_3$addResourceConstraints()

ISOIdentification19115_3$delResourceConstraints()

ISOIdentification19115_3$addAssociatedResource()

ISOIdentification19115_3$delAssociatedResource()

ISOIdentification19115_3$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOIdentification19115_3$new(xml = NULL, defaults = list())

Arguments

xml

object of class XMLInternalNode-class

defaults

defaults list

Method setCitation()

Set citation

Usage

ISOIdentification19115_3$setCitation(citation)

Arguments

citation

object of class ISOCitation

Method setAbstract()

Set abstract

Usage

ISOIdentification19115_3$setAbstract(abstract, locales = NULL)

Arguments

abstract

abstract

locales

list of localized abstracts. Default is NULL

Method setPurpose()

Set purpose

Usage

ISOIdentification19115_3$setPurpose(purpose, locales = NULL)

Arguments

purpose

purpose

locales

list of localized texts. Default is NULL

Method addCredit()

Adds credit

Usage

ISOIdentification19115_3$addCredit(credit, locales = NULL)

Arguments

credit

credit

locales

list of localized texts. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delCredit()

Deletes credit

Usage

ISOIdentification19115_3$delCredit(credit, locales = NULL)

Arguments

credit

credit

locales

list of localized texts. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method addStatus()

Adds status

Usage

ISOIdentification19115_3$addStatus(status)

Arguments

status

object of class ISOProgress or any character among values returned by ISOProgress$values()

Returns

TRUE if added, FALSE otherwise

Method delStatus()

Deletes status

Usage

ISOIdentification19115_3$delStatus(status)

Arguments

status

object of class ISOProgress or any character among values returned by ISOProgress$values()

Returns

TRUE if deleted, FALSE otherwise

Method addPointOfContact()

Adds point of contact

Usage

ISOIdentification19115_3$addPointOfContact(pointOfContact)

Arguments

pointOfContact

object of class ISOResponsibility

Returns

TRUE if added, FALSE otherwise

Method delPointOfContact()

Deletes point of contact

Usage

ISOIdentification19115_3$delPointOfContact(pointOfContact)

Arguments

pointOfContact

object of class ISOResponsibility

Returns

TRUE if deleted, FALSE otherwise

Method addSpatialRepresentationType()

Adds spatial representation type

Usage

ISOIdentification19115_3$addSpatialRepresentationType(
  spatialRepresentationType
)

Arguments

spatialRepresentationType

object of class ISOSpatialRepresentationType or any character among values returned by ISOSpatialRepresentationType$values()

Returns

TRUE if added, FALSE otherwise

Method delSpatialRepresentationType()

Deletes spatial representation type

Usage

ISOIdentification19115_3$delSpatialRepresentationType(
  spatialRepresentationType
)

Arguments

spatialRepresentationType

object of class ISOSpatialRepresentationType or any character among values returned by ISOSpatialRepresentationType$values()

Returns

TRUE if deleted, FALSE otherwise

Method addSpatialResolution()

Adds spatial resolution

Usage

ISOIdentification19115_3$addSpatialResolution(resolution)

Arguments

resolution

object of class ISOResolution or character

Returns

TRUE if added, FALSE otherwise

Method delSpatialResolution()

Deletes spatial resolution

Usage

ISOIdentification19115_3$delSpatialResolution(resolution)

Arguments

resolution

object of class ISOResolution or character

Returns

TRUE if deleted, FALSE otherwise

Method addTemporalResolution()

Adds temporal resolution

Usage

ISOIdentification19115_3$addTemporalResolution(resolution)

Arguments

resolution

object of class ISOPeriodDuration or character

Returns

TRUE if added, FALSE otherwise

Method delTemporalResolution()

Deletes temporal resolution

Usage

ISOIdentification19115_3$delTemporalResolution(resolution)

Arguments

resolution

object of class ISOPeriodDuration or character

Returns

TRUE if deleted, FALSE otherwise

Method addTopicCategory()

Adds topic category

Usage

ISOIdentification19115_3$addTopicCategory(topicCategory)

Arguments

topicCategory

object of class ISOTopicCategory or any character value among those returned by ISOTopicCategory$values()

Returns

TRUE if added, FALSE otherwise

Method delTopicCategory()

Deletes topic category

Usage

ISOIdentification19115_3$delTopicCategory(topicCategory)

Arguments

topicCategory

object of class ISOTopicCategory or any character value among those returned by ISOTopicCategory$values()

Returns

TRUE if deleted, FALSE otherwise

Method addExtent()

Adds extent

Usage

ISOIdentification19115_3$addExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if added, FALSE otherwise

Method delExtent()

Deletes extent

Usage

ISOIdentification19115_3$delExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if deleted, FALSE otherwise

Method addAdditionalDocumentation()

Adds additional documentation

Usage

ISOIdentification19115_3$addAdditionalDocumentation(additionalDocumentation)

Arguments

additionalDocumentation

object of class ISOCitation

Returns

TRUE if added, FALSE otherwise

Method delAdditionalDocumentation()

Deletes additional documentation

Usage

ISOIdentification19115_3$delAdditionalDocumentation(additionalDocumentation)

Arguments

additionalDocumentation

object of class ISOCitation

Returns

TRUE if deleted, FALSE otherwise

Method setProcessingLevel()

Set processing level

Usage

ISOIdentification19115_3$setProcessingLevel(processingLevel)

Arguments

processingLevel

object of class ISOMetaIdentifier or character

Method addResourceMaintenance()

Adds resource maintenance

Usage

ISOIdentification19115_3$addResourceMaintenance(resourceMaintenance)

Arguments

resourceMaintenance

object of class ISOMaintenanceInformation

Returns

TRUE if added, FALSE otherwise

Method delResourceMaintenance()

Deletes resource maintenance

Usage

ISOIdentification19115_3$delResourceMaintenance(resourceMaintenance)

Arguments

resourceMaintenance

object of class ISOMaintenanceInformation

Returns

TRUE if deleted, FALSE otherwise

Method addGraphicOverview()

Adds graphic overview

Usage

ISOIdentification19115_3$addGraphicOverview(graphicOverview)

Arguments

graphicOverview

object of class ISOBrowseGraphic

Returns

TRUE if added, FALSE otherwise

Method delGraphicOverview()

Deletes graphic overview

Usage

ISOIdentification19115_3$delGraphicOverview(graphicOverview)

Arguments

graphicOverview

object of class ISOBrowseGraphic

Returns

TRUE if deleted, FALSE otherwise

Method addFormat()

Adds format

Usage

ISOIdentification19115_3$addFormat(format)

Arguments

format

object of class ISOFormat

Returns

TRUE if added, FALSE otherwise

Method delFormat()

Deletes format

Usage

ISOIdentification19115_3$delFormat(format)

Arguments

format

object of class ISOFormat

Returns

TRUE if deleted, FALSE otherwise

Method addKeywords()

Adds keywords

Usage

ISOIdentification19115_3$addKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if added, FALSE otherwise

Method delKeywords()

Deletes keywords

Usage

ISOIdentification19115_3$delKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if deleted, FALSE otherwise

Method addResourceSpecificUsage()

Adds resource specific usage

Usage

ISOIdentification19115_3$addResourceSpecificUsage(usage)

Arguments

usage

object of class ISOUsage

Returns

TRUE if added, FALSE otherwise

Method delResourceSpecificUsage()

Deletes resource specific usage

Usage

ISOIdentification19115_3$delResourceSpecificUsage(usage)

Arguments

usage

object of class ISOUsage

Returns

TRUE if deleted, FALSE otherwise

Method addResourceConstraints()

Adds resource constraints

Usage

ISOIdentification19115_3$addResourceConstraints(resourceConstraints)

Arguments

resourceConstraints

object of class ISOConstraints

Returns

TRUE if added, FALSE otherwise

Method delResourceConstraints()

Deletes resource constraints

Usage

ISOIdentification19115_3$delResourceConstraints(resourceConstraints)

Arguments

resourceConstraints

object of class ISOConstraints

Returns

TRUE if deleted, FALSE otherwise

Method addAssociatedResource()

Adds associated resource

Usage

ISOIdentification19115_3$addAssociatedResource(associatedResource)

Arguments

associatedResource

object of class ISOAssociatedResource

Returns

TRUE if added, FALSE otherwise

Method delAssociatedResource()

Deletes associated resource

Usage

ISOIdentification19115_3$delAssociatedResource(associatedResource)

Arguments

associatedResource

object of class ISOAssociatedResource

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOIdentification19115_3$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.