ISOImageryAcquisitionInformation

ISOImageryAcquisitionInformation

ISOImageryAcquisitionInformation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Imagery AcquisitionInformation

References

ISO 19115-2:2009 - Geographic information – AcquisitionInformation – Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryAcquisitionInformation

Public fields

instrument

instrument [0..*]: ISOImageryInstrument

operation

operation [0..*]: ISOImageryOperation

platform

platform [0..*]: ISOImageryPlatform

acquisitionPlan

acquisitionPlan [0..*]: ISOImageryPlan

objective

objective [0..*]: ISOImageryObjective

acquisitionRequirement

acquisitionRequirement [0..*]: ISOImageryRequirement

environmentalConditions

environmentalConditions [0..1]: ISOImageryEnvironmentalRecord

Methods

Public methods

ISOImageryAcquisitionInformation$new()

ISOImageryAcquisitionInformation$addInstrument()

ISOImageryAcquisitionInformation$delInstrument()

ISOImageryAcquisitionInformation$addOperation()

ISOImageryAcquisitionInformation$delOperation()

ISOImageryAcquisitionInformation$addPlatform()

ISOImageryAcquisitionInformation$delPlatform()

ISOImageryAcquisitionInformation$addPlan()

ISOImageryAcquisitionInformation$delPlan()

ISOImageryAcquisitionInformation$addObjective()

ISOImageryAcquisitionInformation$delObjective()

ISOImageryAcquisitionInformation$addRequirement()

ISOImageryAcquisitionInformation$delRequirement()

ISOImageryAcquisitionInformation$setEnvironmentConditions()

ISOImageryAcquisitionInformation$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOImageryAcquisitionInformation$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method addInstrument()

Adds instrument

Usage

ISOImageryAcquisitionInformation$addInstrument(instrument)

Arguments

instrument

object of class ISOImageryInstrument

Returns

TRUE if added, FALSE otherwise

Method delInstrument()

Deletes instrument

Usage

ISOImageryAcquisitionInformation$delInstrument(instrument)

Arguments

instrument

object of class ISOImageryInstrument

Returns

TRUE if deleted, FALSE otherwise

Method addOperation()

Adds operation

Usage

ISOImageryAcquisitionInformation$addOperation(operation)

Arguments

operation

object of class ISOImageryOperation

Returns

TRUE if added, FALSE otherwise

Method delOperation()

Deletes operation

Usage

ISOImageryAcquisitionInformation$delOperation(operation)

Arguments

operation

object of class ISOImageryOperation

Returns

TRUE if deleted, FALSE otherwise

Method addPlatform()

Adds platform

Usage

ISOImageryAcquisitionInformation$addPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if added, FALSE otherwise

Method delPlatform()

Deletes platform

Usage

ISOImageryAcquisitionInformation$delPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if deleted, FALSE otherwise

Method addPlan()

Adds plan

Usage

ISOImageryAcquisitionInformation$addPlan(plan)

Arguments

plan

object of class ISOImageryPlan

Returns

TRUE if added, FALSE otherwise

Method delPlan()

Deletes plan

Usage

ISOImageryAcquisitionInformation$delPlan(plan)

Arguments

plan

object of class ISOImageryPlan

Returns

TRUE if deleted, FALSE otherwise

Method addObjective()

Adds objective

Usage

ISOImageryAcquisitionInformation$addObjective(objective)

Arguments

objective

object of class ISOImageryObjective

Returns

TRUE if added, FALSE otherwise

Method delObjective()

Deletes objective

Usage

ISOImageryAcquisitionInformation$delObjective(objective)

Arguments

objective

object of class ISOImageryObjective

Returns

TRUE if deleted, FALSE otherwise

Method addRequirement()

Adds requirement

Usage

ISOImageryAcquisitionInformation$addRequirement(requirement)

Arguments

requirement

object of class ISOImageryRequirement

Returns

TRUE if added, FALSE otherwise

Method delRequirement()

Deletes requirement

Usage

ISOImageryAcquisitionInformation$delRequirement(requirement)

Arguments

requirement

object of class ISOImageryRequirement

Returns

TRUE if deleted, FALSE otherwise

Method setEnvironmentConditions()

Set environment conditions

Usage

ISOImageryAcquisitionInformation$setEnvironmentConditions(conditions)

Arguments

conditions

object of class ISOImageryEnvironmentalRecord

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryAcquisitionInformation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

    md = ISOImageryAcquisitionInformation$new()
    
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_AcquisitionInformation': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryAcquisitionInformation' is INVALID according to ISO 19139 XML schemas!