ISOImageryAlgorithm

ISOImageryAlgorithm

ISOImageryAlgorithm

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery algorithm

References

ISO 19115-2:2009 - Geographic information – Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryAlgorithm

Public fields

citation

citation [1..1]: ISOCitation

description

description [1..1]: character|ISOLocalisedCharacterString

Methods

Public methods

ISOImageryAlgorithm$new()

ISOImageryAlgorithm$setCitation()

ISOImageryAlgorithm$setDescription()

ISOImageryAlgorithm$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initialized object

Usage

ISOImageryAlgorithm$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setCitation()

Set citation

Usage

ISOImageryAlgorithm$setCitation(citation)

Arguments

citation

object of class ISOCitation

Method setDescription()

Set description

Usage

ISOImageryAlgorithm$setDescription(description, locales = NULL)

Arguments

description

description

locales

list of localized texts. Default is NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryAlgorithm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   md <- ISOImageryAlgorithm$new()
   
   #add citation
   rp1 <- ISOResponsibleParty$new()
   rp1$setIndividualName("someone1")
   rp1$setOrganisationName("somewhere1")
   rp1$setPositionName("someposition1")
   rp1$setRole("pointOfContact")
   contact1 <- ISOContact$new()
   phone1 <- ISOTelephone$new()
   phone1$setVoice("myphonenumber1")
   phone1$setFacsimile("myfacsimile1")
   contact1$setPhone(phone1)
   address1 <- ISOAddress$new()
   address1$setDeliveryPoint("theaddress1")
   address1$setCity("thecity1")
   address1$setPostalCode("111")
   address1$setCountry("France")
   address1$setEmail("someone1@theorg.org")
   contact1$setAddress(address1)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact1$setOnlineResource(res)
   rp1$setContactInfo(contact1)
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015,1,1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
   ct$addPresentationForm("mapDigital")
#> [1] TRUE
   ct$addCitedResponsibleParty(rp1)
#> [1] TRUE
   md$setCitation(ct)
   md$setDescription("some description")
   
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}LE_Algorithm': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryAlgorithm' is INVALID according to ISO 19139 XML schemas!