ISOImageryBand

ISOImageryBand

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery band

Fields

bandBoundaryDefinition

[ISOImageryBandDefinition]

nominalSpatialResolution

[numeric]

transferFunctionType

[ISOImageryTransferFunctionType]

transmittedPolarisation

[ISOImageryPolarisationOrientation]

detectedPolarisation

[ISOImageryPolarisationOrientation]

Methods

new(xml)

This method is used to instantiate an ISOImageryBand

setBandBoundaryDefinition(definition)

Set the band boundary definition, object of class ISOImageryBandDefinition or character among values available by ISOImageryBandDefinition$values()

setNominalSpatialResolution(resolution)

Set the nominal spatial reosolution, object of class numeric

setTransmittedPolarisation(polarisation)

Set the transmitted polarisation, object of class ISOImageryPolarisationOrientation or character among values returned by ISOImageryPolarisationOrientation$values()

setDetectedPolarisation(polarisation)

Set the detected polarisation, object of class ISOImageryPolarisationOrientation or character among values returned by ISOImageryPolarisationOrientation$values()

Methods inherited from ISOBand

setMaxValue(maxValue)

Sets the maximum value

setMinValue(minValue)

Sets the minimum value

setUnits(units)

Sets the unit, object of class GMLUnitDefinition

setPeakResponse(peakResponse)

Sets the peak response

setBitsPerValue(bitsPerValue)

Sets the bits per value

setToneGradation

Sets the tone gradation

setScaleFactor(scaleFactor)

Sets the scale factor

setOffset(offset)

Sets the offset

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

   #create band range dimension
   md <- ISOImageryBand$new()
   md$setSequenceIdentifier(ISOMemberName$new(aName = "name", attributeType = "type"))
   md$setDescriptor("descriptor")
   md$setMaxValue(10)
   md$setMinValue(1)
   gml <- GMLBaseUnit$new(id = "ID")
   gml$setDescriptionReference("someref")
   gml$setIdentifier("identifier", "codespace")
   gml$addName("name1", "codespace")
#> [1] TRUE
   gml$addName("name2", "codespace")
#> [1] TRUE
   gml$setQuantityTypeReference("someref")
   gml$setCatalogSymbol("symbol")
   gml$setUnitsSystem("somelink")
   md$setUnits(gml)
   md$setPeakResponse(9)
   md$setBitsPerValue(5)
   md$setToneGradation(100)
   md$setScaleFactor(1)
   md$setOffset(4)
   
   md$setBandBoundaryDefinition("fiftyPercent")
   md$setNominalSpatialResolution(14.5)
   md$setTransferFunctionType("linear")
   md$setTransmittedPolarisation("horizontal")
   md$setDetectedPolarisation("horizontal")
   
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Band': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryBand' is INVALID according to ISO 19139 XML schemas!