ISOImageryGCPCollection

ISOImageryGCPCollection

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery gcp collection

Methods

new(xml)

This method is used to instantiate an ISOImageryGCPCollection

setCollectionIdentification(id)

Set the identifier, object of class integer

setCollectionName(name, locales)

Sets a name (object of class "character"). Locale names can be specified as list with the locales argument.

setCoordinateReferenceSystem(crs)

Sets the crs, object of class ISOReferenceSystem

addGCP(gcp)

Adds a GCP, object of class ISOImageryGCP

delGCP(gcp)

Deletes a GCP, object of class ISOImageryGCP

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

  md <- ISOImageryGCP$new()
  require(sf)
#> Loading required package: sf
#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1; sf_use_s2() is TRUE
  pt <- sf::st_point(c(1,1))
  md$setGeographicCoordinates(sfg = pt) 
  xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_GCP': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryGCP' is INVALID according to ISO 19139 XML schemas!