ISOImageryGeorectified

ISOImageryGeorectified

Format

R6Class object.

Value

Object of R6Class for modelling an ISO image Georectified

Fields

checkPoint

[list of ISOImageryGCP]

Methods

new(xml,value)

This method is used to instantiate an ISOImageryGeorectified

addCheckPoint(sfg,m)

Adds a check point, either an object of class 'sfg' (from sf) or a 'matrix'

delCheckPoint(sfg,m)

Deletes a check point, either an object of class 'sfg' (from sf) or a 'matrix'

Methods inherited from ISOGridSpatialRepresentation

See ISOGridSpatialRepresentation

Methods inherited from ISOGeorectified

See ISOGeorectified

References

ISO 19115-2:2009 - Geographic information -- Metadata -- Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>