ISOImageryPlatform

ISOImageryPlatform

ISOImageryPlatform

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery platform

References

- 19139 https://schemas.isotc211.org/19115/-2/gmi/1.0/gmi/#element_MI_Instrument

- 19115-3 https://schemas.isotc211.org/19115/-3/mac/2.0/mac/#element_MI_Instrument

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryInstrument

Public fields

citation

citation [0..*]: ISOCitation

identifier

identifier [1..1]: ISOMetaIdentifier

type

type [1..1]: character|ISOLocalisedCharacterString

description

description [0..1]: character|ISOLocalisedCharacterString

mountedOn

mountedOn [0..*]: ISOImageryPlatform

otherPropertyType

otherPropertyType [0..1] : ISORecordType (=> ISO 19115-3)

otherProperty

otherProperty [0..1] : ISORecord (=> ISO 19115-3)

sensor

sensor [0..*] : ISOImagerySensor (=> ISO 19115-3)

history

history [0..*] : ISOInstrumentationEventList (=> ISO 19115-3)

Methods

Public methods

ISOImageryInstrument$new()

ISOImageryInstrument$addCitation()

ISOImageryInstrument$delCitation()

ISOImageryInstrument$setIdentifier()

ISOImageryInstrument$setType()

ISOImageryInstrument$setDescription()

ISOImageryInstrument$addPlatform()

ISOImageryInstrument$delPlatform()

ISOImageryInstrument$setOtherPropertyType()

ISOImageryInstrument$setOtherProperty()

ISOImageryInstrument$addSensor()

ISOImageryInstrument$delSensor()

ISOImageryInstrument$addInstrumentationEventList()

ISOImageryInstrument$delInstrumentationEventList()

ISOImageryInstrument$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOImageryInstrument$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method addCitation()

Adds citation

Usage

ISOImageryInstrument$addCitation(citation)

Arguments

citation

object of class ISOCitation

Returns

TRUE if added, FALSE otherwise

Method delCitation()

Deletes citation

Usage

ISOImageryInstrument$delCitation(citation)

Arguments

citation

object of class ISOCitation

Returns

TRUE if deleted, FALSE otherwise

Method setIdentifier()

Set identifier

Usage

ISOImageryInstrument$setIdentifier(identifier)

Arguments

identifier

object of class ISOMetaIdentifier or character

Method setType()

Set type

Usage

ISOImageryInstrument$setType(type, locales = NULL)

Arguments

type

type

locales

list of localized texts. Default is NULL

Method setDescription()

Set description

Usage

ISOImageryInstrument$setDescription(description, locales = NULL)

Arguments

description

description

locales

list of localized texts. Default is NULL

Method addPlatform()

Adds platform

Usage

ISOImageryInstrument$addPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if added, FALSE otherwise

Method delPlatform()

Deletes platform

Usage

ISOImageryInstrument$delPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if deleted, FALSE otherwise

Method setOtherPropertyType()

setOtherPropertyType

Usage

ISOImageryInstrument$setOtherPropertyType(otherPropertyType)

Arguments

otherPropertyType

otherPropertyType object of class ISORecordType

Method setOtherProperty()

setOtherProperty

Usage

ISOImageryInstrument$setOtherProperty(otherProperty)

Arguments

otherProperty

otherProperty object of class ISORecord

Method addSensor()

Adds sensor

Usage

ISOImageryInstrument$addSensor(sensor)

Arguments

sensor

object of class ISOImagerySensor

Returns

TRUE if added, FALSE otherwise

Method delSensor()

Deletes sensor

Usage

ISOImageryInstrument$delSensor(sensor)

Arguments

sensor

object of class ISOImagerySensor

Returns

TRUE if deleted, FALSE otherwise

Method addInstrumentationEventList()

Adds instrumentation event list

Usage

ISOImageryInstrument$addInstrumentationEventList(instrumentEventList)

Arguments

instrumentEventList

object of class ISOInstrumentationEventList

Returns

TRUE if added, FALSE otherwise

Method delInstrumentationEventList()

Adds instrumentation event list

Usage

ISOImageryInstrument$delInstrumentationEventList(instrumentEventList)

Arguments

instrumentEventList

object of class ISOInstrumentationEventList

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryInstrument$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   md <- ISOImageryInstrument$new()
   md$setIdentifier("identifier")
   md$setType("type")
   md$setDescription("description")
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Instrument': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryInstrument' is INVALID according to ISO 19139 XML schemas!