ISOImageryObjective

ISOImageryObjective

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery objective

Fields

identifier

[ISOMetaIdentifier]

priority

[character|ISOLocalisedCharacterString]

type

[list of ISOImageryObjectiveType]

function

[list of character|ISOLocalisedCharacterString]

extent

[list of ISOExtent]

sensingInstrument

[list of ISOImageryInstrument]

pass

[list of ISOImageryPlatformPass]

objectiveOccurance

[list of ISOImageryEvent]

Methods

new(xml)

This method is used to instantiate an ISOImageryObjective

setIdentifier(identifier)

Sets an identifier, object of class character or ISOMetaIdentifier

setPriority(priority, locales)

Sets a priority (object of class "character"). Locale names can be specified as list with the locales argument.

addType(type)

Add type, object of class ISOImageryObjectiveType or 'character' among values given by ISOImageryObjectiveType$values() or any free text.

delType(type)

Deletes type, object of class ISOImageryObjectiveType or 'character' among values given by ISOImageryObjectiveType$values() or any free text.

addFunction(fun, locales)

Adds a function (object of class "character"). Locale names can be specified as list with the locales argument.

delFunction(fun, locales)

Deletes a function (object of class "character"). Locale names can be specified as list with the locales argument.

addExtent(extent)

Adds an extent, object of class ISOExtent

delExtent(extent)

Deletes an extent, object of class ISOExtent

addSensingInstrument(instrument)

Adds an instrument, object of class ISOImageryInstrument

delSensingInstrument(instrument)

Deletes an instrument, object of class ISOImageryInstrument

addPlatformPass(pass)

Adds an platform pass, object of class ISOImageryPlatformPass

delPlatformPass(pass)

Deletes an platform pass, object of class ISOImageryPlatformPass

addObjectiveOccurance(event)

Adds an objective occurance, object of class ISOImageryEvent

delObjectiveOccurance(event)

Deletes an objective occurance, object of class ISOImageryEvent

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

   #encoding
   md <- ISOImageryObjective$new()
   md$setIdentifier("identifier")
   md$setPriority("urgent")
   md$addType("survey")
#> [1] TRUE
   md$addFunction("my_function")
#> [1] TRUE
   evt <- ISOImageryEvent$new()
   evt$setIdentifier("event_1")
   evt$setTrigger("manual")
   evt$setContext("pass")
   evt$setSequence("instantaneous")
   evt$setTime(Sys.time())
   md$addObjectiveOccurance(evt)
#> [1] TRUE
   extent <- ISOExtent$new()
   bbox <- ISOGeographicBoundingBox$new(minx = -180, miny = -90, maxx = 180, maxy = 90)
   extent$setGeographicElement(bbox)
#> [1] TRUE
   time <- ISOTemporalExtent$new()
   start <- ISOdate(2000, 1, 12, 12, 59, 45)
   end <- ISOdate(2010, 8, 22, 13, 12, 43)
   tp <- GMLTimePeriod$new(beginPosition = start, endPosition = end)
   time$setTimePeriod(tp)
   extent$setTemporalElement(time)
#> [1] TRUE
   vert <- ISOVerticalExtent$new()
   vert$setMinimumValue(0)
   vert$setMaximumValue(19)
   extent$setVerticalElement(vert)
#> [1] TRUE
   md$addExtent(extent)
#> [1] TRUE
   md$sensingInstrument = NA
   md$pass = NA
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Objective': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryObjective' is INVALID according to ISO 19139 XML schemas!