ISOImageryOperation

ISOImageryOperation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery Operation

Fields

description

[character|ISOLocalisedCharacterString]

citation

[ISOCitation]

identifier

[ISOMetaIdentifier]

status

[ISOStatus]

type

[ISOImageryOperationType]

parentOperation

[ISOImageryOperation]

childOperation

[list of ISOImageryOperation]

platform

[list of ISOImageryPlatform]

objective

[list of ISOImageryObjective]

plan

[list of ISOImageryPlan]

significantEvent

[list of ISOImageryEvent]

Methods

new(xml)

This method is used to instantiate an ISOImageryOperation

setDescription(description, locales)

Sets a description (object of class "character"). Locale names can be specified as list with the locales argument.

setCitation(citation)

Sets a citation, object of class ISOCitation

setIdentifier(identifier)

Sets an identifier, object of class character or ISOMetaIdentifier

setStatus(status)

Sets status, object of class ISOStatus or 'character', among values given by ISOStatus$values() or any free text code

setType(type)

Sets the operation type, object of class ISOImageryOperationType, or 'character' amont values given by ISOImageryOperationType$values(), or any free text code.

setParentOperation(operation)

Sets the parent operation, object of class ISOImageryOperation.

addChildOperation(operation)

Adds a child operation, object of class ISOImageryOperation

delChildOperation(operation)

Deletes a child operation, object of class ISOImageryOperation

addPlatform(platform)

Adds a platform, object of class ISOImageryPlatform

delPlatform(platform)

Deletes a platform, object of class ISOImageryPlatform

addObjective(objective)

Adds an objective object of class ISOImageryObjective

delObjective(objective)

Deletes an objective, object of class ISOImageryObjective

setPlan(plan)

Sets a plan, object of class ISOImageryPlan

addSignificantEvent(event)

Adds a significant event, object of class ISOImageryEvent

delSignificantEvent(event)

Deletes a significant event, object of class ISOImageryEvent

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>