ISOImageryPlatform

ISOImageryPlatform

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery platform

Fields

citation

[list of ISOCitation]

identifier

[ISOMetaIdentifier]

description

[character|ISOLocalisedCharacterString]

sponsor

[list of ISOResponsibleParty]

instrument

[list of ISOImageryInstrument]

Methods

new(xml)

This method is used to instantiate an ISOImageryPlatform

addCitation(citation)

Adds citation, object of class ISOCitation

delCitation(citation)

Deletes a citation, object of class ISOCitation

setIdentifier(identifier)

Sets an identifier, object of class character or ISOMetaIdentifier

setDescription(description, locales)

Sets a description (object of class "character"). Locale names can be specified as list with the locales argument.

addSponsor(sponsor)

Adds a sponsor, object of class ISOResponsibleParty

delSponsor(sponsor)

Deletes a sponsor, object of class ISOResponsibleParty

addInstrument(instrument)

Add a instrument, object of class ISOImageryInstrument

delInstrument(instrument)

Deletes a instrument, object of class ISOImageryInstrument

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

   md <- ISOImageryPlatform$new()
   
   #add citation
   rp1 <- ISOResponsibleParty$new()
   rp1$setIndividualName("someone1")
   rp1$setOrganisationName("somewhere1")
   rp1$setPositionName("someposition1")
   rp1$setRole("pointOfContact")
   contact1 <- ISOContact$new()
   phone1 <- ISOTelephone$new()
   phone1$setVoice("myphonenumber1")
   phone1$setFacsimile("myfacsimile1")
   contact1$setPhone(phone1)
   address1 <- ISOAddress$new()
   address1$setDeliveryPoint("theaddress1")
   address1$setCity("thecity1")
   address1$setPostalCode("111")
   address1$setCountry("France")
   address1$setEmail("someone1@theorg.org")
   contact1$setAddress(address1)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact1$setOnlineResource(res)
   rp1$setContactInfo(contact1)
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015,1,1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
   ct$addPresentationForm("mapDigital")
#> [1] TRUE
   ct$addCitedResponsibleParty(rp1)
#> [1] TRUE
   md$addCitation(ct)
#> [1] TRUE
   
   md$setIdentifier("identifier")
   md$setDescription("some description")
   
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Platform': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryPlatform' is INVALID according to ISO 19139 XML schemas!