ISOImageryPlatformPass

ISOImageryPlatformPass

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery PlatformPass

Fields

identifier

[ISOMetaIdentifier]

extent

[?]

relatedEvent

[list of ISOImageryEvent]

Methods

new(xml)

This method is used to instantiate an ISOImageryPlatformPass

setIdentifier(identifier)

Sets an identifier, object of class character or ISOMetaIdentifier

setExtent(extent)

Set the extent

addEvent(event)

Add a event, object of class ISOImageryEvent

delEvent(event)

Deletes a event, object of class ISOImageryEvent

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

   md <- ISOImageryPlatformPass$new()
   md$setIdentifier("identifier")
   
   require(sf)
   outer = matrix(c(0,0,10,0,10,10,0,10,0,0),ncol=2, byrow=TRUE)
   hole1 = matrix(c(1,1,1,2,2,2,2,1,1,1),ncol=2, byrow=TRUE)
   hole2 = matrix(c(5,5,5,6,6,6,6,5,5,5),ncol=2, byrow=TRUE)
   pts = list(outer, hole1, hole2)
   pl = st_polygon(pts)
   md$setExtent(pl)
   
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_PlatformPass': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryPlatformPass' is INVALID according to ISO 19139 XML schemas!