ISOImageryProcessStep

ISOImageryProcessStep

ISOImageryProcessStep

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery process step

References

ISO 19115-2:2009 - Geographic information – Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOProcessStep -> ISOImageryProcessStep

Public fields

processingInformation

processingInformation [0..1]: ISOImageryProcessing

output

output [0..*]: list of ISOImagerySource

report

report [0..*]: list of ISOImageryProcessStepReport

Methods

Public methods

ISOImageryProcessStep$new()

ISOImageryProcessStep$setProcessingInformation()

ISOImageryProcessStep$addOutput()

ISOImageryProcessStep$delOutput()

ISOImageryProcessStep$addReport()

ISOImageryProcessStep$delReport()

ISOImageryProcessStep$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

geometa::ISOProcessStep$addProcessor()

geometa::ISOProcessStep$addSource()

geometa::ISOProcessStep$delProcessor()

geometa::ISOProcessStep$delSource()

geometa::ISOProcessStep$setDateTime()

geometa::ISOProcessStep$setDescription()

geometa::ISOProcessStep$setRationale()

Method new()

Initializes object

Usage

ISOImageryProcessStep$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setProcessingInformation()

Set processing info

Usage

ISOImageryProcessStep$setProcessingInformation(processingInfo)

Arguments

processingInfo

object of class ISOImageryProcessing

Method addOutput()

Adds output

Usage

ISOImageryProcessStep$addOutput(output)

Arguments

output

object of class ISOImagerySource

Returns

TRUE if added, FALSE otherwise

Method delOutput()

Deletes output

Usage

ISOImageryProcessStep$delOutput(output)

Arguments

output

object of class ISOImagerySource

Returns

TRUE if deleted, FALSE otherwise

Method addReport()

Adds report

Usage

ISOImageryProcessStep$addReport(report)

Arguments

report

object of class ISOImageryProcessStepReport

Returns

TRUE if added, FALSE otherwise

Method delReport()

Deletes report

Usage

ISOImageryProcessStep$delReport(report)

Arguments

report

object of class ISOImageryProcessStepReport

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryProcessStep$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   ps <- ISOImageryProcessStep$new()
   ps$setDescription("description")
   ps$setRationale("rationale")
   ps$setDateTime( ISOdate(2015, 1, 1, 23, 59, 59))
   rp <- ISOResponsibleParty$new()
   rp$setIndividualName("someone") #and more responsible party properties..
   ps$addProcessor(rp)
#> [1] TRUE
   
   #specific methods to ISO 19115-2
   process <- ISOImageryProcessing$new()
   
   #add citation
   rp1 <- ISOResponsibleParty$new()
   rp1$setIndividualName("someone1")
   rp1$setOrganisationName("somewhere1")
   rp1$setPositionName("someposition1")
   rp1$setRole("pointOfContact")
   contact1 <- ISOContact$new()
   phone1 <- ISOTelephone$new()
   phone1$setVoice("myphonenumber1")
   phone1$setFacsimile("myfacsimile1")
   contact1$setPhone(phone1)
   address1 <- ISOAddress$new()
   address1$setDeliveryPoint("theaddress1")
   address1$setCity("thecity1")
   address1$setPostalCode("111")
   address1$setCountry("France")
   address1$setEmail("someone1@theorg.org")
   contact1$setAddress(address1)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact1$setOnlineResource(res)
   rp1$setContactInfo(contact1)
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015,1,1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
   ct$addPresentationForm("mapDigital")
#> [1] TRUE
   ct$addCitedResponsibleParty(rp1)
#> [1] TRUE
   
   process$setIdentifier("identifier")
   process$setProcedureDescription("some description")
   process$addSoftwareReference(ct)
#> [1] TRUE
   process$addDocumentation(ct)
#> [1] TRUE
   process$setRunTimeParameters("params")
   ps$setProcessingInformation(process)
  
   #output
   trg <- ISOImagerySource$new()
   trg$setProcessedLevel("level")
   res <- ISOImageryNominalResolution$new()
   d <- ISODistance$new(value = 1, uom = "m", useUomURI = TRUE)
   res$setScanningResolution(d)
   trg$setResolution(res)
   ps$addOutput(trg)
#> [1] TRUE
   
   #report
   rep <- ISOImageryProcessStepReport$new()
   rep$setName("report")
   rep$setDescription("description")
   rep$setFileType("filetype")
   ps$addReport(rep)
#> [1] TRUE
   
   xml <- ps$encode()  
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}LE_ProcessStep': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryProcessStep' is INVALID according to ISO 19139 XML schemas!