ISOImageryRequirement

ISOImageryRequirement

ISOImageryRequirement

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery requirement

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryRequirement

Public fields

citation

citation [1..1]: ISOCitation

identifier

identifier [1..1]: ISOMetaIdentifier

requestor

requestor [0..*]: ISOResponsibleParty

recipient

recipient [0..*]: ISOResponsibleParty

priority

priority [1..1]: ISOImageryPriority

requestedDate

requestedDate [1..1]: ISOImageryRequestedDate

expiryDate

expiryDate [1..1]: POSIXt

satisfiedPlan

satisfiedPlan [0..*]: ISOImageryPlan

Methods

Public methods

ISOImageryRequirement$new()

ISOImageryRequirement$setCitation()

ISOImageryRequirement$setIdentifier()

ISOImageryRequirement$addRequestor()

ISOImageryRequirement$delRequestor()

ISOImageryRequirement$addRecipient()

ISOImageryRequirement$delRecipient()

ISOImageryRequirement$setPriority()

ISOImageryRequirement$setRequestedDate()

ISOImageryRequirement$setExpiryDate()

ISOImageryRequirement$addSatisfiedPlan()

ISOImageryRequirement$delSatisfiedPlan()

ISOImageryRequirement$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOImageryRequirement$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setCitation()

Set citation

Usage

ISOImageryRequirement$setCitation(citation)

Arguments

citation

object of class ISOCitation

Method setIdentifier()

Set identifier

Usage

ISOImageryRequirement$setIdentifier(identifier)

Arguments

identifier

object of class ISOMetaIdentifier or character

Method addRequestor()

Adds requestor

Usage

ISOImageryRequirement$addRequestor(requestor)

Arguments

requestor

object of class ISOResponsibleParty

Returns

TRUE if added, FALSE otherwise

Method delRequestor()

Deletes requestor

Usage

ISOImageryRequirement$delRequestor(requestor)

Arguments

requestor

object of class ISOResponsibleParty

Returns

TRUE if deleted, FALSE otherwise

Method addRecipient()

Adds recipient

Usage

ISOImageryRequirement$addRecipient(recipient)

Arguments

recipient

object of class ISOResponsibleParty

Returns

TRUE if added, FALSE otherwise

Method delRecipient()

Deletes recipient

Usage

ISOImageryRequirement$delRecipient(recipient)

Arguments

recipient

object of class ISOResponsibleParty

Returns

TRUE if deleted, FALSE otherwise

Method setPriority()

Set priority

Usage

ISOImageryRequirement$setPriority(priority)

Arguments

priority

object of class ISOImageryPriority pr any character among values returned by ISOImageryPriority$values()

Method setRequestedDate()

Set requested date

Usage

ISOImageryRequirement$setRequestedDate(date)

Arguments

date

object of class ISOImageryRequestedDate

Method setExpiryDate()

Set expiry date

Usage

ISOImageryRequirement$setExpiryDate(date)

Arguments

date

object of class POSIXct

Method addSatisfiedPlan()

Adds satisfied plan

Usage

ISOImageryRequirement$addSatisfiedPlan(plan)

Arguments

plan

object of class ISOImageryPlan

Returns

TRUE if added, FALSE otherwise

Method delSatisfiedPlan()

Deletes satisfied plan

Usage

ISOImageryRequirement$delSatisfiedPlan(plan)

Arguments

plan

object of class ISOImageryPlan

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryRequirement$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   md <- ISOImageryRequirement$new()
   md$setIdentifier("identifier")
   #add citation
   rp1 <- ISOResponsibleParty$new()
   rp1$setIndividualName("someone1")
   rp1$setOrganisationName("somewhere1")
   rp1$setPositionName("someposition1")
   rp1$setRole("pointOfContact")
   contact1 <- ISOContact$new()
   phone1 <- ISOTelephone$new()
   phone1$setVoice("myphonenumber1")
   phone1$setFacsimile("myfacsimile1")
   contact1$setPhone(phone1)
   address1 <- ISOAddress$new()
   address1$setDeliveryPoint("theaddress1")
   address1$setCity("thecity1")
   address1$setPostalCode("111")
   address1$setCountry("France")
   address1$setEmail("someone1@theorg.org")
   contact1$setAddress(address1)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact1$setOnlineResource(res)
   rp2 <- ISOResponsibleParty$new()
   rp2$setIndividualName("someone2")
   rp2$setOrganisationName("somewhere2")
   rp2$setPositionName("someposition2")
   rp2$setRole("pointOfContact")
   contact2 <- ISOContact$new()
   phone2 <- ISOTelephone$new()
   phone2$setVoice("myphonenumber2")
   phone2$setFacsimile("myfacsimile2")
   contact1$setPhone(phone2)
   address2 <- ISOAddress$new()
   address2$setDeliveryPoint("theaddress2")
   address2$setCity("thecity2")
   address2$setPostalCode("111")
   address2$setCountry("France")
   address2$setEmail("someone2@theorg.org")
   contact2$setAddress(address2)
   contact2$setOnlineResource(res)
   rp2$setContactInfo(contact2)
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015,1,1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
   ct$addPresentationForm("mapDigital")
#> [1] TRUE
   ct$addCitedResponsibleParty(rp1)
#> [1] TRUE
   md$setCitation(ct)
   md$addRequestor(rp1)
#> [1] TRUE
   md$addRecipient(rp2)
#> [1] TRUE
   md$setPriority("highImportance")
   
   rd <- ISOImageryRequestedDate$new()
   rd$setRequestedDateOfCollection(Sys.time())
   rd$setLatestAcceptableDate(Sys.time())
   md$setRequestedDate(rd)
   md$setExpiryDate(Sys.time())
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Requirement': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryRequirement' is INVALID according to ISO 19139 XML schemas!