ISOKeywords

ISOKeywords

Format

R6Class object.

Value

Object of R6Class for modelling a ISO set of keywords

Fields

keyword

[character] keyword(s)

type

[ISOKeywordType] keyword type

thesaurusName

[ISOCitation] thesaurus citation

Methods

new(xml)

This method is used to instantiate an ISOKeywords

addKeyword(keyword, locales)

Adds a keyword. Locale names can be specified as list with the locales argument.

delKeyword(keyword, locales)

Deletes a keyword. Locale names can be specified as list with the locales argument. Local names should match those of the keyword to be deleted, otherwise nothing will be deleted.

setKeywordType(keywordType)

Sets the keyword type

setThesaurusName(thesaurusName)

Sets the thesaurus name

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

  #a basic keyword set
  md <- ISOKeywords$new()
  md$addKeyword("keyword1")
#> [1] TRUE
  md$addKeyword("keyword2")
#> [1] TRUE
  md$setKeywordType("theme")
  th <- ISOCitation$new()
  th$setTitle("General")
  md$setThesaurusName(th)
  xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmd}CI_Citation': Missing child element(s). Expected is one of ( {http://www.isotc211.org/2005/gmd}alternateTitle, {http://www.isotc211.org/2005/gmd}date ) at line 12. 
#> [geometa][WARN] Object 'ISOKeywords' is INVALID according to ISO 19139 XML schemas! 
  
  #a keyword set with anchors
  md <- ISOKeywords$new()
  kwd1 <- ISOAnchor$new(
    name = "keyword1",
    href = "http://myvocabulary.geometa/keyword1"
  )
  md$addKeyword(kwd1)
#> [1] TRUE
  kwd2 <- ISOAnchor$new(
    name = "keyword2",
    href = "http://myvocabulary.geometa/keyword2"
  )
  md$addKeyword(kwd2)
#> [1] TRUE
  md$setKeywordType("theme")
  xml <- md$encode()
#> [geometa][INFO] Object 'ISOKeywords' is VALID according to ISO 19139 XML schemas! 
  
  #Example for INSPIRE (GEMET Spatial Data Theme)
  inspire_kwd <- ISOKeywords$new()
  anc1 <- ISOAnchor$new(
    name = "Environmental monitoring facilities",
    href = "http://inspire.ec.europa.eu/theme/ef"
  )
  inspire_kwd$addKeyword(anc1)
#> [1] TRUE
  inspire_kwd$setKeywordType("theme")
  th <- ISOCitation$new()
  th$setTitle(
    ISOAnchor$new(
     name = "GEMET - INSPIRE themes, version 1.0",
     href="http://www.eionet.europa.eu/gemet/inspire_themes"
    )
  )
  inspire_date <- ISODate$new()
  inspire_date$setDate(as.Date("2008-06-01"))
  inspire_date$setDateType("publication")
  th$addDate(inspire_date)
  inspire_kwd$setThesaurusName(th)