ISOLocale

ISOLocale

ISOLocale

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Locale

References

- ISO 19139 https://schemas.isotc211.org/19139/-/gmd/1.0/gmd/#element_PT_Locale

- ISO 19115-3 https://schemas.isotc211.org/19115/-3/lan/1.0/lan/#element_PT_Locale

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOLocale

Public fields

languageCode

languageCode [1..1]: ISOLanguage (ISO 19139)

language

language [1..1]: ISOLanguage (ISO 19115-3)

country

country [0..1]: ISOCountry

characterEncoding

characterEncoding [1..1]: ISOCharacterSet

Methods

Public methods

ISOLocale$new()

ISOLocale$setId()

ISOLocale$setLanguage()

ISOLocale$setCountry()

ISOLocale$setCharacterSet()

ISOLocale$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOLocale$new(
  xml = NULL,
  id = NULL,
  language = NULL,
  country = NULL,
  characterEncoding = NULL
)

Arguments

xml

object of class XMLInternalNode-class

id

id

language

language

country

country

characterEncoding

characterEncoding

Method setId()

Set ID

Usage

ISOLocale$setId(id)

Arguments

id

id

Method setLanguage()

Set language

Usage

ISOLocale$setLanguage(language)

Arguments

language

object of class ISOLanguage or any character among values returned by ISOLanguage$values()

Method setCountry()

Set country

Usage

ISOLocale$setCountry(country)

Arguments

country

object of class ISOCountry or any character among values returned by ISOCountry$values() or any other ISO-2 country code

Method setCharacterSet()

Set character set

Usage

ISOLocale$setCharacterSet(charset)

Arguments

charset

object of class ISOCharacterSet or any character among values returned by ISOCharacterSet$values()

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOLocale$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  loc <- ISOLocale$new()
  loc$setId("eng")
  loc$setLanguage("eng")
  loc$setCountry("UK")
  loc$setCharacterSet("utf8")