ISOOperationMetadata

ISOOperationMetadata

ISOOperationMetadata

Format

R6Class object.

Value

Object of R6Class for modelling an ISOOperationMetadata

References

ISO 19119:2005 - Geographic information -- Services

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOOperationMetadata

Public fields

operationName

operationName [1..1]: character

DCP

DCP [1..*]: ISODCPList

operationDescription

operationDescription [0..1]: character

invocationName

invocationName [0..1]: character

parameters

parameters [0..*]: ISOParameter

connectPoint

connectPoint [1..*]: ISOOnlineResource

dependsOn

dependsOn [0..*]: ISOOperationMetadata

Methods

Public methods

ISOOperationMetadata$new()

ISOOperationMetadata$setOperationName()

ISOOperationMetadata$addDCP()

ISOOperationMetadata$delDCP()

ISOOperationMetadata$setOperationDescription()

ISOOperationMetadata$setInvocationName()

ISOOperationMetadata$addParameter()

ISOOperationMetadata$delParameter()

ISOOperationMetadata$addConnectPoint()

ISOOperationMetadata$delConnectPoint()

ISOOperationMetadata$addDependentOperationMetadata()

ISOOperationMetadata$delDependentOperationMetadata()

ISOOperationMetadata$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOOperationMetadata$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setOperationName()

Set operation name

Usage

ISOOperationMetadata$setOperationName(operationName, locales = NULL)

Arguments

operationName

operation name

locales

list of localized texts. Default is NULL

Method addDCP()

Adds DCP

Usage

ISOOperationMetadata$addDCP(dcp)

Arguments

dcp

object of class ISODCPList or any character among values returned by ISODCPList$values()

Returns

TRUE if added, FALSE otherwise

Method delDCP()

Deletes DCP

Usage

ISOOperationMetadata$delDCP(dcp)

Arguments

dcp

object of class ISODCPList or any character among values returned by ISODCPList$values()

Returns

TRUE if deleted, FALSE otherwise

Method setOperationDescription()

Set operation description

Usage

ISOOperationMetadata$setOperationDescription(
  operationDescription,
  locales = NULL
)

Arguments

operationDescription

operation description

locales

list of localized texts. Default is NULL

Method setInvocationName()

Set invocation name

Usage

ISOOperationMetadata$setInvocationName(invocationName, locales = NULL)

Arguments

invocationName

invocation name

locales

list of localized texts. Default is NULL

Method addParameter()

Adds parameter

Usage

ISOOperationMetadata$addParameter(parameter)

Arguments

parameter

object of class ISOParameter

Returns

TRUE if added, FALSE otherwise

Method delParameter()

Deletes parameter

Usage

ISOOperationMetadata$delParameter(parameter)

Arguments

parameter

object of class ISOParameter

Returns

TRUE if deleted, FALSE otherwise

Method addConnectPoint()

Adds connection point

Usage

ISOOperationMetadata$addConnectPoint(connectPoint)

Arguments

connectPoint

object of class ISOOnlineResource

Returns

TRUE if added, FALSE otherwise

Method delConnectPoint()

Deletes connection point

Usage

ISOOperationMetadata$delConnectPoint(connectPoint)

Arguments

connectPoint

object of class ISOOnlineResource

Returns

TRUE if deleted, FALSE otherwise

Method addDependentOperationMetadata()

Adds operation metadata

Usage

ISOOperationMetadata$addDependentOperationMetadata(operationMetadata)

Arguments

operationMetadata

object of class ISOOperationMetadata

Returns

TRUE if added, FALSE otherwise

Method delDependentOperationMetadata()

Deletes operation metadata

Usage

ISOOperationMetadata$delDependentOperationMetadata(operationMetadata)

Arguments

operationMetadata

object of class ISOOperationMetadata

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOOperationMetadata$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  md <- ISOOperationMetadata$new()
  xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/srv}SV_OperationMetadata': Missing child element(s). Expected is ( {http://www.isotc211.org/2005/srv}operationName ) at line 1. 
#> [geometa][WARN] Object 'ISOOperationMetadata' is INVALID according to ISO 19139 XML schemas!