ISOSRVServiceIdentification19115_3

ISOSRVServiceIdentification19115_3

ISOSRVServiceIdentification19115_3

Format

R6Class object.

Value

Object of R6Class for modelling an ISO SRV ServiceIdentification in ISO 19115-3

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOIdentification -> geometa::ISOIdentification19115_3 -> geometa::ISOServiceIdentification19115_3 -> ISOSRVServiceIdentification19115_3

Public fields

serviceType

serviceType [1..1]: ISOGenericName

serviceTypeVersion

serviceTypeVersion [0..*]: character

accessProperties

accessProperties [0..1]: ISOStandardOrderProcess

restrictions

restrictions [0..1]: ISOConstraints

keywords

keywords [0..*]: ISOKeywords

extent

extent [0..*]: ISOExtent

coupledResource

coupledResource [0..*]: ISOCoupledResource

couplingType

couplingType [1..1]: ISOCouplingType

containsOperations

containsOperations [1..*]: ISOOperationMetadata

operatesOn

operatesOn [0..*]: ISODataIdentification

Methods

Public methods

ISOSRVServiceIdentification19115_3$new()

ISOSRVServiceIdentification19115_3$setServiceType()

ISOSRVServiceIdentification19115_3$addServiceTypeVersion()

ISOSRVServiceIdentification19115_3$delServiceTypeVersion()

ISOSRVServiceIdentification19115_3$setAccessProperties()

ISOSRVServiceIdentification19115_3$setRestrictions()

ISOSRVServiceIdentification19115_3$addKeywords()

ISOSRVServiceIdentification19115_3$delKeywords()

ISOSRVServiceIdentification19115_3$addExtent()

ISOSRVServiceIdentification19115_3$delExtent()

ISOSRVServiceIdentification19115_3$addCoupledResource()

ISOSRVServiceIdentification19115_3$delCoupledResource()

ISOSRVServiceIdentification19115_3$setCouplingType()

ISOSRVServiceIdentification19115_3$addOperationMetadata()

ISOSRVServiceIdentification19115_3$delOperationMetadata()

ISOSRVServiceIdentification19115_3$addOperatesOn()

ISOSRVServiceIdentification19115_3$delOperatesOn()

ISOSRVServiceIdentification19115_3$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

geometa::ISOIdentification19115_3$addAdditionalDocumentation()

geometa::ISOIdentification19115_3$addAssociatedResource()

geometa::ISOIdentification19115_3$addCredit()

geometa::ISOIdentification19115_3$addFormat()

geometa::ISOIdentification19115_3$addGraphicOverview()

geometa::ISOIdentification19115_3$addPointOfContact()

geometa::ISOIdentification19115_3$addResourceConstraints()

geometa::ISOIdentification19115_3$addResourceMaintenance()

geometa::ISOIdentification19115_3$addResourceSpecificUsage()

geometa::ISOIdentification19115_3$addSpatialRepresentationType()

geometa::ISOIdentification19115_3$addSpatialResolution()

geometa::ISOIdentification19115_3$addStatus()

geometa::ISOIdentification19115_3$addTemporalResolution()

geometa::ISOIdentification19115_3$addTopicCategory()

geometa::ISOIdentification19115_3$delAdditionalDocumentation()

geometa::ISOIdentification19115_3$delAssociatedResource()

geometa::ISOIdentification19115_3$delCredit()

geometa::ISOIdentification19115_3$delFormat()

geometa::ISOIdentification19115_3$delGraphicOverview()

geometa::ISOIdentification19115_3$delPointOfContact()

geometa::ISOIdentification19115_3$delResourceConstraints()

geometa::ISOIdentification19115_3$delResourceMaintenance()

geometa::ISOIdentification19115_3$delResourceSpecificUsage()

geometa::ISOIdentification19115_3$delSpatialRepresentationType()

geometa::ISOIdentification19115_3$delSpatialResolution()

geometa::ISOIdentification19115_3$delStatus()

geometa::ISOIdentification19115_3$delTemporalResolution()

geometa::ISOIdentification19115_3$delTopicCategory()

geometa::ISOIdentification19115_3$setAbstract()

geometa::ISOIdentification19115_3$setCitation()

geometa::ISOIdentification19115_3$setProcessingLevel()

geometa::ISOIdentification19115_3$setPurpose()

Method new()

Initializes object

Usage

ISOSRVServiceIdentification19115_3$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setServiceType()

Set service type

Usage

ISOSRVServiceIdentification19115_3$setServiceType(serviceType)

Arguments

serviceType

object of class ISOLocalName, ISOScopedName or character

Method addServiceTypeVersion()

Adds service type version

Usage

ISOSRVServiceIdentification19115_3$addServiceTypeVersion(version)

Arguments

version

version

Returns

TRUE if added, FALSE otherwise

Method delServiceTypeVersion()

Deletes service type version

Usage

ISOSRVServiceIdentification19115_3$delServiceTypeVersion(version)

Arguments

version

version

Returns

TRUE if deleted, FALSE otherwise

Method setAccessProperties()

Set access properties

Usage

ISOSRVServiceIdentification19115_3$setAccessProperties(accessProperties)

Arguments

accessProperties

object of class ISOStandardOrderProcess

Method setRestrictions()

Set restrictions

Usage

ISOSRVServiceIdentification19115_3$setRestrictions(restrictions)

Arguments

restrictions

object of class ISOConstraints

Method addKeywords()

Adds keywords

Usage

ISOSRVServiceIdentification19115_3$addKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if added, FALSE otherwise

Method delKeywords()

Deletes keywords

Usage

ISOSRVServiceIdentification19115_3$delKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if deleted, FALSE otherwise

Method addExtent()

Adds extent

Usage

ISOSRVServiceIdentification19115_3$addExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if added, FALSE otherwise

Method delExtent()

Deletes extent

Usage

ISOSRVServiceIdentification19115_3$delExtent(extent)

Arguments

extent

object of class ISOExtent

Returns

TRUE if deleted, FALSE otherwise

Method addCoupledResource()

Adds coupled resource

Usage

ISOSRVServiceIdentification19115_3$addCoupledResource(resource)

Arguments

resource

object of class ISOCoupledResource

Returns

TRUE if added, FALSE otherwise

Method delCoupledResource()

Deletes coupled resource

Usage

ISOSRVServiceIdentification19115_3$delCoupledResource(resource)

Arguments

resource

object of class ISOCoupledResource

Returns

TRUE if deleted, FALSE otherwise

Method setCouplingType()

Set coupling type

Usage

ISOSRVServiceIdentification19115_3$setCouplingType(couplingType)

Arguments

couplingType

object of class ISOCouplingType or any character among values returned by ISOCouplingType$values()

Method addOperationMetadata()

Adds operation metadata

Usage

ISOSRVServiceIdentification19115_3$addOperationMetadata(operationMetadata)

Arguments

operationMetadata

object of class ISOOperationMetadata

Returns

TRUE if added, FALSE otherwise

Method delOperationMetadata()

Deletes operation metadata

Usage

ISOSRVServiceIdentification19115_3$delOperationMetadata(operationMetadata)

Arguments

operationMetadata

object of class ISOOperationMetadata

Returns

TRUE if deleted, FALSE otherwise

Method addOperatesOn()

Adds operates on

Usage

ISOSRVServiceIdentification19115_3$addOperatesOn(dataIdentification)

Arguments

dataIdentification

object of class ISODataIdentification

Returns

TRUE if added, FALSE otherwise

Method delOperatesOn()

Deletes operates on

Usage

ISOSRVServiceIdentification19115_3$delOperatesOn(dataIdentification)

Arguments

dataIdentification

object of class ISODataIdentification

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOSRVServiceIdentification19115_3$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.