ISOServiceIdentification

ISOServiceIdentification

Format

R6Class object.

Value

Object of R6Class for modelling an ISO ServiceIdentification

Methods inherited from ISOIdentification

setCitation(citation)

Sets an object of class ISOCitation

setAbstract(abstract, locales)

Sets an abstract (object of class "character"). Locale names can be specified as list with the locales argument.

setPurpose(purpose, locales)

Sets a purpose (object of class "character"). Locale names can be specified as list with the locales argument.

addCredit(credit, locales)

Adds a credit (object of class "character"). Locale names can be specified as list with the locales argument.

delCredit(credit, locales)

Deletes a credit (object of class "character"). Locale names can be specified as list with the locales argument.

addStatus(status)

Adds a status, as object of class "character" or class ISOStatus. If an object of class "character" is specified, it must match the accepted progress status values ISOStatus$values().

delStatus(status)

Deletes a status, as object of class "character" or class ISOStatus. If an object of class "character" is specified, it must match the accepted progress status values ISOStatus$values().

addPointOfContact(pointOfContact)

Adds an object of class ISOResponsibleParty

delPointOfContact(pointOfContact)

Deletes an object of class ISOResponsibleParty

addResourceMaintenance(resourceMaintenance)

Adds a resource maintenance information as object of class ISOMaintenanceInformation.

setResourceMaintenance(resourceMaintenance)

Sets a resource maintenance information as object of class ISOMaintenanceInformation.

delResourceMaintenance(resourceMaintenance)

Deletes a resource maintenance information as object of class ISOMaintenanceInformation.

addGraphicOverview(graphicOverview)

Adds an object of class ISOBrowseGraphic

setGraphicOverview(graphicOverview)

Sets an object of class ISOBrowseGraphic

delGraphicOverview(graphicOverview)

Deletes an object of class ISOBrowseGraphic

addKeywords(keywords)

Adds a set of keywords as object of class ISOKeywords

setKeywords(keywords)

Sets a set of keywords as object of class ISOKeywords

delKeywords(keywords)

Deletes a set of keywords as object of class ISOKeywords

addResourceConstraints(resourceConstraints)

Adds an object of class ISOLegalConstraints

setResourceConstraints(resourceConstraints)

Sets an object of class ISOLegalConstraints

addResourceConstraints(resourceConstraints)

Deletes an object of class ISOLegalConstraints

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

  #encoding
  md <- ISOServiceIdentification$new()
  md$setAbstract("abstract")
  md$setPurpose("purpose")

  #adding a point of contact
  rp <- ISOResponsibleParty$new()
  rp$setIndividualName("someone")
  rp$setOrganisationName("somewhere")
  rp$setPositionName("someposition")
  rp$setRole("pointOfContact")
  contact <- ISOContact$new()
  phone <- ISOTelephone$new()
  phone$setVoice("myphonenumber")
  phone$setFacsimile("myfacsimile")
  contact$setPhone(phone)
  address <- ISOAddress$new()
  address$setDeliveryPoint("theaddress")
  address$setCity("thecity")
  address$setPostalCode("111")
  address$setCountry("France")
  address$setEmail("someone@theorg.org")
  contact$setAddress(address)
  res <- ISOOnlineResource$new()
  res$setLinkage("http://www.somewhereovertheweb.org")
  res$setName("somename")
  contact$setOnlineResource(res)
  rp$setContactInfo(contact)
  md$addPointOfContact(rp)
#> [1] TRUE

  #citation
  ct <- ISOCitation$new()
  ct$setTitle("sometitle")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  ct$addDate(d)
  ct$setEdition("1.0")
  ct$setEditionDate(ISOdate(2015,1,1))
  ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
  ct$addPresentationForm("mapDigital")
#> [1] TRUE
  ct$addCitedResponsibleParty(rp)
#> [1] TRUE
  md$setCitation(ct)

  #graphic overview
  go <- ISOBrowseGraphic$new(
    fileName = "http://wwww.somefile.org/png",
    fileDescription = "Map Overview",
    fileType = "image/png"
  )
  md$setGraphicOverview(go)
#> [1] TRUE

  #maintenance information
  mi <- ISOMaintenanceInformation$new()
  mi$setMaintenanceFrequency("daily")
  md$setResourceMaintenance(mi)
#> [1] TRUE

  #adding legal constraints
  lc <- ISOLegalConstraints$new()
  lc$addUseLimitation("limitation1")
#> [1] TRUE
  lc$addUseLimitation("limitation2")
#> [1] TRUE
  lc$addUseLimitation("limitation3")
#> [1] TRUE
  lc$addAccessConstraint("copyright")
#> [1] TRUE
  lc$addAccessConstraint("license")
#> [1] TRUE
  lc$addUseConstraint("copyright")
#> [1] TRUE
  lc$addUseConstraint("license")
#> [1] TRUE
  md$setResourceConstraints(lc)
#> [1] TRUE

  xml <- md$encode()
#> [geometa][INFO] Object 'ISOServiceIdentification' is VALID according to ISO 19139 XML schemas!