getISOClasses

get the list of ISO classes, ie classes extending ISOAbstractObject super class, including classes eventually defined outside geometa. In case the latter is on the search path, the list of ISO classes will be cached for optimized used by geometa encoder/decoder.

getISOClasses()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOClasses()
#>   [1] "GMLAbstractCRS"                          
#>   [2] "GMLAbstractCoordinateOperation"          
#>   [3] "GMLAbstractCoordinateSystem"             
#>   [4] "GMLAbstractCoverage"                     
#>   [5] "GMLAbstractCurve"                        
#>   [6] "GMLAbstractDiscreteCoverage"             
#>   [7] "GMLAbstractFeature"                      
#>   [8] "GMLAbstractGML"                          
#>   [9] "GMLAbstractGeneralConversion"            
#>  [10] "GMLAbstractGeneralDerivedCRS"            
#>  [11] "GMLAbstractGeneralOperationParameter"    
#>  [12] "GMLAbstractGeneralParameterValue"        
#>  [13] "GMLAbstractGeometricAggregate"           
#>  [14] "GMLAbstractGeometricPrimitive"           
#>  [15] "GMLAbstractGeometry"                     
#>  [16] "GMLAbstractImplicitGeometry"             
#>  [17] "GMLAbstractObject"                       
#>  [18] "GMLAbstractReferenceableGrid"            
#>  [19] "GMLAbstractRing"                         
#>  [20] "GMLAbstractSingleCRS"                    
#>  [21] "GMLAbstractSingleOperation"              
#>  [22] "GMLAbstractSurface"                      
#>  [23] "GMLAbstractTimeGeometricPrimitive"       
#>  [24] "GMLAbstractTimeObject"                   
#>  [25] "GMLAbstractTimePrimitive"                
#>  [26] "GMLAffineCS"                             
#>  [27] "GMLBaseUnit"                             
#>  [28] "GMLCOVAbstractCoverage"                  
#>  [29] "GMLCOVExtension"                         
#>  [30] "GMLCartesianCS"                          
#>  [31] "GMLCodeType"                             
#>  [32] "GMLCompoundCRS"                          
#>  [33] "GMLConventionalUnit"                     
#>  [34] "GMLConversion"                           
#>  [35] "GMLCoordinateSystemAxis"                 
#>  [36] "GMLCylindricalCS"                        
#>  [37] "GMLDefinition"                           
#>  [38] "GMLDerivedCRS"                           
#>  [39] "GMLDerivedUnit"                          
#>  [40] "GMLElement"                              
#>  [41] "GMLEllipsoidalCS"                        
#>  [42] "GMLEnvelope"                             
#>  [43] "GMLEnvelopeWithTimePeriod"               
#>  [44] "GMLGeneralGridAxis"                      
#>  [45] "GMLGeodeticCRS"                          
#>  [46] "GMLGrid"                                 
#>  [47] "GMLGridCoverage"                         
#>  [48] "GMLGridEnvelope"                         
#>  [49] "GMLGridFunction"                         
#>  [50] "GMLLineString"                           
#>  [51] "GMLLinearCS"                             
#>  [52] "GMLLinearRing"                           
#>  [53] "GMLMultiCurve"                           
#>  [54] "GMLMultiCurveCoverage"                   
#>  [55] "GMLMultiPoint"                           
#>  [56] "GMLMultiPointCoverage"                   
#>  [57] "GMLMultiSolidCoverage"                   
#>  [58] "GMLMultiSurface"                         
#>  [59] "GMLMultiSurfaceCoverage"                 
#>  [60] "GMLObliqueCartesianCS"                   
#>  [61] "GMLOperationMethod"                      
#>  [62] "GMLOperationParameter"                   
#>  [63] "GMLOperationParameterGroup"              
#>  [64] "GMLParameterValue"                       
#>  [65] "GMLParameterValueGroup"                  
#>  [66] "GMLPoint"                                
#>  [67] "GMLPolarCS"                              
#>  [68] "GMLPolygon"                              
#>  [69] "GMLProjectedCRS"                         
#>  [70] "GMLRectifiedGrid"                        
#>  [71] "GMLRectifiedGridCoverage"                
#>  [72] "GMLReferenceableGridByArray"             
#>  [73] "GMLReferenceableGridByTransformation"    
#>  [74] "GMLReferenceableGridByVectors"           
#>  [75] "GMLSphericalCS"                          
#>  [76] "GMLTemporalCRS"                          
#>  [77] "GMLTemporalCS"                           
#>  [78] "GMLTimeCS"                               
#>  [79] "GMLTimeInstant"                          
#>  [80] "GMLTimePeriod"                           
#>  [81] "GMLUnitDefinition"                       
#>  [82] "GMLUserDefinedCS"                        
#>  [83] "GMLVerticalCRS"                          
#>  [84] "GMLVerticalCS"                           
#>  [85] "ISOAbsoluteExternalPositionalAccuracy"   
#>  [86] "ISOAbstractAcquisitionInformation"       
#>  [87] "ISOAbstractAggregate"                    
#>  [88] "ISOAbstractApplicationSchemaInformation" 
#>  [89] "ISOAbstractCarrierOfCharacteristics"     
#>  [90] "ISOAbstractCatalogue"                    
#>  [91] "ISOAbstractCitation"                     
#>  [92] "ISOAbstractCompleteness"                 
#>  [93] "ISOAbstractConstraints"                  
#>  [94] "ISOAbstractContentInformation"           
#>  [95] "ISOAbstractDataQuality"                  
#>  [96] "ISOAbstractDistribution"                 
#>  [97] "ISOAbstractExtent"                       
#>  [98] "ISOAbstractFormat"                       
#>  [99] "ISOAbstractGenericName"                  
#> [100] "ISOAbstractLineageInformation"           
#> [101] "ISOAbstractLogicalConsistency"           
#> [102] "ISOAbstractMDContentInformation"         
#> [103] "ISOAbstractMaintenanceInformation"       
#> [104] "ISOAbstractMetadata"                     
#> [105] "ISOAbstractMetadataExtension"            
#> [106] "ISOAbstractOnlineResource"               
#> [107] "ISOAbstractParameter"                    
#> [108] "ISOAbstractParty"                        
#> [109] "ISOAbstractPlatform"                     
#> [110] "ISOAbstractPortrayalCatalogueInformation"
#> [111] "ISOAbstractPositionalAccuracy"           
#> [112] "ISOAbstractPropertyType"                 
#> [113] "ISOAbstractQualityElement"               
#> [114] "ISOAbstractRSReferenceSystem"            
#> [115] "ISOAbstractReferenceSystem"              
#> [116] "ISOAbstractResourceDescription"          
#> [117] "ISOAbstractResponsibility"               
#> [118] "ISOAbstractResult"                       
#> [119] "ISOAbstractSpatialRepresentation"        
#> [120] "ISOAbstractSpatialResolution"            
#> [121] "ISOAbstractStandardOrderProcess"         
#> [122] "ISOAbstractTemporalAccuracy"             
#> [123] "ISOAbstractThematicAccuracy"             
#> [124] "ISOAbstractTypedDate"                    
#> [125] "ISOAccuracyOfATimeMeasurement"           
#> [126] "ISOAddress"                              
#> [127] "ISOAggregateInformation"                 
#> [128] "ISOAnchor"                               
#> [129] "ISOAngle"                                
#> [130] "ISOApplicationSchemaInformation"         
#> [131] "ISOAssociatedResource"                   
#> [132] "ISOAssociation"                          
#> [133] "ISOAssociationRole"                      
#> [134] "ISOAssociationType"                      
#> [135] "ISOBand"                                 
#> [136] "ISOBaseBoolean"                          
#> [137] "ISOBaseCharacterString"                  
#> [138] "ISOBaseDate"                             
#> [139] "ISOBaseDateTime"                         
#> [140] "ISOBaseDecimal"                          
#> [141] "ISOBaseInteger"                          
#> [142] "ISOBaseReal"                             
#> [143] "ISOBinary"                               
#> [144] "ISOBinding"                              
#> [145] "ISOBoundAssociationRole"                 
#> [146] "ISOBoundFeatureAttribute"                
#> [147] "ISOBoundingPolygon"                      
#> [148] "ISOBrowseGraphic"                        
#> [149] "ISOCarrierOfCharacteristics"             
#> [150] "ISOCellGeometry"                         
#> [151] "ISOCharacterSet"                         
#> [152] "ISOCitation"                             
#> [153] "ISOCitationSeries"                       
#> [154] "ISOClassification"                       
#> [155] "ISOCodeDefinition"                       
#> [156] "ISOCodeListDictionary"                   
#> [157] "ISOCodeListValue"                        
#> [158] "ISOCodelist"                             
#> [159] "ISOCodelistCatalogue"                    
#> [160] "ISOCodelistValue"                        
#> [161] "ISOCompletenessCommission"               
#> [162] "ISOCompletenessOmission"                 
#> [163] "ISOConceptualConsistency"                
#> [164] "ISOConformanceResult"                    
#> [165] "ISOConstraint"                           
#> [166] "ISOConstraints"                          
#> [167] "ISOContact"                              
#> [168] "ISOCountry"                              
#> [169] "ISOCoupledResource"                      
#> [170] "ISOCouplingType"                         
#> [171] "ISOCoverageContentType"                  
#> [172] "ISOCoverageDescription"                  
#> [173] "ISODCPList"                              
#> [174] "ISODataFile"                             
#> [175] "ISODataIdentification"                   
#> [176] "ISODataQuality"                          
#> [177] "ISODataQualityAbstractElement"           
#> [178] "ISODataQualityScope"                     
#> [179] "ISODataSet"                              
#> [180] "ISODatatype"                             
#> [181] "ISODate"                                 
#> [182] "ISODateType"                             
#> [183] "ISODefinitionReference"                  
#> [184] "ISODefinitionSource"                     
#> [185] "ISODigitalTransferOptions"               
#> [186] "ISODimension"                            
#> [187] "ISODimensionNameType"                    
#> [188] "ISODistance"                             
#> [189] "ISODistribution"                         
#> [190] "ISODistributionUnits"                    
#> [191] "ISODistributor"                          
#> [192] "ISODomainConsistency"                    
#> [193] "ISOElementSequence"                      
#> [194] "ISOEvaluationMethodType"                 
#> [195] "ISOExtendedElementInformation"           
#> [196] "ISOExtent"                               
#> [197] "ISOFeatureAssociation"                   
#> [198] "ISOFeatureAttribute"                     
#> [199] "ISOFeatureCatalogue"                     
#> [200] "ISOFeatureCatalogueDescription"          
#> [201] "ISOFeatureOperation"                     
#> [202] "ISOFeatureType"                          
#> [203] "ISOFileName"                             
#> [204] "ISOFormat"                               
#> [205] "ISOFormatConsistency"                    
#> [206] "ISOFreeText"                             
#> [207] "ISOGeographicBoundingBox"                
#> [208] "ISOGeographicDescription"                
#> [209] "ISOGeographicExtent"                     
#> [210] "ISOGeometricObjectType"                  
#> [211] "ISOGeometricObjects"                     
#> [212] "ISOGeorectified"                         
#> [213] "ISOGeoreferenceable"                     
#> [214] "ISOGridSpatialRepresentation"            
#> [215] "ISOGriddedDataPositionalAccuracy"        
#> [216] "ISOIdentification"                       
#> [217] "ISOIdentifier"                           
#> [218] "ISOImageDescription"                     
#> [219] "ISOImageryAbstractGeolocationInformation"
#> [220] "ISOImageryAcquisitionInformation"        
#> [221] "ISOImageryAlgorithm"                     
#> [222] "ISOImageryBand"                          
#> [223] "ISOImageryBandDefinition"                
#> [224] "ISOImageryContext"                       
#> [225] "ISOImageryCoverageDescription"           
#> [226] "ISOImageryCoverageResult"                
#> [227] "ISOImageryEnvironmentalRecord"           
#> [228] "ISOImageryEvent"                         
#> [229] "ISOImageryGCP"                           
#> [230] "ISOImageryGCPCollection"                 
#> [231] "ISOImageryGeometryType"                  
#> [232] "ISOImageryGeorectified"                  
#> [233] "ISOImageryGeoreferenceable"              
#> [234] "ISOImageryImageDescription"              
#> [235] "ISOImageryInstrument"                    
#> [236] "ISOImageryMetadata"                      
#> [237] "ISOImageryNominalResolution"             
#> [238] "ISOImageryObjective"                     
#> [239] "ISOImageryObjectiveType"                 
#> [240] "ISOImageryOperation"                     
#> [241] "ISOImageryOperationType"                 
#> [242] "ISOImageryPlan"                          
#> [243] "ISOImageryPlatform"                      
#> [244] "ISOImageryPlatformPass"                  
#> [245] "ISOImageryPolarisationOrientation"       
#> [246] "ISOImageryPriority"                      
#> [247] "ISOImageryProcessStep"                   
#> [248] "ISOImageryProcessStepReport"             
#> [249] "ISOImageryProcessing"                    
#> [250] "ISOImageryRangeElementDescription"       
#> [251] "ISOImageryRequestedDate"                 
#> [252] "ISOImageryRequirement"                   
#> [253] "ISOImagerySensorType"                    
#> [254] "ISOImagerySequence"                      
#> [255] "ISOImagerySource"                        
#> [256] "ISOImageryTransferFunctionType"          
#> [257] "ISOImageryTrigger"                       
#> [258] "ISOImageryUsability"                     
#> [259] "ISOImagingCondition"                     
#> [260] "ISOIndividual"                           
#> [261] "ISOInheritanceRelation"                  
#> [262] "ISOInitiative"                           
#> [263] "ISOInitiativeType"                       
#> [264] "ISOKeywordClass"                         
#> [265] "ISOKeywordType"                          
#> [266] "ISOKeywords"                             
#> [267] "ISOLanguage"                             
#> [268] "ISOLegalConstraints"                     
#> [269] "ISOLength"                               
#> [270] "ISOLineage"                              
#> [271] "ISOListedValue"                          
#> [272] "ISOLocalName"                            
#> [273] "ISOLocale"                               
#> [274] "ISOLocaleContainer"                      
#> [275] "ISOLocalisedCharacterString"             
#> [276] "ISOMLCodeDefinition"                     
#> [277] "ISOMLCodeListDictionary"                 
#> [278] "ISOMaintenanceFrequency"                 
#> [279] "ISOMaintenanceInformation"               
#> [280] "ISOMeasure"                              
#> [281] "ISOMedium"                               
#> [282] "ISOMediumFormat"                         
#> [283] "ISOMediumName"                           
#> [284] "ISOMemberName"                           
#> [285] "ISOMetaIdentifier"                       
#> [286] "ISOMetadata"                             
#> [287] "ISOMetadataExtensionInformation"         
#> [288] "ISOMetadataScope"                        
#> [289] "ISOMimeFileType"                         
#> [290] "ISOMultiplicity"                         
#> [291] "ISOMultiplicityRange"                    
#> [292] "ISONonQuantitativeAttributeAccuracy"     
#> [293] "ISOObligation"                           
#> [294] "ISOOnLineFunction"                       
#> [295] "ISOOnlineResource"                       
#> [296] "ISOOperationMetadata"                    
#> [297] "ISOOrganisation"                         
#> [298] "ISOOtherAggregate"                       
#> [299] "ISOParameter"                            
#> [300] "ISOParameterDirection"                   
#> [301] "ISOPeriodDuration"                       
#> [302] "ISOPixelOrientation"                     
#> [303] "ISOPlatform"                             
#> [304] "ISOPortrayalCatalogueReference"          
#> [305] "ISOPresentationForm"                     
#> [306] "ISOProcessStep"                          
#> [307] "ISOProductionSeries"                     
#> [308] "ISOProgress"                             
#> [309] "ISOPropertyType"                         
#> [310] "ISOQuantitativeAttributeAccuracy"        
#> [311] "ISOQuantitativeResult"                   
#> [312] "ISORangeDimension"                       
#> [313] "ISORecord"                               
#> [314] "ISORecordType"                           
#> [315] "ISOReferenceIdentifier"                  
#> [316] "ISOReferenceSystem"                      
#> [317] "ISOReferenceSystemType"                  
#> [318] "ISORelativeInternalPositionalAccuracy"   
#> [319] "ISOReleasability"                        
#> [320] "ISORepresentativeFraction"               
#> [321] "ISOResolution"                           
#> [322] "ISOResponsibility"                       
#> [323] "ISOResponsibleParty"                     
#> [324] "ISORestriction"                          
#> [325] "ISORole"                                 
#> [326] "ISORoleType"                             
#> [327] "ISOSRVServiceIdentification"             
#> [328] "ISOSampleDimension"                      
#> [329] "ISOScale"                                
#> [330] "ISOScope"                                
#> [331] "ISOScopeCode"                            
#> [332] "ISOScopeDescription"                     
#> [333] "ISOScopedName"                           
#> [334] "ISOSecurityConstraints"                  
#> [335] "ISOSensor"                               
#> [336] "ISOSeries"                               
#> [337] "ISOServiceIdentification"                
#> [338] "ISOSource"                               
#> [339] "ISOSpatialRepresentation"                
#> [340] "ISOSpatialRepresentationType"            
#> [341] "ISOSpatialTemporalExtent"                
#> [342] "ISOStandardOrderProcess"                 
#> [343] "ISOStatus"                               
#> [344] "ISOStereoMate"                           
#> [345] "ISOTelephone"                            
#> [346] "ISOTelephoneType"                        
#> [347] "ISOTemporalConsistency"                  
#> [348] "ISOTemporalExtent"                       
#> [349] "ISOTemporalValidity"                     
#> [350] "ISOThematicClassificationCorrectness"    
#> [351] "ISOTopicCategory"                        
#> [352] "ISOTopologicalConsistency"               
#> [353] "ISOTopologyLevel"                        
#> [354] "ISOTypeName"                             
#> [355] "ISOURI"                                  
#> [356] "ISOURL"                                  
#> [357] "ISOUnlimitedInteger"                     
#> [358] "ISOUomIdentifier"                        
#> [359] "ISOUsage"                                
#> [360] "ISOVectorSpatialRepresentation"          
#> [361] "ISOVerticalExtent"                       
#> [362] "SWEAbstractDataComponent"                
#> [363] "SWEAbstractEncoding"                     
#> [364] "SWEAbstractObject"                       
#> [365] "SWEAbstractSWE"                          
#> [366] "SWEAbstractSWEIdentifiable"              
#> [367] "SWEAbstractSimpleComponent"              
#> [368] "SWECategory"                             
#> [369] "SWECategoryRange"                        
#> [370] "SWECount"                                
#> [371] "SWECountRange"                           
#> [372] "SWEDataRecord"                           
#> [373] "SWEElement"                              
#> [374] "SWENilValues"                            
#> [375] "SWEQuantity"                             
#> [376] "SWEQuantityRange"                        
#> [377] "SWEText"                                 
#> [378] "SWETextEncoding"                         
#> [379] "SWETime"                                 
#> [380] "SWETimeRange"                            
#> [381] "SWEXMLEncoding"