getISOCodelist

getISOCodelist allows to get a registered ISO codelist by id registered in geometa

getISOCodelist(id)

Arguments

id

identifier of the codelist

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOCodelist(id = "LanguageCode")
#> <ISOCodelist>
#> ....|-- id: LanguageCode
#> ....|-- refFile: ISO-639-2_utf-8.txt
#> ....|-- codeSpace: ISO 639-2
#> ....|-- identifier: LanguageCode
#> ....|-- description: Language : ISO 639-2 (3 characters)
#> ....|-- codeEntry: c("aar", "abk", "ace", "ach", "ada", "ady", "afa", "afh", "afr", "ain", "aka", "akk", "alb", "sqi", "ale", "alg", "alt", "amh", "ang", "anp", "apa", "ara", "arc", "arg", "arm", "hye", "arn", "arp", "art", "arw", "asm", "ast", "ath", "aus", "ava", "ave", "awa", "aym", "aze", "bad", "bai", "bak", "bal", "bam", "ban", "baq", "eus", "bas", "bat", "bej", "bel", "bem", "ben", "ber", "bho", "bih", "bik", "bin", "bis", "bla", "bnt", "bos", "bra", "bre", "btk", "bua", "bug", "bul", "bur", "mya", "byn", "cad", 
#> "cai", "car", "cat", "cau", "ceb", "cel", "cha", "chb", "che", "chg", "chi", "zho", "chk", "chm", "chn", "cho", "chp", "chr", "chu", "chv", "chy", "cmc", "cop", "cor", "cos", "cpe", "cpf", "cpp", "cre", "crh", "crp", "csb", "cus", "cze", "ces", "dak", "dan", "dar", "day", "del", "den", "dgr", "din", "div", "doi", "dra", "dsb", "dua", "dum", "dut", "nld", "dyu", "dzo", "efi", "egy", "eka", "elx", "eng", "enm", "epo", "est", "ewe", "ewo", "fan", "fao", "fat", "fij", "fil", "fin", "fiu", "fon", "fre", 
#> "fra", "frm", "fro", "frr", "frs", "fry", "ful", "fur", "gaa", "gay", "gba", "gem", "geo", "kat", "ger", "deu", "gez", "gil", "gla", "gle", "glg", "glv", "gmh", "goh", "gon", "gor", "got", "grb", "grc", "gre", "ell", "grn", "gsw", "guj", "gwi", "hai", "hat", "hau", "haw", "heb", "her", "hil", "him", "hin", "hit", "hmn", "hmo", "hrv", "hsb", "hun", "hup", "iba", "ibo", "ice", "isl", "ido", "iii", "ijo", "iku", "ile", "ilo", "ina", "inc", "ind", "ine", "inh", "ipk", "ira", "iro", "ita", "jav", "jbo", 
#> "jpn", "jpr", "jrb", "kaa", "kab", "kac", "kal", "kam", "kan", "kar", "kas", "kau", "kaw", "kaz", "kbd", "kha", "khi", "khm", "kho", "kik", "kin", "kir", "kmb", "kok", "kom", "kon", "kor", "kos", "kpe", "krc", "krl", "kro", "kru", "kua", "kum", "kur", "kut", "lad", "lah", "lam", "lao", "lat", "lav", "lez", "lim", "lin", "lit", "lol", "loz", "ltz", "lua", "lub", "lug", "lui", "lun", "luo", "lus", "mac", "mkd", "mad", "mag", "mah", "mai", "mak", "mal", "man", "mao", "mri", "map", "mar", "mas", "may", 
#> "msa", "mdf", "mdr", "men", "mga", "mic", "min", "mis", "mkh", "mlg", "mlt", "mnc", "mni", "mno", "moh", "mon", "mos", "mul", "mun", "mus", "mwl", "mwr", "myn", "myv", "nah", "nai", "nap", "nau", "nav", "nbl", "nde", "ndo", "nds", "nep", "new", "nia", "nic", "niu", "nno", "nob", "nog", "non", "nor", "nqo", "nso", "nub", "nwc", "nya", "nym", "nyn", "nyo", "nzi", "oci", "oji", "ori", "orm", "osa", "oss", "ota", "oto", "paa", "pag", "pal", "pam", "pan", "pap", "pau", "peo", "per", "fas", "phi", "phn", 
#> "pli", "pol", "pon", "por", "pra", "pro", "pus", "qaa-qtz", "que", "raj", "rap", "rar", "roa", "roh", "rom", "rum", "ron", "run", "rup", "rus", "sad", "sag", "sah", "sai", "sal", "sam", "san", "sas", "sat", "scn", "sco", "sel", "sem", "sga", "sgn", "shn", "sid", "sin", "sio", "sit", "sla", "slo", "slk", "slv", "sma", "sme", "smi", "smj", "smn", "smo", "sms", "sna", "snd", "snk", "sog", "som", "son", "sot", "spa", "srd", "srn", "srp", "srr", "ssa", "ssw", "suk", "sun", "sus", "sux", "swa", "swe", 
#> "syc", "syr", "tah", "tai", "tam", "tat", "tel", "tem", "ter", "tet", "tgk", "tgl", "tha", "tib", "bod", "tig", "tir", "tiv", "tkl", "tlh", "tli", "tmh", "tog", "ton", "tpi", "tsi", "tsn", "tso", "tuk", "tum", "tup", "tur", "tut", "tvl", "twi", "tyv", "udm", "uga", "uig", "ukr", "umb", "und", "urd", "uzb", "vai", "ven", "vie", "vol", "vot", "wak", "wal", "war", "was", "wel", "cym", "wen", "wln", "wol", "xal", "xho", "yao", "yap", "yid", "yor", "ypk", "zap", "zbl", "zen", "zgh", "zha", "znd", "zul", 
#> "zun", "zxx", "zza") 
#> ....|-- codeEntry: c("Afar", "Abkhazian", "Achinese", "Acoli", "Adangme", "Adyghe, Adygei", "Afro-Asiatic languages", "Afrihili", "Afrikaans", "Ainu", "Akan", "Akkadian", "Albanian", "Albanian", "Aleut", "Algonquian languages", "Southern Altai", "Amharic", "English, Old (ca.450-1100)", "Angika", "Apache languages", "Arabic", "Official Aramaic (700-300 BCE), Imperial Aramaic (700-300 BCE)", "Aragonese", "Armenian", "Armenian", "Mapudungun, Mapuche", "Arapaho", "Artificial languages", "Arawak", "Assamese", "Asturian, Bable, Leonese, Asturleonese", 
#> "Athapascan languages", "Australian languages", "Avaric", "Avestan", "Awadhi", "Aymara", "Azerbaijani", "Banda languages", "Bamileke languages", "Bashkir", "Baluchi", "Bambara", "Balinese", "Basque", "Basque", "Basa", "Baltic languages", "Beja, Bedawiyet", "Belarusian", "Bemba", "Bengali", "Berber languages", "Bhojpuri", "Bihari languages", "Bikol", "Bini, Edo", "Bislama", "Siksika", "Bantu (Other)", "Bosnian", "Braj", "Breton", "Batak languages", "Buriat", "Buginese", "Bulgarian", "Burmese", "Burmese", 
#> "Blin, Bilin", "Caddo", "Central American Indian languages", "Galibi Carib", "Catalan, Valencian", "Caucasian languages", "Cebuano", "Celtic languages", "Chamorro", "Chibcha", "Chechen", "Chagatai", "Chinese", "Chinese", "Chuukese", "Mari", "Chinook jargon", "Choctaw", "Chipewyan, Dene Suline", "Cherokee", "Church Slavic, Old Slavonic, Church Slavonic, Old Bulgarian, Old Church Slavonic", "Chuvash", "Cheyenne", "Chamic languages", "Coptic", "Cornish", "Corsican", "Creoles and pidgins, English based", 
#> "Creoles and pidgins, French-based ", "Creoles and pidgins, Portuguese-based ", "Cree", "Crimean Tatar, Crimean Turkish", "Creoles and pidgins ", "Kashubian", "Cushitic languages", "Czech", "Czech", "Dakota", "Danish", "Dargwa", "Land Dayak languages", "Delaware", "Slave (Athapascan)", "Dogrib", "Dinka", "Divehi, Dhivehi, Maldivian", "Dogri", "Dravidian languages", "Lower Sorbian", "Duala", "Dutch, Middle (ca.1050-1350)", "Dutch, Flemish", "Dutch, Flemish", "Dyula", "Dzongkha", "Efik", "Egyptian (Ancient)", 
#> "Ekajuk", "Elamite", "English", "English, Middle (1100-1500)", "Esperanto", "Estonian", "Ewe", "Ewondo", "Fang", "Faroese", "Fanti", "Fijian", "Filipino, Pilipino", "Finnish", "Finno-Ugrian languages", "Fon", "French", "French", "French, Middle (ca.1400-1600)", "French, Old (842-ca.1400)", "Northern Frisian", "Eastern Frisian", "Western Frisian", "Fulah", "Friulian", "Ga", "Gayo", "Gbaya", "Germanic languages", "Georgian", "Georgian", "German", "German", "Geez", "Gilbertese", "Gaelic, Scottish Gaelic", 
#> "Irish", "Galician", "Manx", "German, Middle High (ca.1050-1500)", "German, Old High (ca.750-1050)", "Gondi", "Gorontalo", "Gothic", "Grebo", "Greek, Ancient (to 1453)", "Greek, Modern (1453-)", "Greek, Modern (1453-)", "Guarani", "Swiss German, Alemannic, Alsatian", "Gujarati", "Gwich'in", "Haida", "Haitian, Haitian Creole", "Hausa", "Hawaiian", "Hebrew", "Herero", "Hiligaynon", "Himachali languages, Western Pahari languages", "Hindi", "Hittite", "Hmong, Mong", "Hiri Motu", "Croatian", "Upper Sorbian", 
#> "Hungarian", "Hupa", "Iban", "Igbo", "Icelandic", "Icelandic", "Ido", "Sichuan Yi, Nuosu", "Ijo languages", "Inuktitut", "Interlingue, Occidental", "Iloko", "Interlingua (International Auxiliary Language Association)", "Indic languages", "Indonesian", "Indo-European languages", "Ingush", "Inupiaq", "Iranian languages", "Iroquoian languages", "Italian", "Javanese", "Lojban", "Japanese", "Judeo-Persian", "Judeo-Arabic", "Kara-Kalpak", "Kabyle", "Kachin, Jingpho", "Kalaallisut, Greenlandic", "Kamba", 
#> "Kannada", "Karen languages", "Kashmiri", "Kanuri", "Kawi", "Kazakh", "Kabardian", "Khasi", "Khoisan languages", "Central Khmer", "Khotanese, Sakan", "Kikuyu, Gikuyu", "Kinyarwanda", "Kirghiz, Kyrgyz", "Kimbundu", "Konkani", "Komi", "Kongo", "Korean", "Kosraean", "Kpelle", "Karachay-Balkar", "Karelian", "Kru languages", "Kurukh", "Kuanyama, Kwanyama", "Kumyk", "Kurdish", "Kutenai", "Ladino", "Lahnda", "Lamba", "Lao", "Latin", "Latvian", "Lezghian", "Limburgan, Limburger, Limburgish", "Lingala", "Lithuanian", 
#> "Mongo", "Lozi", "Luxembourgish, Letzeburgesch", "Luba-Lulua", "Luba-Katanga", "Ganda", "Luiseno", "Lunda", "Luo (Kenya and Tanzania)", "Lushai", "Macedonian", "Macedonian", "Madurese", "Magahi", "Marshallese", "Maithili", "Makasar", "Malayalam", "Mandingo", "Maori", "Maori", "Austronesian languages", "Marathi", "Masai", "Malay", "Malay", "Moksha", "Mandar", "Mende", "Irish, Middle (900-1200)", "Mi'kmaq, Micmac", "Minangkabau", "Uncoded languages", "Mon-Khmer languages", "Malagasy", "Maltese", "Manchu", 
#> "Manipuri", "Manobo languages", "Mohawk", "Mongolian", "Mossi", "Multiple languages", "Munda languages", "Creek", "Mirandese", "Marwari", "Mayan languages", "Erzya", "Nahuatl languages", "North American Indian languages", "Neapolitan", "Nauru", "Navajo, Navaho", "Ndebele, South, South Ndebele", "Ndebele, North, North Ndebele", "Ndonga", "Low German, Low Saxon, German, Low, Saxon, Low", "Nepali", "Nepal Bhasa, Newari", "Nias", "Niger-Kordofanian languages", "Niuean", "Norwegian Nynorsk, Nynorsk, Norwegian", 
#> "Bokmål, Norwegian, Norwegian Bokmål", "Nogai", "Norse, Old", "Norwegian", "N'Ko", "Pedi, Sepedi, Northern Sotho", "Nubian languages", "Classical Newari, Old Newari, Classical Nepal Bhasa", "Chichewa, Chewa, Nyanja", "Nyamwezi", "Nyankole", "Nyoro", "Nzima", "Occitan (post 1500), Provençal", "Ojibwa", "Oriya", "Oromo", "Osage", "Ossetian, Ossetic", "Turkish, Ottoman (1500-1928)", "Otomian languages", "Papuan languages", "Pangasinan", "Pahlavi", "Pampanga, Kapampangan", "Panjabi, Punjabi", "Papiamento", 
#> "Palauan", "Persian, Old (ca.600-400 B.C.)", "Persian", "Persian", "Philippine languages", "Phoenician", "Pali", "Polish", "Pohnpeian", "Portuguese", "Prakrit languages", "Provençal, Old (to 1500)", "Pushto, Pashto", "Reserved for local use", "Quechua", "Rajasthani", "Rapanui", "Rarotongan, Cook Islands Maori", "Romance languages", "Romansh", "Romany", "Romanian, Moldavian, Moldovan", "Romanian, Moldavian, Moldovan", "Rundi", "Aromanian, Arumanian, Macedo-Romanian", "Russian", "Sandawe", "Sango", 
#> "Yakut", "South American Indian (Other)", "Salishan languages", "Samaritan Aramaic", "Sanskrit", "Sasak", "Santali", "Sicilian", "Scots", "Selkup", "Semitic languages", "Irish, Old (to 900)", "Sign Languages", "Shan", "Sidamo", "Sinhala, Sinhalese", "Siouan languages", "Sino-Tibetan languages", "Slavic languages", "Slovak", "Slovak", "Slovenian", "Southern Sami", "Northern Sami", "Sami languages", "Lule Sami", "Inari Sami", "Samoan", "Skolt Sami", "Shona", "Sindhi", "Soninke", "Sogdian", "Somali", 
#> "Songhai languages", "Sotho, Southern", "Spanish, Castilian", "Sardinian", "Sranan Tongo", "Serbian", "Serer", "Nilo-Saharan languages", "Swati", "Sukuma", "Sundanese", "Susu", "Sumerian", "Swahili", "Swedish", "Classical Syriac", "Syriac", "Tahitian", "Tai languages", "Tamil", "Tatar", "Telugu", "Timne", "Tereno", "Tetum", "Tajik", "Tagalog", "Thai", "Tibetan", "Tibetan", "Tigre", "Tigrinya", "Tiv", "Tokelau", "Klingon, tlhIngan-Hol", "Tlingit", "Tamashek", "Tonga (Nyasa)", "Tonga (Tonga Islands)", 
#> "Tok Pisin", "Tsimshian", "Tswana", "Tsonga", "Turkmen", "Tumbuka", "Tupi languages", "Turkish", "Altaic languages", "Tuvalu", "Twi", "Tuvinian", "Udmurt", "Ugaritic", "Uighur, Uyghur", "Ukrainian", "Umbundu", "Undetermined", "Urdu", "Uzbek", "Vai", "Venda", "Vietnamese", "Volapük", "Votic", "Wakashan languages", "Walamo", "Waray", "Washo", "Welsh", "Welsh", "Sorbian languages", "Walloon", "Wolof", "Kalmyk, Oirat", "Xhosa", "Yao", "Yapese", "Yiddish", "Yoruba", "Yupik languages", "Zapotec", "Blissymbols, Blissymbolics, Bliss", 
#> "Zenaga", "Standard Moroccan Tamazight", "Zhuang, Chuang", "Zande languages", "Zulu", "Zuni", "No linguistic content, Not applicable", "Zaza, Dimili, Dimli, Kirdki, Kirmanjki, Zazaki") 
#> ....|-- codeEntry: c("Afar", "Abkhazian", "Achinese", "Acoli", "Adangme", "Adyghe, Adygei", "Afro-Asiatic languages", "Afrihili", "Afrikaans", "Ainu", "Akan", "Akkadian", "Albanian", "Albanian", "Aleut", "Algonquian languages", "Southern Altai", "Amharic", "English, Old (ca.450-1100)", "Angika", "Apache languages", "Arabic", "Official Aramaic (700-300 BCE), Imperial Aramaic (700-300 BCE)", "Aragonese", "Armenian", "Armenian", "Mapudungun, Mapuche", "Arapaho", "Artificial languages", "Arawak", "Assamese", "Asturian, Bable, Leonese, Asturleonese", 
#> "Athapascan languages", "Australian languages", "Avaric", "Avestan", "Awadhi", "Aymara", "Azerbaijani", "Banda languages", "Bamileke languages", "Bashkir", "Baluchi", "Bambara", "Balinese", "Basque", "Basque", "Basa", "Baltic languages", "Beja, Bedawiyet", "Belarusian", "Bemba", "Bengali", "Berber languages", "Bhojpuri", "Bihari languages", "Bikol", "Bini, Edo", "Bislama", "Siksika", "Bantu (Other)", "Bosnian", "Braj", "Breton", "Batak languages", "Buriat", "Buginese", "Bulgarian", "Burmese", "Burmese", 
#> "Blin, Bilin", "Caddo", "Central American Indian languages", "Galibi Carib", "Catalan, Valencian", "Caucasian languages", "Cebuano", "Celtic languages", "Chamorro", "Chibcha", "Chechen", "Chagatai", "Chinese", "Chinese", "Chuukese", "Mari", "Chinook jargon", "Choctaw", "Chipewyan, Dene Suline", "Cherokee", "Church Slavic, Old Slavonic, Church Slavonic, Old Bulgarian, Old Church Slavonic", "Chuvash", "Cheyenne", "Chamic languages", "Coptic", "Cornish", "Corsican", "Creoles and pidgins, English based", 
#> "Creoles and pidgins, French-based ", "Creoles and pidgins, Portuguese-based ", "Cree", "Crimean Tatar, Crimean Turkish", "Creoles and pidgins ", "Kashubian", "Cushitic languages", "Czech", "Czech", "Dakota", "Danish", "Dargwa", "Land Dayak languages", "Delaware", "Slave (Athapascan)", "Dogrib", "Dinka", "Divehi, Dhivehi, Maldivian", "Dogri", "Dravidian languages", "Lower Sorbian", "Duala", "Dutch, Middle (ca.1050-1350)", "Dutch, Flemish", "Dutch, Flemish", "Dyula", "Dzongkha", "Efik", "Egyptian (Ancient)", 
#> "Ekajuk", "Elamite", "English", "English, Middle (1100-1500)", "Esperanto", "Estonian", "Ewe", "Ewondo", "Fang", "Faroese", "Fanti", "Fijian", "Filipino, Pilipino", "Finnish", "Finno-Ugrian languages", "Fon", "French", "French", "French, Middle (ca.1400-1600)", "French, Old (842-ca.1400)", "Northern Frisian", "Eastern Frisian", "Western Frisian", "Fulah", "Friulian", "Ga", "Gayo", "Gbaya", "Germanic languages", "Georgian", "Georgian", "German", "German", "Geez", "Gilbertese", "Gaelic, Scottish Gaelic", 
#> "Irish", "Galician", "Manx", "German, Middle High (ca.1050-1500)", "German, Old High (ca.750-1050)", "Gondi", "Gorontalo", "Gothic", "Grebo", "Greek, Ancient (to 1453)", "Greek, Modern (1453-)", "Greek, Modern (1453-)", "Guarani", "Swiss German, Alemannic, Alsatian", "Gujarati", "Gwich'in", "Haida", "Haitian, Haitian Creole", "Hausa", "Hawaiian", "Hebrew", "Herero", "Hiligaynon", "Himachali languages, Western Pahari languages", "Hindi", "Hittite", "Hmong, Mong", "Hiri Motu", "Croatian", "Upper Sorbian", 
#> "Hungarian", "Hupa", "Iban", "Igbo", "Icelandic", "Icelandic", "Ido", "Sichuan Yi, Nuosu", "Ijo languages", "Inuktitut", "Interlingue, Occidental", "Iloko", "Interlingua (International Auxiliary Language Association)", "Indic languages", "Indonesian", "Indo-European languages", "Ingush", "Inupiaq", "Iranian languages", "Iroquoian languages", "Italian", "Javanese", "Lojban", "Japanese", "Judeo-Persian", "Judeo-Arabic", "Kara-Kalpak", "Kabyle", "Kachin, Jingpho", "Kalaallisut, Greenlandic", "Kamba", 
#> "Kannada", "Karen languages", "Kashmiri", "Kanuri", "Kawi", "Kazakh", "Kabardian", "Khasi", "Khoisan languages", "Central Khmer", "Khotanese, Sakan", "Kikuyu, Gikuyu", "Kinyarwanda", "Kirghiz, Kyrgyz", "Kimbundu", "Konkani", "Komi", "Kongo", "Korean", "Kosraean", "Kpelle", "Karachay-Balkar", "Karelian", "Kru languages", "Kurukh", "Kuanyama, Kwanyama", "Kumyk", "Kurdish", "Kutenai", "Ladino", "Lahnda", "Lamba", "Lao", "Latin", "Latvian", "Lezghian", "Limburgan, Limburger, Limburgish", "Lingala", "Lithuanian", 
#> "Mongo", "Lozi", "Luxembourgish, Letzeburgesch", "Luba-Lulua", "Luba-Katanga", "Ganda", "Luiseno", "Lunda", "Luo (Kenya and Tanzania)", "Lushai", "Macedonian", "Macedonian", "Madurese", "Magahi", "Marshallese", "Maithili", "Makasar", "Malayalam", "Mandingo", "Maori", "Maori", "Austronesian languages", "Marathi", "Masai", "Malay", "Malay", "Moksha", "Mandar", "Mende", "Irish, Middle (900-1200)", "Mi'kmaq, Micmac", "Minangkabau", "Uncoded languages", "Mon-Khmer languages", "Malagasy", "Maltese", "Manchu", 
#> "Manipuri", "Manobo languages", "Mohawk", "Mongolian", "Mossi", "Multiple languages", "Munda languages", "Creek", "Mirandese", "Marwari", "Mayan languages", "Erzya", "Nahuatl languages", "North American Indian languages", "Neapolitan", "Nauru", "Navajo, Navaho", "Ndebele, South, South Ndebele", "Ndebele, North, North Ndebele", "Ndonga", "Low German, Low Saxon, German, Low, Saxon, Low", "Nepali", "Nepal Bhasa, Newari", "Nias", "Niger-Kordofanian languages", "Niuean", "Norwegian Nynorsk, Nynorsk, Norwegian", 
#> "Bokmål, Norwegian, Norwegian Bokmål", "Nogai", "Norse, Old", "Norwegian", "N'Ko", "Pedi, Sepedi, Northern Sotho", "Nubian languages", "Classical Newari, Old Newari, Classical Nepal Bhasa", "Chichewa, Chewa, Nyanja", "Nyamwezi", "Nyankole", "Nyoro", "Nzima", "Occitan (post 1500), Provençal", "Ojibwa", "Oriya", "Oromo", "Osage", "Ossetian, Ossetic", "Turkish, Ottoman (1500-1928)", "Otomian languages", "Papuan languages", "Pangasinan", "Pahlavi", "Pampanga, Kapampangan", "Panjabi, Punjabi", "Papiamento", 
#> "Palauan", "Persian, Old (ca.600-400 B.C.)", "Persian", "Persian", "Philippine languages", "Phoenician", "Pali", "Polish", "Pohnpeian", "Portuguese", "Prakrit languages", "Provençal, Old (to 1500)", "Pushto, Pashto", "Reserved for local use", "Quechua", "Rajasthani", "Rapanui", "Rarotongan, Cook Islands Maori", "Romance languages", "Romansh", "Romany", "Romanian, Moldavian, Moldovan", "Romanian, Moldavian, Moldovan", "Rundi", "Aromanian, Arumanian, Macedo-Romanian", "Russian", "Sandawe", "Sango", 
#> "Yakut", "South American Indian (Other)", "Salishan languages", "Samaritan Aramaic", "Sanskrit", "Sasak", "Santali", "Sicilian", "Scots", "Selkup", "Semitic languages", "Irish, Old (to 900)", "Sign Languages", "Shan", "Sidamo", "Sinhala, Sinhalese", "Siouan languages", "Sino-Tibetan languages", "Slavic languages", "Slovak", "Slovak", "Slovenian", "Southern Sami", "Northern Sami", "Sami languages", "Lule Sami", "Inari Sami", "Samoan", "Skolt Sami", "Shona", "Sindhi", "Soninke", "Sogdian", "Somali", 
#> "Songhai languages", "Sotho, Southern", "Spanish, Castilian", "Sardinian", "Sranan Tongo", "Serbian", "Serer", "Nilo-Saharan languages", "Swati", "Sukuma", "Sundanese", "Susu", "Sumerian", "Swahili", "Swedish", "Classical Syriac", "Syriac", "Tahitian", "Tai languages", "Tamil", "Tatar", "Telugu", "Timne", "Tereno", "Tetum", "Tajik", "Tagalog", "Thai", "Tibetan", "Tibetan", "Tigre", "Tigrinya", "Tiv", "Tokelau", "Klingon, tlhIngan-Hol", "Tlingit", "Tamashek", "Tonga (Nyasa)", "Tonga (Tonga Islands)", 
#> "Tok Pisin", "Tsimshian", "Tswana", "Tsonga", "Turkmen", "Tumbuka", "Tupi languages", "Turkish", "Altaic languages", "Tuvalu", "Twi", "Tuvinian", "Udmurt", "Ugaritic", "Uighur, Uyghur", "Ukrainian", "Umbundu", "Undetermined", "Urdu", "Uzbek", "Vai", "Venda", "Vietnamese", "Volapük", "Votic", "Wakashan languages", "Walamo", "Waray", "Washo", "Welsh", "Welsh", "Sorbian languages", "Walloon", "Wolof", "Kalmyk, Oirat", "Xhosa", "Yao", "Yapese", "Yiddish", "Yoruba", "Yupik languages", "Zapotec", "Blissymbols, Blissymbolics, Bliss", 
#> "Zenaga", "Standard Moroccan Tamazight", "Zhuang, Chuang", "Zande languages", "Zulu", "Zuni", "No linguistic content, Not applicable", "Zaza, Dimili, Dimli, Kirdki, Kirmanjki, Zazaki")