getISOInternalCodelists

getISOInternalCodelists allows to get the list of ISO codelists registered in geometa

getISOInternalCodelists()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOInternalCodelists()
#> $FC_RoleType
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19110
#>     description:  code list for the classification of roles 
#>     entries: data.frame
#>     id: FC_RoleType
#>     identifier: FC_RoleType
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $CI_DateTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: identification of when a given event occurred
#>     entries: data.frame
#>     id: CI_DateTypeCode
#>     identifier: CI_DateTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $CI_PresentationFormCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: mode in which the data is represented
#>     entries: data.frame
#>     id: CI_PresentationFormCode
#>     identifier: CI_PresentationFormCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $CI_RoleCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: function performed by the responsible party
#>     entries: data.frame
#>     id: CI_RoleCode
#>     identifier: CI_RoleCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $CI_OnLineFunctionCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: function performed by the resource
#>     entries: data.frame
#>     id: CI_OnLineFunctionCode
#>     identifier: CI_OnLineFunctionCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $Country
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISO 3166-2
#>     description: Country : ISO 3166-2 (2 characters)
#>     entries: data.frame
#>     id: Country
#>     identifier: Country
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: ML_gmxCodelists.xml
#> 
#> $DCPList
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19119
#>     description: distributed computing platforms on which the operation h ...
#>     entries: data.frame
#>     id: DCPList
#>     identifier: DCPList
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $DQ_EvaluationMethodTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: type or method for evaluating an identified data quality ...
#>     entries: data.frame
#>     id: DQ_EvaluationMethodTypeCode
#>     identifier: DQ_EvaluationMethodTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $DS_AssociationTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: justification for the correlation of two datasets
#>     entries: data.frame
#>     id: DS_AssociationTypeCode
#>     identifier: DS_AssociationTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $DS_InitiativeTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: type of aggregation activity in which datasets are related
#>     entries: data.frame
#>     id: DS_InitiativeTypeCode
#>     identifier: DS_InitiativeTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $LanguageCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISO 639-2
#>     description: Language : ISO 639-2 (3 characters)
#>     entries: data.frame
#>     id: LanguageCode
#>     identifier: LanguageCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: ISO-639-2_utf-8.txt
#> 
#> $MD_CellGeometryCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: code indicating whether grid data is point or area
#>     entries: data.frame
#>     id: MD_CellGeometryCode
#>     identifier: MD_CellGeometryCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_CharacterSetCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: name of the character coding standard used in the resource
#>     entries: data.frame
#>     id: MD_CharacterSetCode
#>     identifier: MD_CharacterSetCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: ML_gmxCodelists.xml
#> 
#> $MD_ClassificationCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: name of the handling restrictions on the dataset
#>     entries: data.frame
#>     id: MD_ClassificationCode
#>     identifier: MD_ClassificationCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_CoverageContentTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: specific type of information represented in the cell
#>     entries: data.frame
#>     id: MD_CoverageContentTypeCode
#>     identifier: MD_CoverageContentTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_DatatypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: datatype of element or entity
#>     entries: data.frame
#>     id: MD_DatatypeCode
#>     identifier: MD_DatatypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_DimensionNameTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: name of the dimension
#>     entries: data.frame
#>     id: MD_DimensionNameTypeCode
#>     identifier: MD_DimensionNameTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_DistributionUnits
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: NA
#>     description: NA
#>     entries: NULL
#>     id: MD_DistributionUnits
#>     identifier: NA
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_GeometricObjectTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: name of point or vector objects used to locate zero-, on ...
#>     entries: data.frame
#>     id: MD_GeometricObjectTypeCode
#>     identifier: MD_GeometricObjectTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_KeywordTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: methods used to group similar keywords
#>     entries: data.frame
#>     id: MD_KeywordTypeCode
#>     identifier: MD_KeywordTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_ImagingConditionCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: code which indicates conditions which may affect the image
#>     entries: data.frame
#>     id: MD_ImagingConditionCode
#>     identifier: MD_ImagingConditionCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_MaintenanceFrequencyCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: frequency with which modifications and deletions are mad ...
#>     entries: data.frame
#>     id: MD_MaintenanceFrequencyCode
#>     identifier: MD_MaintenanceFrequencyCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_MediumFormatCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: method used to write to the medium
#>     entries: data.frame
#>     id: MD_MediumFormatCode
#>     identifier: MD_MediumFormatCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_MediumNameCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: name of the medium
#>     entries: data.frame
#>     id: MD_MediumNameCode
#>     identifier: MD_MediumNameCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_ObligationCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: obligation of the element or entity
#>     entries: data.frame
#>     id: MD_ObligationCode
#>     identifier: MD_ObligationCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_PixelOrientationCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: point in a pixel corresponding to the Earth location of  ...
#>     entries: data.frame
#>     id: MD_PixelOrientationCode
#>     identifier: MD_PixelOrientationCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_ProgressCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: status of the dataset or progress of a review
#>     entries: data.frame
#>     id: MD_ProgressCode
#>     identifier: MD_ProgressCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_RestrictionCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: limitation(s) placed upon the access or use of the data
#>     entries: data.frame
#>     id: MD_RestrictionCode
#>     identifier: MD_RestrictionCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_SpatialRepresentationTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: method used to represent geographic information in the d ...
#>     entries: data.frame
#>     id: MD_SpatialRepresentationTypeCode
#>     identifier: MD_SpatialRepresentationTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_TopicCategoryCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: high-level geographic data thematic classification to as ...
#>     entries: data.frame
#>     id: MD_TopicCategoryCode
#>     identifier: MD_TopicCategoryCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MD_TopologyLevelCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: degree of complexity of the spatial relationships
#>     entries: data.frame
#>     id: MD_TopologyLevelCode
#>     identifier: MD_TopologyLevelCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_BandDefinition
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: designation of criterion for defining maximum and minimu ...
#>     entries: data.frame
#>     id: MI_BandDefinition
#>     identifier: MI_BandDefinition
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_ContextCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: designation of criterion for defining the context of the ...
#>     entries: data.frame
#>     id: MI_ContextCode
#>     identifier: MI_ContextCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_GeometryTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: geometric description of the collection
#>     entries: data.frame
#>     id: MI_GeometryTypeCode
#>     identifier: MI_GeometryTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_ObjectiveTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: temporal persistence of collection objective
#>     entries: data.frame
#>     id: MI_ObjectiveTypeCode
#>     identifier: MI_ObjectiveTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_OperationTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: code indicating whether the data contained in this packe ...
#>     entries: data.frame
#>     id: MI_OperationTypeCode
#>     identifier: MI_OperationTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_PolarisationOrientationCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: Polarisation of the antenna relative to the waveform
#>     entries: data.frame
#>     id: MI_PolarisationOrientationCode
#>     identifier: MI_PolarisationOrientationCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_PriorityCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: ordered list of priorities
#>     entries: data.frame
#>     id: MI_PriorityCode
#>     identifier: MI_PriorityCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_SequenceCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: temporal relation of activation
#>     entries: data.frame
#>     id: MI_SequenceCode
#>     identifier: MI_SequenceCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_TransferFunctionTypeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: transform function to be used when scaling a physical va ...
#>     entries: data.frame
#>     id: MI_TransferFunctionTypeCode
#>     identifier: MI_TransferFunctionTypeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MI_TriggerCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115-2
#>     description: mechanism of activation
#>     entries: data.frame
#>     id: MI_TriggerCode
#>     identifier: MI_TriggerCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $SV_CouplingType
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19119
#>     description: indication of service instance coupling
#>     entries: data.frame
#>     id: SV_CouplingType
#>     identifier: SV_CouplingType
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $SV_ParameterDirection
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19119
#>     description:  indication if the parameter is an input to the service, ...
#>     entries: data.frame
#>     id: SV_ParameterDirection
#>     identifier: SV_ParameterDirection
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#> 
#> $MX_ScopeCode
#> <ISOCodelist>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     codeSpace: ISOTC211/19115
#>     description: Extension of MD_ScopeCode for the needs of GMX applicati ...
#>     entries: data.frame
#>     id: MX_ScopeCode
#>     identifier: MX_ScopeCode
#>     initialize: function (refFile, id) 
#>     parse: function (refFile, id) 
#>     refFile: gmxCodelists.xml
#>