INSPIREMetadataValidator

INSPIREMetadataValidator

INSPIREMetadataValidator

Format

R6Class object.

Value

Object of R6Class for setting an INSPIREMetadataValidator

Note

Starting 2025-05-02, The API key is not needed anymore and metadata can be then validated as anonymous user.

References

INSPIRE Reference Validator Web Service (https://inspire.ec.europa.eu/validator/swagger-ui.html)

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

geometa::geometaLogger -> INSPIREMetadataValidator

Public fields

url

url of the INSPIRE metadata validator

running

wether the service is up and running

status

status

Methods

Public methods

INSPIREMetadataValidator$new()

INSPIREMetadataValidator$uploadFile()

INSPIREMetadataValidator$getAPIKey()

INSPIREMetadataValidator$getValidationReport()

INSPIREMetadataValidator$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

Method new()

Method used to instantiate an INSPIRE Metadata validator. To check metadata with the INSPIRE metadata validator, a user API key is now required, and should be specified with the apiKey. By default, the url will be the INSPIRE production service https://inspire.ec.europa.eu/validator/swagger-ui.html.

The keyring_backend can be set to use a different backend for storing the INSPIRE metadata validator API key with keyring (Default value is 'env').

Usage

INSPIREMetadataValidator$new(
  url = "https://inspire.ec.europa.eu/validator-api",
  apiKey = NULL,
  keyring_backend = "env"
)

Arguments

url

url

apiKey

API key. Default is NULL.

keyring_backend

backend name to use with keyring to store API key

Method uploadFile()

Uploads a file. Upload a XML metadata file to INSPIRE web-service. Method called internally through getValidationReport.

Usage

INSPIREMetadataValidator$uploadFile(path)

Arguments

path

path

Returns

the response from the web-service

Method getAPIKey()

Retrieves the API key

Usage

INSPIREMetadataValidator$getAPIKey()

Returns

the API key as character

Method getValidationReport()

Get validation report for a metadata specified either as R object of class ISOMetadata (from geometa package) or XMLInternalNode-class (from XML package), or as XML file, providing the path of the XML file to be sent to the INSPIRE metadata validator web-service. By default, a summary report is returned. To append the raw response of INSPIRE validation web-service to the summary report, set raw = TRUE.

Usage

INSPIREMetadataValidator$getValidationReport(
  obj = NULL,
  file = NULL,
  raw = FALSE
)

Arguments

obj

obj

file

file

raw

raw

Returns

an object of class list

Method clone()

The objects of this class are cloneable with this method.

Usage

INSPIREMetadataValidator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 # \donttest{
  apiKey <- ""
  if(nzchar(apiKey)){
    inspireValidator <- INSPIREMetadataValidator$new(apiKey = apiKey)
    inspireReport <- inspireValidator$getValidationReport(obj = ISOMetadata$new())
  }
 # }