ISOAggregateInformation

ISOAggregateInformation

Format

R6Class object.

Value

Object of R6Class for modelling a ISO AggregateInformation

Fields

aggregateDataSetName

[ISOCitation]

aggregateDataSetIdentifier

[ISOMetaIdentifier]

associationType

[ISOAssociationType]

initiativeType

[ISOInitiativeType]

Methods

new(xml)

This method is used to instantiate an ISOAggregateInformation

setAggregateDataSetName(datasetName)

Sets aggregate dataset name, as an object of class ISOCitation

setAggregateDataSetIdentifier(datasetIdentifier)

Sets aggregate dataset identifier, as an object of class ISOMetaIdentifier

setAssociationType(associationType)

Sets the association type

setInitiativeType(initiativeType))

Sets the initiative type

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

  #encoding
  md <- ISOAggregateInformation$new()
  
  #adding a point of contact
  rp <- ISOResponsibleParty$new()
  rp$setIndividualName("someone")
  rp$setOrganisationName("somewhere")
  rp$setPositionName("someposition")
  rp$setRole("pointOfContact")
  contact <- ISOContact$new()
  phone <- ISOTelephone$new()
  phone$setVoice("myphonenumber")
  phone$setFacsimile("myfacsimile")
  contact$setPhone(phone)
  address <- ISOAddress$new()
  address$setDeliveryPoint("theaddress")
  address$setCity("thecity")
  address$setPostalCode("111")
  address$setCountry("France")
  address$setEmail("someone@theorg.org")
  contact$setAddress(address)
  res <- ISOOnlineResource$new()
  res$setLinkage("http://www.somewhereovertheweb.org")
  res$setName("somename")
  contact$setOnlineResource(res)
  rp$setContactInfo(contact)
  #citation
  ct <- ISOCitation$new()
  ct$setTitle("sometitle")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  ct$addDate(d)
  ct$setEdition("1.0")
  ct$setEditionDate(ISOdate(2015,1,1))
  ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
  ct$addPresentationForm("mapDigital")
#> [1] TRUE
  ct$addCitedResponsibleParty(rp)
#> [1] TRUE
  md$setAggregateDataSetName(ct)
  
  md$setAssociationType("source")
  md$setInitiativeType("investigation")
  
  xml <- md$encode()
#> [geometa][INFO] Object 'ISOAggregateInformation' is VALID according to ISO 19139 XML schemas!