ISOCitation

ISOCitation

ISOCitation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Citation

References

ISO 19115:2003 - Geographic information – Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOCitation

Public fields

title

title

alternateTitle

alternate title

date

date list

edition

edition

editionDate

edition date

identifier

identifier list

citedResponsibleParty

list of cited responsible parties

presentationForm

list of presentation forms

series

series

otherCitationDetails

other citation details

collectiveTitle

collective title

ISBN

ISBN

ISSN

ISSN

Methods

Public methods

ISOCitation$new()

ISOCitation$setTitle()

ISOCitation$setAlternateTitle()

ISOCitation$addAlternateTitle()

ISOCitation$delAlternateTitle()

ISOCitation$addDate()

ISOCitation$setEdition()

ISOCitation$setEditionDate()

ISOCitation$setIdentifier()

ISOCitation$addIdentifier()

ISOCitation$delIdentifier()

ISOCitation$setCitedResponsibleParty()

ISOCitation$addCitedResponsibleParty()

ISOCitation$delCitedResponsibleParty()

ISOCitation$setPresentationForm()

ISOCitation$addPresentationForm()

ISOCitation$delPresentationForm()

ISOCitation$setSeries()

ISOCitation$setOtherCitationDetails()

ISOCitation$setCollectiveTitle()

ISOCitation$setISBN()

ISOCitation$setISSN()

ISOCitation$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOCitation$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setTitle()

Set title

Usage

ISOCitation$setTitle(title, locales = NULL)

Arguments

title

title

locales

list of localized names. Default is NULL

Method setAlternateTitle()

Set alternate title

Usage

ISOCitation$setAlternateTitle(alternateTitle, locales = NULL)

Arguments

alternateTitle

alternate title

locales

list of localized names. Default is NULL

Method addAlternateTitle()

Adds alternate title

Usage

ISOCitation$addAlternateTitle(alternateTitle, locales = NULL)

Arguments

alternateTitle

alternate title

locales

list of localized titles. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delAlternateTitle()

Deletes alternate title

Usage

ISOCitation$delAlternateTitle(alternateTitle, locales = NULL)

Arguments

alternateTitle

alternate title

locales

list of localized titles. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method addDate()

Adds date

Usage

ISOCitation$addDate(date)

Arguments

date

date

Returns

TRUE if added, FALSE otherwise

Method setEdition()

Set edition

Usage

ISOCitation$setEdition(edition, locales = NULL)

Arguments

edition

edition

locales

list of localized editions. Default is NULL

Method setEditionDate()

Sets the edition date, either an ISODate object containing date and dateType or a simple R date "POSIXct"/"POSIXt" object. For thesaurus citations, an ISODate should be used while for the general citation of ISODataIdentification, a simple R date should be used.

Usage

ISOCitation$setEditionDate(editionDate)

Arguments

editionDate

object of class Date or POSIXct

Method setIdentifier()

Set identifier

Usage

ISOCitation$setIdentifier(identifier)

Arguments

identifier

identifier, object of class ISOMetaIdentifier

Method addIdentifier()

Adds identifier

Usage

ISOCitation$addIdentifier(identifier)

Arguments

identifier

identifier, object of class ISOMetaIdentifier

locales

list of localized identifiers. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delIdentifier()

Deletes identifier

Usage

ISOCitation$delIdentifier(identifier)

Arguments

identifier

identifier, object of class ISOMetaIdentifier

locales

list of localized identifiers. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method setCitedResponsibleParty()

Set cited responsible party

Usage

ISOCitation$setCitedResponsibleParty(rp)

Arguments

rp

cited responsible party, object of class ISOResponsibleParty

Method addCitedResponsibleParty()

Adds cited responsible party

Usage

ISOCitation$addCitedResponsibleParty(rp)

Arguments

rp

cited responsible party, object of class ISOResponsibleParty

locales

list of localized responsible parties. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delCitedResponsibleParty()

Deletes cited responsible party

Usage

ISOCitation$delCitedResponsibleParty(rp)

Arguments

rp

cited responsible party, object of class ISOResponsibleParty

locales

list of localized responsible parties. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method setPresentationForm()

Sets presentation form

Usage

ISOCitation$setPresentationForm(presentationForm)

Arguments

presentationForm

presentation form, object of class ISOPresentationForm or character among values returned by ISOPresentationForm$values()

Method addPresentationForm()

Adds presentation form

Usage

ISOCitation$addPresentationForm(presentationForm)

Arguments

presentationForm

presentation form, object of class ISOPresentationForm or character among values returned by ISOPresentationForm$values()

Returns

TRUE if added, FALSE otherwise

Method delPresentationForm()

Deletes presentation form

Usage

ISOCitation$delPresentationForm(presentationForm)

Arguments

presentationForm

presentation form, object of class ISOPresentationForm or character among values returned by ISOPresentationForm$values()

Returns

TRUE if deleted, FALSE otherwise

Method setSeries()

Set series

Usage

ISOCitation$setSeries(series)

Arguments

series

object of class ISOCitationSeries

Method setOtherCitationDetails()

Set other citation details

Usage

ISOCitation$setOtherCitationDetails(otherCitationDetails, locales = NULL)

Arguments

otherCitationDetails

other citation details

locales

list of localized other citation details. Default is NULL

Method setCollectiveTitle()

Set collective title

Usage

ISOCitation$setCollectiveTitle(collectiveTitle, locales = NULL)

Arguments

collectiveTitle

collective title

locales

list of localized titles. Default is NULL

Method setISBN()

Set ISBN

Usage

ISOCitation$setISBN(isbn)

Arguments

isbn

isbn

Method setISSN()

Set ISSN

Usage

ISOCitation$setISSN(issn)

Arguments

issn

issn

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOCitation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 #create ISOCitation
 md <- ISOCitation$new()
 md$setTitle("sometitle")
 md$setEdition("1.0")
 md$setEditionDate(ISOdate(2015,1,1))
 md$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
#> [1] TRUE
 md$addPresentationForm("mapDigital")
#> [1] TRUE
 
 #add a cited responsible party
 rp <- ISOResponsibleParty$new()
 rp$setIndividualName("someone")
 rp$setOrganisationName("somewhere")
 rp$setPositionName("someposition")
 rp$setRole("pointOfContact")
 contact <- ISOContact$new()
 phone <- ISOTelephone$new()
 phone$setVoice("myphonenumber")
 phone$setFacsimile("myfacsimile")
 contact$setPhone(phone)
 address <- ISOAddress$new()
 address$setDeliveryPoint("theaddress")
 address$setCity("thecity")
 address$setPostalCode("111")
 address$setCountry("France")
 address$setEmail("someone@theorg.org")
 contact$setAddress(address)
 res <- ISOOnlineResource$new()
 res$setLinkage("http://www.somewhereovertheweb.org")
 res$setName("somename")
 contact$setOnlineResource(res)
 rp$setContactInfo(contact)
 md$addCitedResponsibleParty(rp)
#> [1] TRUE
 xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmd}edition': This element is not expected. Expected is one of ( {http://www.isotc211.org/2005/gmd}alternateTitle, {http://www.isotc211.org/2005/gmd}date ) at line 5. 
#> [geometa][WARN] Object 'ISOCitation' is INVALID according to ISO 19139 XML schemas!