ISODataQualityAbstractElement

ISODataQualityAbstractElement

ISODataQualityAbstractElement

Format

R6Class object.

Value

Object of R6Class for modelling an ISODataQualityAbstractElement

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISODataQualityAbstractElement

Public fields

nameOfMeasure

nameOfMeasure [0..*]: character

measureIdentification

measureIdentification [0..1]: ISOMetaIdentifier

measureDescription

measureDescription [0..1]: character

evaluationMethodType

evaluationMethodType [0..1]: ISOEvaluationMethodType

evaluationMethodDescription

evaluationMethodDescription [0..1]: character

evaluationProcedure

evaluationProcedure [0..1]: ISOCitation

dateTime

dateTime [0..1]: ISODateTime

result

result [1..2]: ISOConformanceResult

Methods

Public methods

ISODataQualityAbstractElement$new()

ISODataQualityAbstractElement$addNameOfMeasure()

ISODataQualityAbstractElement$delNameOfMeasure()

ISODataQualityAbstractElement$setMeasureIdentification()

ISODataQualityAbstractElement$setMeasureDescription()

ISODataQualityAbstractElement$setEvaluationMethodType()

ISODataQualityAbstractElement$setEvaluationMethodDescription()

ISODataQualityAbstractElement$setEvaluationProcedure()

ISODataQualityAbstractElement$setDateTime()

ISODataQualityAbstractElement$addResult()

ISODataQualityAbstractElement$delResult()

ISODataQualityAbstractElement$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISODataQualityAbstractElement$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method addNameOfMeasure()

Adds name of measure

Usage

ISODataQualityAbstractElement$addNameOfMeasure(name, locales = NULL)

Arguments

name

name

locales

list of localized names. Default is NULL

Returns

TRUE if added, FALSE

Method delNameOfMeasure()

Deletes name of measure

Usage

ISODataQualityAbstractElement$delNameOfMeasure(name, locales = NULL)

Arguments

name

name

locales

list of localized names. Default is NULL

Returns

TRUE if deleted, FALSE

Method setMeasureIdentification()

Set measure identification

Usage

ISODataQualityAbstractElement$setMeasureIdentification(identification)

Arguments

identification

object of class ISOMetaIdentifier

Method setMeasureDescription()

Set measure description

Usage

ISODataQualityAbstractElement$setMeasureDescription(
  description,
  locales = NULL
)

Arguments

description

object of class character

locales

list of localized descriptions. Default is NULL

Method setEvaluationMethodType()

Set evaluation method type

Usage

ISODataQualityAbstractElement$setEvaluationMethodType(type)

Arguments

type

object of class ISOEvaluationMethodType or any character value from those returned by ISOEvaluationMethodType$values()

Method setEvaluationMethodDescription()

Set evaluation method description

Usage

ISODataQualityAbstractElement$setEvaluationMethodDescription(
  description,
  locales = NULL
)

Arguments

description

description

locales

list of localized descriptions. Default is NULL

Method setEvaluationProcedure()

Set evaluation procedure

Usage

ISODataQualityAbstractElement$setEvaluationProcedure(procedure)

Arguments

procedure

procedure, object of class ISOCitation

Method setDateTime()

Set date time

Usage

ISODataQualityAbstractElement$setDateTime(dateTime)

Arguments

dateTime

date time, object of class POSIXct

Method addResult()

Adds result

Usage

ISODataQualityAbstractElement$addResult(result)

Arguments

result

object of class ISOConformanceResult

Returns

TRUE if added, FALSE otherwise

Method delResult()

Deletes result

Usage

ISODataQualityAbstractElement$delResult(result)

Arguments

result

object of class ISOConformanceResult

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISODataQualityAbstractElement$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.