ISOExtendedElementInformation

ISOExtendedElementInformation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO ExtendedElementInformation

Fields

name

[character]

shortName

[character]

domainCode

[integer]

definition

[character]

obligation

[ISOObligation]

condition

[character]

dataType

[ISODateType]

maximumOccurrence

[character]

domainValue

[character]

parentEntity

[character]

rule

[character]

rationale

[character]

source

[ISOResponsibleParty]

Methods

new(xml)

This method is used to instantiate an ISOExtendedElementInformation

setName(name, locales)

Sets the element name, object of class Character. Locale names can be specified as list with the locales argument.

setShortName(shortName, locales)

Sets the element shortname, object of class character. Locale names can be specified as list with the locales argument.

setDomainCode(domainCode)

Sets the element domain code, object of class integer

setDefinition(definition, locales)

Sets the element definition, object of class character. Locale names can be specified as list with the locales argument.

setObligation(obligation)

Sets an obligation, as object of class character or class ISOObligation. If an object of class "character" is specified, it must match the accepted obligation values ISOObligation$values().

setCondition(condition, locales)

Sets the element condition, object of class character. Locale names can be specified as list with the locales argument.

setDatatype(dataType)

Sets the element datatype, as object of class character or class ISODatatype. If an object of class "character" is specified, it must match the accepted datatype values ISODatatype$values().

setMaximumOccurrrence(maximumOccurrence)

Sets the element maximum occurrence, object of class character

setDomainValue(domainValue)

Sets the element domain value, object of class character

addParentyEntity(parentEntity)

Adds a parent Entity, object of class character

delParentEntity(parentEntity)

Deletes a parent Entity, object of class character

setRule(rule, locales)

Sets a rule, object of class character. Locale names can be specified as list with the locales argument.

addRationale(rationale, locales)

Adds a rationale, object of class character. Locale names can be specified as list with the locales argument.

delRationale(rationale, locales)

Deletes a rationale, object of class character. Locale names can be specified as list with the locales argument. Local names should match those of the keyword to be deleted, otherwise nothing will be deleted.

addSource(source)

Adds a source, object of class ISOResponsibleParty

delSource(source)

Deletes a source, object of class ISOResponsibleParty

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Examples

  md <- ISOExtendedElementInformation$new()
  md$setName("name")
  md$setShortName("shortName")
  md$setDomainCode(1L)
  md$setDefinition("some definition")
  md$setObligation("mandatory")
  md$setCondition("no condition")
  md$setDatatype("characterString")
  md$setMaximumOccurrence("string")
  md$setDomainValue("value")
  md$addParentEntity("none")
#> [1] TRUE
  md$setRule("rule")
  md$addRationale("rationale")
#> [1] TRUE
  
  #adding a source
  rp <- ISOResponsibleParty$new()
  rp$setIndividualName("someone")
  rp$setOrganisationName("somewhere")
  rp$setPositionName("someposition")
  rp$setRole("pointOfContact")
  contact <- ISOContact$new()
  phone <- ISOTelephone$new()
  phone$setVoice("myphonenumber")
  phone$setFacsimile("myfacsimile")
  contact$setPhone(phone)
  address <- ISOAddress$new()
  address$setDeliveryPoint("theaddress")
  address$setCity("thecity")
  address$setPostalCode("111")
  address$setCountry("France")
  address$setEmail("someone@theorg.org")
  contact$setAddress(address)
  res <- ISOOnlineResource$new()
  res$setLinkage("http://www.somewhereovertheweb.org")
  res$setName("somename")
  contact$setOnlineResource(res)
  rp$setContactInfo(contact)
  
  md$addSource(rp)
#> [1] TRUE
  
  xml <- md$encode()
#> [geometa][INFO] Object 'ISOExtendedElementInformation' is VALID according to ISO 19139 XML schemas!