ISOGeorectified

ISOGeorectified

ISOGeorectified

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Georectified

References

- ISO 19139 https://schemas.isotc211.org/19139/-/gmd/1.0/gmd/#element_MD_Georectified

- ISO 19115-3 https://schemas.isotc211.org/19115/-3/msr/1.0/msr/#element_MD_Georectified

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> geometa::ISOAbstractSpatialRepresentation -> geometa::ISOSpatialRepresentation -> geometa::ISOGridSpatialRepresentation -> ISOGeorectified

Public fields

checkPointAvailability

checkPointAvailability [1..1]

checkPointDescription

checkPointDescription [0..1]

cornerPoints

cornerPoints [0..*]

centerPoint

centerPoint [0..1]

pointInPixel

pointInPixel [1..1]

transformationDimensionDescription

transformationDimensionDescription [0..1]

transformationDimensionMapping

transformationDimensionMapping [0..2]

Methods

Public methods

ISOGeorectified$new()

ISOGeorectified$setCheckPointAvailability()

ISOGeorectified$setCheckPointDescription()

ISOGeorectified$addCornerPoint()

ISOGeorectified$delCornerPoint()

ISOGeorectified$setCenterPoint()

ISOGeorectified$setPixelOrientation()

ISOGeorectified$setTransformationDimensionDescription()

ISOGeorectified$addTransformationDimensionMapping()

ISOGeorectified$delTransformationDimensionMapping()

ISOGeorectified$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$checkMetadataStandardCompliance()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

geometa::ISOGridSpatialRepresentation$addDimension()

geometa::ISOGridSpatialRepresentation$delDimension()

geometa::ISOGridSpatialRepresentation$setCellGeometry()

geometa::ISOGridSpatialRepresentation$setNumberOfDimensions()

geometa::ISOGridSpatialRepresentation$setTransformationParameterAvailability()

Method new()

Initializes object

Usage

ISOGeorectified$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setCheckPointAvailability()

Set check point availability

Usage

ISOGeorectified$setCheckPointAvailability(availability)

Arguments

availability

object of class logical

Method setCheckPointDescription()

Set check point description

Usage

ISOGeorectified$setCheckPointDescription(description, locales = NULL)

Arguments

description

object of class character

locales

list of localized descriptions. Default is NULL

Method addCornerPoint()

Adds corner point

Usage

ISOGeorectified$addCornerPoint(sfg = NULL, m = NULL)

Arguments

sfg

simple feature object from sf

m

simple feature object of class matrix

Returns

TRUE if added, FALSE otherwise

Method delCornerPoint()

Deletes corner point

Usage

ISOGeorectified$delCornerPoint(sfg = NULL, m = NULL)

Arguments

sfg

simple feature object from sf

m

simple feature object of class matrix

Returns

TRUE if deleted, FALSE otherwise

Method setCenterPoint()

Sets center point

Usage

ISOGeorectified$setCenterPoint(sfg = NULL, m = NULL)

Arguments

sfg

simple feature object from sf

m

simple feature object of class matrix

Method setPixelOrientation()

Set pixel orientation

Usage

ISOGeorectified$setPixelOrientation(pixelOrientation)

Arguments

pixelOrientation

object of class ISOPixelOrientation or character among values among those returned by ISOPixelOrientation$values()

Method setTransformationDimensionDescription()

Set transformation dimension description

Usage

ISOGeorectified$setTransformationDimensionDescription(
  description,
  locales = NULL
)

Arguments

description

description

locales

list of localized descriptions. Default is NULL

Method addTransformationDimensionMapping()

Adds transformation dimension mapping

Usage

ISOGeorectified$addTransformationDimensionMapping(mapping)

Arguments

mapping

mapping

Returns

TRUE if added, FALSE otherwise

Method delTransformationDimensionMapping()

Deletes transformation dimension mapping

Usage

ISOGeorectified$delTransformationDimensionMapping(mapping)

Arguments

mapping

mapping

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOGeorectified$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.