ISOIdentification

ISOIdentification

ISOIdentification

Format

R6Class object.

Value

Object of R6Class for modelling an ISO Identification

References

ISO 19115:2003 - Geographic information -- Metadata

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOIdentification

Public fields

citation

citation: ISOCitation

abstract

abstract: character

purpose

purpose [0..1]: character

credit

credit [0..*]: character

status

status [0..*]: ISOStatus

pointOfContact

pointOfContact [0..*]: ISOResponsibleParty

resourceMaintenance

resourceMaintenance [0..*]: ISOMaintenanceInformation

graphicOverview

graphicOverview [0..*]: ISOBrowseGraphic

resourceFormat

resourceFormat [0..*]: ISOFormat

descriptiveKeywords

descriptiveKeywords [0..*]: ISOKeywords

resourceConstraints

resourceConstraints [0..*]: ISOLegalConstraints

resourceSpecificUsage

resourceSpecificUsage [0..*]: MD_Usage (ISOUsage - to implement)

aggregationInfo

aggregationInfo [0..*]: ISOAggregateInformation

Methods

Public methods

ISOIdentification$new()

ISOIdentification$setCitation()

ISOIdentification$setAbstract()

ISOIdentification$setPurpose()

ISOIdentification$addCredit()

ISOIdentification$delCredit()

ISOIdentification$addStatus()

ISOIdentification$delStatus()

ISOIdentification$addPointOfContact()

ISOIdentification$delPointOfContact()

ISOIdentification$addResourceMaintenance()

ISOIdentification$setResourceMaintenance()

ISOIdentification$delResourceMaintenance()

ISOIdentification$addGraphicOverview()

ISOIdentification$setGraphicOverview()

ISOIdentification$delGraphicOverview()

ISOIdentification$addFormat()

ISOIdentification$delFormat()

ISOIdentification$addKeywords()

ISOIdentification$setKeywords()

ISOIdentification$delKeywords()

ISOIdentification$addResourceConstraints()

ISOIdentification$setResourceConstraints()

ISOIdentification$delResourceConstraints()

ISOIdentification$addAggregateInformation()

ISOIdentification$delAggregateInformation()

ISOIdentification$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOIdentification$new(xml = NULL, defaults = list())

Arguments

xml

object of class XMLInternalNode-class

defaults

defaults list

Method setCitation()

Set citation

Usage

ISOIdentification$setCitation(citation)

Arguments

citation

object of class ISOCitation

Method setAbstract()

Set abstract

Usage

ISOIdentification$setAbstract(abstract, locales = NULL)

Arguments

abstract

abstract

locales

list of localized abstracts. Default is NULL

Method setPurpose()

Set purpose

Usage

ISOIdentification$setPurpose(purpose, locales = NULL)

Arguments

purpose

purpose

locales

list of localized texts. Default is NULL

Method addCredit()

Adds credit

Usage

ISOIdentification$addCredit(credit, locales = NULL)

Arguments

credit

credit

locales

list of localized texts. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delCredit()

Deletes credit

Usage

ISOIdentification$delCredit(credit, locales = NULL)

Arguments

credit

credit

locales

list of localized texts. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method addStatus()

Adds status

Usage

ISOIdentification$addStatus(status)

Arguments

status

object of class ISOStatus or any character among values returned by ISOStatus$values()

Returns

TRUE if added, FALSE otherwise

Method delStatus()

Deletes status

Usage

ISOIdentification$delStatus(status)

Arguments

status

object of class ISOStatus or any character among values returned by ISOStatus$values()

Returns

TRUE if deleted, FALSE otherwise

Method addPointOfContact()

Adds point of contact

Usage

ISOIdentification$addPointOfContact(pointOfContact)

Arguments

pointOfContact

object of class ISOResponsibleParty

Returns

TRUE if added, FALSE otherwise

Method delPointOfContact()

Deletes point of contact

Usage

ISOIdentification$delPointOfContact(pointOfContact)

Arguments

pointOfContact

object of class ISOResponsibleParty

Returns

TRUE if deleted, FALSE otherwise

Method addResourceMaintenance()

Adds resource maintenance

Usage

ISOIdentification$addResourceMaintenance(resourceMaintenance)

Arguments

resourceMaintenance

object of class ISOMaintenanceInformation

Returns

TRUE if added, FALSE otherwise

Method setResourceMaintenance()

Set resource maintenance

Usage

ISOIdentification$setResourceMaintenance(resourceMaintenance)

Arguments

resourceMaintenance

object of class ISOMaintenanceInformation

Returns

TRUE if set, FALSE otherwise

Method delResourceMaintenance()

Deletes resource maintenance

Usage

ISOIdentification$delResourceMaintenance(resourceMaintenance)

Arguments

resourceMaintenance

object of class ISOMaintenanceInformation

Returns

TRUE if deleted, FALSE otherwise

Method addGraphicOverview()

Adds graphic overview

Usage

ISOIdentification$addGraphicOverview(graphicOverview)

Arguments

graphicOverview

object of class ISOBrowseGraphic

Returns

TRUE if added, FALSE otherwise

Method setGraphicOverview()

Sets graphic overview

Usage

ISOIdentification$setGraphicOverview(graphicOverview)

Arguments

graphicOverview

object of class ISOBrowseGraphic

Returns

TRUE if set, FALSE otherwise

Method delGraphicOverview()

Deletes graphic overview

Usage

ISOIdentification$delGraphicOverview(graphicOverview)

Arguments

graphicOverview

object of class ISOBrowseGraphic

Returns

TRUE if deleted, FALSE otherwise

Method addFormat()

Adds format

Usage

ISOIdentification$addFormat(format)

Arguments

format

object of class ISOFormat

Returns

TRUE if added, FALSE otherwise

Method delFormat()

Deletes format

Usage

ISOIdentification$delFormat(format)

Arguments

format

object of class ISOFormat

Returns

TRUE if deleted, FALSE otherwise

Method addKeywords()

Adds keywords

Usage

ISOIdentification$addKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if added, FALSE otherwise

Method setKeywords()

Set keywords

Usage

ISOIdentification$setKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if set, FALSE otherwise

Method delKeywords()

Deletes keywords

Usage

ISOIdentification$delKeywords(keywords)

Arguments

keywords

object of class ISOKeywords

Returns

TRUE if deleted, FALSE otherwise

Method addResourceConstraints()

Adds resource constraints

Usage

ISOIdentification$addResourceConstraints(resourceConstraints)

Arguments

resourceConstraints

object of class ISOConstraints

Returns

TRUE if added, FALSE otherwise

Method setResourceConstraints()

Sets resource constraints

Usage

ISOIdentification$setResourceConstraints(resourceConstraints)

Arguments

resourceConstraints

object of class ISOConstraints

Returns

TRUE if set, FALSE otherwise

Method delResourceConstraints()

Deletes resource constraints

Usage

ISOIdentification$delResourceConstraints(resourceConstraints)

Arguments

resourceConstraints

object of class ISOConstraints

Returns

TRUE if deleted, FALSE otherwise

Method addAggregateInformation()

Adds aggregate information

Usage

ISOIdentification$addAggregateInformation(aggregateInfo)

Arguments

aggregateInfo

object of class ISOAggregateInformation

Returns

TRUE if added, FALSE otherwise

Method delAggregateInformation()

Deletes aggregate information

Usage

ISOIdentification$delAggregateInformation(aggregateInfo)

Arguments

aggregateInfo

object of class ISOAggregateInformation

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOIdentification$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.