ISOImageryObjective

ISOImageryObjective

ISOImageryObjective

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery objective

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryObjective

Public fields

identifier

identifier [1..1]: ISOMetaIdentifier

priority

priority [0..1]: character|ISOLocalisedCharacterString

type

type [0..*]: ISOImageryObjectiveType

function

function [0..*]: character|ISOLocalisedCharacterString

extent

extent [0..*]: ISOExtent

sensingInstrument

sensingInstrument [0..*]: ISOImageryInstrument

pass

pass [0..*]: ISOImageryPlatformPass

objectiveOccurance

objectiveOccurance [1..*]: ISOImageryEvent

Methods

Public methods

ISOImageryObjective$new()

ISOImageryObjective$setIdentifier()

ISOImageryObjective$setPriority()

ISOImageryObjective$addType()

ISOImageryObjective$delType()

ISOImageryObjective$addFunction()

ISOImageryObjective$delFunction()

ISOImageryObjective$addExtent()

ISOImageryObjective$delExtent()

ISOImageryObjective$addSensingInstrument()

ISOImageryObjective$delSensingInstrument()

ISOImageryObjective$addPlatformPass()

ISOImageryObjective$delPlatformPass()

ISOImageryObjective$addObjectiveOccurance()

ISOImageryObjective$delObjectiveOccurance()

ISOImageryObjective$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOImageryObjective$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setIdentifier()

Set identifier

Usage

ISOImageryObjective$setIdentifier(identifier)

Arguments

identifier

object of class ISOMetaIdentifier or character

Method setPriority()

Set priority

Usage

ISOImageryObjective$setPriority(priority, locales = NULL)

Arguments

priority

priority

locales

list of localized texts. Default is NULL

Method addType()

Adds type

Usage

ISOImageryObjective$addType(type)

Arguments

type

object of class ISOImageryObjectiveType or any character among values returned by ISOImageryObjectiveType$values()

Returns

TRUE if added, FALSE otherwise

Method delType()

Deletes type

Usage

ISOImageryObjective$delType(type)

Arguments

type

object of class ISOImageryObjectiveType or any character among values returned by ISOImageryObjectiveType$values()

Returns

TRUE if deleted, FALSE otherwise

Method addFunction()

Adds function

Usage

ISOImageryObjective$addFunction(fun, locales = NULL)

Arguments

fun

fun

locales

list of localized texts. Default is NULL

Returns

TRUE if added, FALSE otherwise

Method delFunction()

Deletes function

Usage

ISOImageryObjective$delFunction(fun, locales = NULL)

Arguments

fun

fun

locales

list of localized texts. Default is NULL

Returns

TRUE if deleted, FALSE otherwise

Method addExtent()

Adds extent

Usage

ISOImageryObjective$addExtent(extent)

Arguments

extent

extent, object of class ISOExtent

Returns

TRUE if added, FALSE otherwise

Method delExtent()

Deletes extent

Usage

ISOImageryObjective$delExtent(extent)

Arguments

extent

extent, object of class ISOExtent

Returns

TRUE if deleted, FALSE otherwise

Method addSensingInstrument()

Adds sensing instrument

Usage

ISOImageryObjective$addSensingInstrument(instrument)

Arguments

instrument

object of class ISOImageryInstrument

Returns

TRUE if added, FALSE otherwise

Method delSensingInstrument()

Deletes sensing instrument

Usage

ISOImageryObjective$delSensingInstrument(instrument)

Arguments

instrument

object of class ISOImageryInstrument

Returns

TRUE if deleted, FALSE otherwise

Method addPlatformPass()

Adds platform pass

Usage

ISOImageryObjective$addPlatformPass(pass)

Arguments

pass

object of class ISOImageryPlatformPass

Returns

TRUE if added, FALSE otherwise

Method delPlatformPass()

Deletes platform pass

Usage

ISOImageryObjective$delPlatformPass(pass)

Arguments

pass

object of class ISOImageryPlatformPass

Returns

TRUE if deleted, FALSE otherwise

Method addObjectiveOccurance()

Adds objective occurance

Usage

ISOImageryObjective$addObjectiveOccurance(event)

Arguments

event

object of class ISOImageryEvent

Returns

TRUE if added, FALSE otherwise

Method delObjectiveOccurance()

Deletes objective occurance

Usage

ISOImageryObjective$delObjectiveOccurance(event)

Arguments

event

object of class ISOImageryEvent

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryObjective$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

   #encoding
   md <- ISOImageryObjective$new()
   md$setIdentifier("identifier")
   md$setPriority("urgent")
   md$addType("survey")
#> [1] TRUE
   md$addFunction("my_function")
#> [1] TRUE
   evt <- ISOImageryEvent$new()
   evt$setIdentifier("event_1")
   evt$setTrigger("manual")
   evt$setContext("pass")
   evt$setSequence("instantaneous")
   evt$setTime(Sys.time())
   md$addObjectiveOccurance(evt)
#> [1] TRUE
   extent <- ISOExtent$new()
   bbox <- ISOGeographicBoundingBox$new(minx = -180, miny = -90, maxx = 180, maxy = 90)
   extent$addGeographicElement(bbox)
#> [1] TRUE
   time <- ISOTemporalExtent$new()
   start <- ISOdate(2000, 1, 12, 12, 59, 45)
   end <- ISOdate(2010, 8, 22, 13, 12, 43)
   tp <- GMLTimePeriod$new(beginPosition = start, endPosition = end)
   time$setTimePeriod(tp)
   extent$addTemporalElement(time)
#> [1] TRUE
   vert <- ISOVerticalExtent$new()
   vert$setMinimumValue(0)
   vert$setMaximumValue(19)
   extent$addVerticalElement(vert)
#> [1] TRUE
   md$addExtent(extent)
#> [1] TRUE
   md$sensingInstrument = NA
   md$pass = NA
   xml <- md$encode()
#> [geometa][WARN] Element '{http://www.isotc211.org/2005/gmi}MI_Objective': No matching global declaration available for the validation root at line 1. 
#> [geometa][WARN] Object 'ISOImageryObjective' is INVALID according to ISO 19139 XML schemas!