ISOImageryOperation

ISOImageryOperation

ISOImageryOperation

Format

R6Class object.

Value

Object of R6Class for modelling an ISO imagery Operation

References

ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

geometa::geometaLogger -> geometa::ISOAbstractObject -> ISOImageryOperation

Public fields

description

description [0..1]: character|ISOLocalisedCharacterString

citation

citation [0..1]: ISOCitation

identifier

identifier [1..1]: ISOMetaIdentifier

status

status [1..1]: ISOStatus

type

type [0..1]: ISOImageryOperationType

parentOperation

parentOperation [1..1]: ISOImageryOperation

childOperation

childOperation [0..*]: ISOImageryOperation

platform

platform [0..*]: ISOImageryPlatform

objective

objective [0..*]: ISOImageryObjective

plan

plan [0..1]: ISOImageryPlan

significantEvent

significantEvent [0..*]: ISOImageryEvent

Methods

Public methods

ISOImageryOperation$new()

ISOImageryOperation$setDescription()

ISOImageryOperation$setCitation()

ISOImageryOperation$setIdentifier()

ISOImageryOperation$setStatus()

ISOImageryOperation$setType()

ISOImageryOperation$setParentOperation()

ISOImageryOperation$addChildOperation()

ISOImageryOperation$delChildOperation()

ISOImageryOperation$addPlatform()

ISOImageryOperation$delPlatform()

ISOImageryOperation$addObjective()

ISOImageryOperation$delObjective()

ISOImageryOperation$setPlan()

ISOImageryOperation$addSignificantEvent()

ISOImageryOperation$delSignificantEvent()

ISOImageryOperation$clone()

Inherited methods

geometa::geometaLogger$ERROR()

geometa::geometaLogger$INFO()

geometa::geometaLogger$WARN()

geometa::ISOAbstractObject$addFieldAttrs()

geometa::ISOAbstractObject$addListElement()

geometa::ISOAbstractObject$contains()

geometa::ISOAbstractObject$createLocalisedProperty()

geometa::ISOAbstractObject$decode()

geometa::ISOAbstractObject$delListElement()

geometa::ISOAbstractObject$encode()

geometa::ISOAbstractObject$getClass()

geometa::ISOAbstractObject$getClassName()

geometa::ISOAbstractObject$getNamespaceDefinition()

geometa::ISOAbstractObject$isDocument()

geometa::ISOAbstractObject$isFieldInheritedFrom()

geometa::ISOAbstractObject$print()

geometa::ISOAbstractObject$save()

geometa::ISOAbstractObject$setAttr()

geometa::ISOAbstractObject$setCodeList()

geometa::ISOAbstractObject$setCodeListValue()

geometa::ISOAbstractObject$setCodeSpace()

geometa::ISOAbstractObject$setHref()

geometa::ISOAbstractObject$setId()

geometa::ISOAbstractObject$setIsNull()

geometa::ISOAbstractObject$setValue()

geometa::ISOAbstractObject$validate()

geometa::ISOAbstractObject$wrapBaseElement()

Method new()

Initializes object

Usage

ISOImageryOperation$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class

Method setDescription()

Set description

Usage

ISOImageryOperation$setDescription(description, locales = NULL)

Arguments

description

description

locales

list of localized texts. Default is NULL

Method setCitation()

Set citation

Usage

ISOImageryOperation$setCitation(citation)

Arguments

citation

object of class ISOCitation

Method setIdentifier()

Set identifier

Usage

ISOImageryOperation$setIdentifier(identifier)

Arguments

identifier

object of class ISOMetaIdentifier or character

Method setStatus()

Set status

Usage

ISOImageryOperation$setStatus(status)

Arguments

status

object of class ISOStatus or any character among values returned by ISOStatus$values()

Method setType()

Set type

Usage

ISOImageryOperation$setType(type)

Arguments

type

object of class ISOImageryOperationType or any character among values returned by ISOImageryOperationType$values()

Method setParentOperation()

Set parent operation

Usage

ISOImageryOperation$setParentOperation(operation)

Arguments

operation

object of class ISOImageryOperation

Method addChildOperation()

Adds child operation

Usage

ISOImageryOperation$addChildOperation(operation)

Arguments

operation

object of class ISOImageryOperation

Returns

TRUE if added, FALSE otherwise

Method delChildOperation()

Deletes child operation

Usage

ISOImageryOperation$delChildOperation(operation)

Arguments

operation

object of class ISOImageryOperation

Returns

TRUE if deleted, FALSE otherwise

Method addPlatform()

Adds platform

Usage

ISOImageryOperation$addPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if added, FALSE otherwise

Method delPlatform()

Deletes platform

Usage

ISOImageryOperation$delPlatform(platform)

Arguments

platform

object of class ISOImageryPlatform

Returns

TRUE if deleted, FALSE otherwise

Method addObjective()

Adds objective

Usage

ISOImageryOperation$addObjective(objective)

Arguments

objective

object of class ISOImageryObjective

Returns

TRUE if added, FALSE otherwise

Method delObjective()

Deletes objective

Usage

ISOImageryOperation$delObjective(objective)

Arguments

objective

object of class ISOImageryObjective

Returns

TRUE if deleted, FALSE otherwise

Method setPlan()

Set plan

Usage

ISOImageryOperation$setPlan(plan)

Arguments

plan

object of class ISOImageryPlan

Method addSignificantEvent()

Adds significative event

Usage

ISOImageryOperation$addSignificantEvent(event)

Arguments

event

object of class ISOImageryEvent

Returns

TRUE if added, FALSE otherwise

Method delSignificantEvent()

Deletes significative event

Usage

ISOImageryOperation$delSignificantEvent(event)

Arguments

event

object of class ISOImageryEvent

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

ISOImageryOperation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.