getISOClasses

get the list of ISO classes, ie classes extending ISOAbstractObject super class, including classes eventually defined outside geometa. In case the latter is on the search path, the list of ISO classes will be cached for optimized used by geometa encoder/decoder.

getISOClasses()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOClasses()
#>   [1] "GMLAbstractCRS"                          
#>   [2] "GMLAbstractCoordinateOperation"          
#>   [3] "GMLAbstractCoordinateSystem"             
#>   [4] "GMLAbstractCoverage"                     
#>   [5] "GMLAbstractCurve"                        
#>   [6] "GMLAbstractDiscreteCoverage"             
#>   [7] "GMLAbstractFeature"                      
#>   [8] "GMLAbstractGML"                          
#>   [9] "GMLAbstractGeneralConversion"            
#>  [10] "GMLAbstractGeneralDerivedCRS"            
#>  [11] "GMLAbstractGeneralOperationParameter"    
#>  [12] "GMLAbstractGeneralParameterValue"        
#>  [13] "GMLAbstractGeometricAggregate"           
#>  [14] "GMLAbstractGeometricPrimitive"           
#>  [15] "GMLAbstractGeometry"                     
#>  [16] "GMLAbstractImplicitGeometry"             
#>  [17] "GMLAbstractObject"                       
#>  [18] "GMLAbstractReferenceableGrid"            
#>  [19] "GMLAbstractRing"                         
#>  [20] "GMLAbstractSingleCRS"                    
#>  [21] "GMLAbstractSingleOperation"              
#>  [22] "GMLAbstractSurface"                      
#>  [23] "GMLAbstractTimeGeometricPrimitive"       
#>  [24] "GMLAbstractTimeObject"                   
#>  [25] "GMLAbstractTimePrimitive"                
#>  [26] "GMLAffineCS"                             
#>  [27] "GMLBaseUnit"                             
#>  [28] "GMLCOVAbstractCoverage"                  
#>  [29] "GMLCOVExtension"                         
#>  [30] "GMLCartesianCS"                          
#>  [31] "GMLCodeType"                             
#>  [32] "GMLCompoundCRS"                          
#>  [33] "GMLConventionalUnit"                     
#>  [34] "GMLConversion"                           
#>  [35] "GMLCoordinateSystemAxis"                 
#>  [36] "GMLCylindricalCS"                        
#>  [37] "GMLDefinition"                           
#>  [38] "GMLDerivedCRS"                           
#>  [39] "GMLDerivedUnit"                          
#>  [40] "GMLElement"                              
#>  [41] "GMLEllipsoidalCS"                        
#>  [42] "GMLEnvelope"                             
#>  [43] "GMLEnvelopeWithTimePeriod"               
#>  [44] "GMLGeneralGridAxis"                      
#>  [45] "GMLGeodeticCRS"                          
#>  [46] "GMLGrid"                                 
#>  [47] "GMLGridCoverage"                         
#>  [48] "GMLGridEnvelope"                         
#>  [49] "GMLGridFunction"                         
#>  [50] "GMLLineString"                           
#>  [51] "GMLLinearCS"                             
#>  [52] "GMLLinearRing"                           
#>  [53] "GMLMultiCurve"                           
#>  [54] "GMLMultiCurveCoverage"                   
#>  [55] "GMLMultiPoint"                           
#>  [56] "GMLMultiPointCoverage"                   
#>  [57] "GMLMultiSolidCoverage"                   
#>  [58] "GMLMultiSurface"                         
#>  [59] "GMLMultiSurfaceCoverage"                 
#>  [60] "GMLObliqueCartesianCS"                   
#>  [61] "GMLOperationMethod"                      
#>  [62] "GMLOperationParameter"                   
#>  [63] "GMLOperationParameterGroup"              
#>  [64] "GMLParameterValue"                       
#>  [65] "GMLParameterValueGroup"                  
#>  [66] "GMLPoint"                                
#>  [67] "GMLPolarCS"                              
#>  [68] "GMLPolygon"                              
#>  [69] "GMLProjectedCRS"                         
#>  [70] "GMLRectifiedGrid"                        
#>  [71] "GMLRectifiedGridCoverage"                
#>  [72] "GMLReferenceableGridByArray"             
#>  [73] "GMLReferenceableGridByTransformation"    
#>  [74] "GMLReferenceableGridByVectors"           
#>  [75] "GMLSphericalCS"                          
#>  [76] "GMLTemporalCRS"                          
#>  [77] "GMLTemporalCS"                           
#>  [78] "GMLTimeCS"                               
#>  [79] "GMLTimeInstant"                          
#>  [80] "GMLTimePeriod"                           
#>  [81] "GMLUnitDefinition"                       
#>  [82] "GMLUserDefinedCS"                        
#>  [83] "GMLVerticalCRS"                          
#>  [84] "GMLVerticalCS"                           
#>  [85] "ISOAbsoluteExternalPositionalAccuracy"   
#>  [86] "ISOAbstractAggregate"                    
#>  [87] "ISOAbstractCarrierOfCharacteristics"     
#>  [88] "ISOAbstractCatalogue"                    
#>  [89] "ISOAbstractCompleteness"                 
#>  [90] "ISOAbstractContentInformation"           
#>  [91] "ISOAbstractGenericName"                  
#>  [92] "ISOAbstractLogicalConsistency"           
#>  [93] "ISOAbstractParty"                        
#>  [94] "ISOAbstractPositionalAccuracy"           
#>  [95] "ISOAbstractPropertyType"                 
#>  [96] "ISOAbstractReferenceSystem"              
#>  [97] "ISOAbstractResponsibility"               
#>  [98] "ISOAbstractResult"                       
#>  [99] "ISOAbstractTemporalAccuracy"             
#> [100] "ISOAbstractThematicAccuracy"             
#> [101] "ISOAccuracyOfATimeMeasurement"           
#> [102] "ISOAddress"                              
#> [103] "ISOAggregateInformation"                 
#> [104] "ISOAnchor"                               
#> [105] "ISOAngle"                                
#> [106] "ISOApplicationSchemaInformation"         
#> [107] "ISOAssociation"                          
#> [108] "ISOAssociationRole"                      
#> [109] "ISOAssociationType"                      
#> [110] "ISOBand"                                 
#> [111] "ISOBaseBoolean"                          
#> [112] "ISOBaseCharacterString"                  
#> [113] "ISOBaseDate"                             
#> [114] "ISOBaseDateTime"                         
#> [115] "ISOBaseDecimal"                          
#> [116] "ISOBaseInteger"                          
#> [117] "ISOBaseReal"                             
#> [118] "ISOBinary"                               
#> [119] "ISOBinding"                              
#> [120] "ISOBoundAssociationRole"                 
#> [121] "ISOBoundFeatureAttribute"                
#> [122] "ISOBoundingPolygon"                      
#> [123] "ISOBrowseGraphic"                        
#> [124] "ISOCarrierOfCharacteristics"             
#> [125] "ISOCellGeometry"                         
#> [126] "ISOCharacterSet"                         
#> [127] "ISOCitation"                             
#> [128] "ISOCitationSeries"                       
#> [129] "ISOClassification"                       
#> [130] "ISOCodeDefinition"                       
#> [131] "ISOCodeListDictionary"                   
#> [132] "ISOCodeListValue"                        
#> [133] "ISOCodelist"                             
#> [134] "ISOCodelistCatalogue"                    
#> [135] "ISOCodelistValue"                        
#> [136] "ISOCompletenessCommission"               
#> [137] "ISOCompletenessOmission"                 
#> [138] "ISOConceptualConsistency"                
#> [139] "ISOConformanceResult"                    
#> [140] "ISOConstraint"                           
#> [141] "ISOConstraints"                          
#> [142] "ISOContact"                              
#> [143] "ISOCountry"                              
#> [144] "ISOCoupledResource"                      
#> [145] "ISOCouplingType"                         
#> [146] "ISOCoverageContentType"                  
#> [147] "ISOCoverageDescription"                  
#> [148] "ISODCPList"                              
#> [149] "ISODataFile"                             
#> [150] "ISODataIdentification"                   
#> [151] "ISODataQuality"                          
#> [152] "ISODataQualityAbstractElement"           
#> [153] "ISODataQualityScope"                     
#> [154] "ISODataSet"                              
#> [155] "ISODatatype"                             
#> [156] "ISODate"                                 
#> [157] "ISODateType"                             
#> [158] "ISODefinitionReference"                  
#> [159] "ISODefinitionSource"                     
#> [160] "ISODigitalTransferOptions"               
#> [161] "ISODimension"                            
#> [162] "ISODimensionNameType"                    
#> [163] "ISODistance"                             
#> [164] "ISODistribution"                         
#> [165] "ISODistributionUnits"                    
#> [166] "ISODistributor"                          
#> [167] "ISODomainConsistency"                    
#> [168] "ISOElementSequence"                      
#> [169] "ISOEvaluationMethodType"                 
#> [170] "ISOExtendedElementInformation"           
#> [171] "ISOExtent"                               
#> [172] "ISOFeatureAssociation"                   
#> [173] "ISOFeatureAttribute"                     
#> [174] "ISOFeatureCatalogue"                     
#> [175] "ISOFeatureCatalogueDescription"          
#> [176] "ISOFeatureOperation"                     
#> [177] "ISOFeatureType"                          
#> [178] "ISOFileName"                             
#> [179] "ISOFormat"                               
#> [180] "ISOFormatConsistency"                    
#> [181] "ISOFreeText"                             
#> [182] "ISOGeographicBoundingBox"                
#> [183] "ISOGeographicDescription"                
#> [184] "ISOGeographicExtent"                     
#> [185] "ISOGeometricObjectType"                  
#> [186] "ISOGeometricObjects"                     
#> [187] "ISOGeorectified"                         
#> [188] "ISOGeoreferenceable"                     
#> [189] "ISOGridSpatialRepresentation"            
#> [190] "ISOGriddedDataPositionalAccuracy"        
#> [191] "ISOIdentification"                       
#> [192] "ISOIdentifier"                           
#> [193] "ISOImageDescription"                     
#> [194] "ISOImageryAbstractGeolocationInformation"
#> [195] "ISOImageryAcquisitionInformation"        
#> [196] "ISOImageryAlgorithm"                     
#> [197] "ISOImageryBand"                          
#> [198] "ISOImageryBandDefinition"                
#> [199] "ISOImageryContext"                       
#> [200] "ISOImageryCoverageDescription"           
#> [201] "ISOImageryCoverageResult"                
#> [202] "ISOImageryEnvironmentalRecord"           
#> [203] "ISOImageryEvent"                         
#> [204] "ISOImageryGCP"                           
#> [205] "ISOImageryGCPCollection"                 
#> [206] "ISOImageryGeometryType"                  
#> [207] "ISOImageryGeorectified"                  
#> [208] "ISOImageryGeoreferenceable"              
#> [209] "ISOImageryImageDescription"              
#> [210] "ISOImageryInstrument"                    
#> [211] "ISOImageryMetadata"                      
#> [212] "ISOImageryNominalResolution"             
#> [213] "ISOImageryObjective"                     
#> [214] "ISOImageryObjectiveType"                 
#> [215] "ISOImageryOperation"                     
#> [216] "ISOImageryOperationType"                 
#> [217] "ISOImageryPlan"                          
#> [218] "ISOImageryPlatform"                      
#> [219] "ISOImageryPlatformPass"                  
#> [220] "ISOImageryPolarisationOrientation"       
#> [221] "ISOImageryPriority"                      
#> [222] "ISOImageryProcessStep"                   
#> [223] "ISOImageryProcessStepReport"             
#> [224] "ISOImageryProcessing"                    
#> [225] "ISOImageryRangeElementDescription"       
#> [226] "ISOImageryRequestedDate"                 
#> [227] "ISOImageryRequirement"                   
#> [228] "ISOImagerySensorType"                    
#> [229] "ISOImagerySequence"                      
#> [230] "ISOImagerySource"                        
#> [231] "ISOImageryTransferFunctionType"          
#> [232] "ISOImageryTrigger"                       
#> [233] "ISOImageryUsability"                     
#> [234] "ISOImagingCondition"                     
#> [235] "ISOIndividual"                           
#> [236] "ISOInheritanceRelation"                  
#> [237] "ISOInitiative"                           
#> [238] "ISOInitiativeType"                       
#> [239] "ISOKeywordType"                          
#> [240] "ISOKeywords"                             
#> [241] "ISOLanguage"                             
#> [242] "ISOLegalConstraints"                     
#> [243] "ISOLength"                               
#> [244] "ISOLineage"                              
#> [245] "ISOListedValue"                          
#> [246] "ISOLocalName"                            
#> [247] "ISOLocale"                               
#> [248] "ISOLocaleContainer"                      
#> [249] "ISOLocalisedCharacterString"             
#> [250] "ISOMLCodeDefinition"                     
#> [251] "ISOMLCodeListDictionary"                 
#> [252] "ISOMaintenanceFrequency"                 
#> [253] "ISOMaintenanceInformation"               
#> [254] "ISOMeasure"                              
#> [255] "ISOMedium"                               
#> [256] "ISOMediumFormat"                         
#> [257] "ISOMediumName"                           
#> [258] "ISOMemberName"                           
#> [259] "ISOMetaIdentifier"                       
#> [260] "ISOMetadata"                             
#> [261] "ISOMetadataExtensionInformation"         
#> [262] "ISOMetadataScope"                        
#> [263] "ISOMimeFileType"                         
#> [264] "ISOMultiplicity"                         
#> [265] "ISOMultiplicityRange"                    
#> [266] "ISONonQuantitativeAttributeAccuracy"     
#> [267] "ISOObligation"                           
#> [268] "ISOOnLineFunction"                       
#> [269] "ISOOnlineResource"                       
#> [270] "ISOOperationMetadata"                    
#> [271] "ISOOrganisation"                         
#> [272] "ISOOtherAggregate"                       
#> [273] "ISOParameter"                            
#> [274] "ISOParameterDirection"                   
#> [275] "ISOPeriodDuration"                       
#> [276] "ISOPixelOrientation"                     
#> [277] "ISOPlatform"                             
#> [278] "ISOPortrayalCatalogueReference"          
#> [279] "ISOPresentationForm"                     
#> [280] "ISOProcessStep"                          
#> [281] "ISOProductionSeries"                     
#> [282] "ISOProgress"                             
#> [283] "ISOPropertyType"                         
#> [284] "ISOQuantitativeAttributeAccuracy"        
#> [285] "ISOQuantitativeResult"                   
#> [286] "ISORangeDimension"                       
#> [287] "ISORecord"                               
#> [288] "ISORecordType"                           
#> [289] "ISOReferenceIdentifier"                  
#> [290] "ISOReferenceSystem"                      
#> [291] "ISOReferenceSystemType"                  
#> [292] "ISORelativeInternalPositionalAccuracy"   
#> [293] "ISOReleasability"                        
#> [294] "ISORepresentativeFraction"               
#> [295] "ISOResolution"                           
#> [296] "ISOResponsibility"                       
#> [297] "ISOResponsibleParty"                     
#> [298] "ISORestriction"                          
#> [299] "ISORole"                                 
#> [300] "ISORoleType"                             
#> [301] "ISOSRVServiceIdentification"             
#> [302] "ISOSampleDimension"                      
#> [303] "ISOScale"                                
#> [304] "ISOScope"                                
#> [305] "ISOScopeDescription"                     
#> [306] "ISOScopedName"                           
#> [307] "ISOSecurityConstraints"                  
#> [308] "ISOSensor"                               
#> [309] "ISOSeries"                               
#> [310] "ISOServiceIdentification"                
#> [311] "ISOSource"                               
#> [312] "ISOSpatialRepresentation"                
#> [313] "ISOSpatialRepresentationType"            
#> [314] "ISOSpatialTemporalExtent"                
#> [315] "ISOStandardOrderProcess"                 
#> [316] "ISOStatus"                               
#> [317] "ISOStereoMate"                           
#> [318] "ISOTelephone"                            
#> [319] "ISOTelephoneType"                        
#> [320] "ISOTemporalConsistency"                  
#> [321] "ISOTemporalExtent"                       
#> [322] "ISOTemporalValidity"                     
#> [323] "ISOThematicClassificationCorrectness"    
#> [324] "ISOTopicCategory"                        
#> [325] "ISOTopologicalConsistency"               
#> [326] "ISOTopologyLevel"                        
#> [327] "ISOTypeName"                             
#> [328] "ISOURL"                                  
#> [329] "ISOUnlimitedInteger"                     
#> [330] "ISOUomIdentifier"                        
#> [331] "ISOUsage"                                
#> [332] "ISOVectorSpatialRepresentation"          
#> [333] "ISOVerticalExtent"                       
#> [334] "SWEAbstractDataComponent"                
#> [335] "SWEAbstractEncoding"                     
#> [336] "SWEAbstractObject"                       
#> [337] "SWEAbstractSWE"                          
#> [338] "SWEAbstractSWEIdentifiable"              
#> [339] "SWEAbstractSimpleComponent"              
#> [340] "SWECategory"                             
#> [341] "SWECategoryRange"                        
#> [342] "SWECount"                                
#> [343] "SWECountRange"                           
#> [344] "SWEDataRecord"                           
#> [345] "SWEElement"                              
#> [346] "SWENilValues"                            
#> [347] "SWEQuantity"                             
#> [348] "SWEQuantityRange"                        
#> [349] "SWEText"                                 
#> [350] "SWETextEncoding"                         
#> [351] "SWETime"                                 
#> [352] "SWETimeRange"                            
#> [353] "SWEXMLEncoding"