getISOClasses

get the list of ISO classes, ie classes extending ISOAbstractObject super class, including classes eventually defined outside geometa. In case the latter is on the search path, the list of ISO classes will be cached for optimized used by geometa encoder/decoder.

getISOClasses()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOClasses()
#>   [1] "GMLAbstractCRS"                          
#>   [2] "GMLAbstractCoordinateOperation"          
#>   [3] "GMLAbstractCoordinateSystem"             
#>   [4] "GMLAbstractCoverage"                     
#>   [5] "GMLAbstractCurve"                        
#>   [6] "GMLAbstractDiscreteCoverage"             
#>   [7] "GMLAbstractFeature"                      
#>   [8] "GMLAbstractGML"                          
#>   [9] "GMLAbstractGeneralConversion"            
#>  [10] "GMLAbstractGeneralDerivedCRS"            
#>  [11] "GMLAbstractGeneralOperationParameter"    
#>  [12] "GMLAbstractGeneralParameterValue"        
#>  [13] "GMLAbstractGeometricAggregate"           
#>  [14] "GMLAbstractGeometricPrimitive"           
#>  [15] "GMLAbstractGeometry"                     
#>  [16] "GMLAbstractImplicitGeometry"             
#>  [17] "GMLAbstractObject"                       
#>  [18] "GMLAbstractReferenceableGrid"            
#>  [19] "GMLAbstractRing"                         
#>  [20] "GMLAbstractSingleCRS"                    
#>  [21] "GMLAbstractSingleOperation"              
#>  [22] "GMLAbstractSurface"                      
#>  [23] "GMLAbstractTimeGeometricPrimitive"       
#>  [24] "GMLAbstractTimeObject"                   
#>  [25] "GMLAbstractTimePrimitive"                
#>  [26] "GMLAffineCS"                             
#>  [27] "GMLBaseUnit"                             
#>  [28] "GMLCOVAbstractCoverage"                  
#>  [29] "GMLCOVExtension"                         
#>  [30] "GMLCartesianCS"                          
#>  [31] "GMLCodeType"                             
#>  [32] "GMLCompoundCRS"                          
#>  [33] "GMLConventionalUnit"                     
#>  [34] "GMLConversion"                           
#>  [35] "GMLCoordinateSystemAxis"                 
#>  [36] "GMLCylindricalCS"                        
#>  [37] "GMLDefinition"                           
#>  [38] "GMLDerivedCRS"                           
#>  [39] "GMLDerivedUnit"                          
#>  [40] "GMLElement"                              
#>  [41] "GMLEllipsoidalCS"                        
#>  [42] "GMLEnvelope"                             
#>  [43] "GMLEnvelopeWithTimePeriod"               
#>  [44] "GMLGeneralGridAxis"                      
#>  [45] "GMLGeodeticCRS"                          
#>  [46] "GMLGrid"                                 
#>  [47] "GMLGridCoverage"                         
#>  [48] "GMLGridEnvelope"                         
#>  [49] "GMLGridFunction"                         
#>  [50] "GMLLineString"                           
#>  [51] "GMLLinearCS"                             
#>  [52] "GMLLinearRing"                           
#>  [53] "GMLMultiCurve"                           
#>  [54] "GMLMultiCurveCoverage"                   
#>  [55] "GMLMultiPoint"                           
#>  [56] "GMLMultiPointCoverage"                   
#>  [57] "GMLMultiSolidCoverage"                   
#>  [58] "GMLMultiSurface"                         
#>  [59] "GMLMultiSurfaceCoverage"                 
#>  [60] "GMLObliqueCartesianCS"                   
#>  [61] "GMLOperationMethod"                      
#>  [62] "GMLOperationParameter"                   
#>  [63] "GMLOperationParameterGroup"              
#>  [64] "GMLParameterValue"                       
#>  [65] "GMLParameterValueGroup"                  
#>  [66] "GMLPoint"                                
#>  [67] "GMLPolarCS"                              
#>  [68] "GMLPolygon"                              
#>  [69] "GMLProjectedCRS"                         
#>  [70] "GMLRectifiedGrid"                        
#>  [71] "GMLRectifiedGridCoverage"                
#>  [72] "GMLReferenceableGridByArray"             
#>  [73] "GMLReferenceableGridByTransformation"    
#>  [74] "GMLReferenceableGridByVectors"           
#>  [75] "GMLSphericalCS"                          
#>  [76] "GMLTemporalCRS"                          
#>  [77] "GMLTemporalCS"                           
#>  [78] "GMLTimeCS"                               
#>  [79] "GMLTimeInstant"                          
#>  [80] "GMLTimePeriod"                           
#>  [81] "GMLUnitDefinition"                       
#>  [82] "GMLUserDefinedCS"                        
#>  [83] "GMLVerticalCRS"                          
#>  [84] "GMLVerticalCS"                           
#>  [85] "ISOAbsoluteExternalPositionalAccuracy"   
#>  [86] "ISOAbstractAcquisitionInformation"       
#>  [87] "ISOAbstractAggregate"                    
#>  [88] "ISOAbstractApplicationSchemaInformation" 
#>  [89] "ISOAbstractCarrierOfCharacteristics"     
#>  [90] "ISOAbstractCatalogue"                    
#>  [91] "ISOAbstractCitation"                     
#>  [92] "ISOAbstractCompleteness"                 
#>  [93] "ISOAbstractConstraints"                  
#>  [94] "ISOAbstractContentInformation"           
#>  [95] "ISOAbstractDataQuality"                  
#>  [96] "ISOAbstractDistribution"                 
#>  [97] "ISOAbstractExtent"                       
#>  [98] "ISOAbstractFeatureType"                  
#>  [99] "ISOAbstractFormat"                       
#> [100] "ISOAbstractGenericName"                  
#> [101] "ISOAbstractLineageInformation"           
#> [102] "ISOAbstractLogicalConsistency"           
#> [103] "ISOAbstractMDContentInformation"         
#> [104] "ISOAbstractMaintenanceInformation"       
#> [105] "ISOAbstractMetadata"                     
#> [106] "ISOAbstractMetadataExtension"            
#> [107] "ISOAbstractOnlineResource"               
#> [108] "ISOAbstractParameter"                    
#> [109] "ISOAbstractParty"                        
#> [110] "ISOAbstractPlatform"                     
#> [111] "ISOAbstractPortrayalCatalogueInformation"
#> [112] "ISOAbstractPositionalAccuracy"           
#> [113] "ISOAbstractPropertyType"                 
#> [114] "ISOAbstractQualityElement"               
#> [115] "ISOAbstractRSReferenceSystem"            
#> [116] "ISOAbstractReferenceSystem"              
#> [117] "ISOAbstractResourceDescription"          
#> [118] "ISOAbstractResponsibility"               
#> [119] "ISOAbstractResult"                       
#> [120] "ISOAbstractSpatialRepresentation"        
#> [121] "ISOAbstractSpatialResolution"            
#> [122] "ISOAbstractStandardOrderProcess"         
#> [123] "ISOAbstractTemporalAccuracy"             
#> [124] "ISOAbstractThematicAccuracy"             
#> [125] "ISOAbstractTypedDate"                    
#> [126] "ISOAccuracyOfATimeMeasurement"           
#> [127] "ISOAddress"                              
#> [128] "ISOAggregateInformation"                 
#> [129] "ISOAnchor"                               
#> [130] "ISOAngle"                                
#> [131] "ISOApplicationSchemaInformation"         
#> [132] "ISOAssociatedResource"                   
#> [133] "ISOAssociation"                          
#> [134] "ISOAssociationRole"                      
#> [135] "ISOAssociationType"                      
#> [136] "ISOAttributeGroup"                       
#> [137] "ISOBand"                                 
#> [138] "ISOBaseBoolean"                          
#> [139] "ISOBaseCharacterString"                  
#> [140] "ISOBaseDate"                             
#> [141] "ISOBaseDateTime"                         
#> [142] "ISOBaseDecimal"                          
#> [143] "ISOBaseInteger"                          
#> [144] "ISOBaseReal"                             
#> [145] "ISOBinary"                               
#> [146] "ISOBinding"                              
#> [147] "ISOBoundAssociationRole"                 
#> [148] "ISOBoundFeatureAttribute"                
#> [149] "ISOBoundingPolygon"                      
#> [150] "ISOBrowseGraphic"                        
#> [151] "ISOCarrierOfCharacteristics"             
#> [152] "ISOCellGeometry"                         
#> [153] "ISOCharacterSet"                         
#> [154] "ISOCitation"                             
#> [155] "ISOCitationSeries"                       
#> [156] "ISOClassification"                       
#> [157] "ISOCodeDefinition"                       
#> [158] "ISOCodeListDictionary"                   
#> [159] "ISOCodeListValue"                        
#> [160] "ISOCodelist"                             
#> [161] "ISOCodelistCatalogue"                    
#> [162] "ISOCodelistValue"                        
#> [163] "ISOCompletenessCommission"               
#> [164] "ISOCompletenessOmission"                 
#> [165] "ISOConceptualConsistency"                
#> [166] "ISOConformanceResult"                    
#> [167] "ISOConstraint"                           
#> [168] "ISOConstraints"                          
#> [169] "ISOContact"                              
#> [170] "ISOCountry"                              
#> [171] "ISOCoupledResource"                      
#> [172] "ISOCouplingType"                         
#> [173] "ISOCoverageContentType"                  
#> [174] "ISOCoverageDescription"                  
#> [175] "ISODCPList"                              
#> [176] "ISODataFile"                             
#> [177] "ISODataIdentification"                   
#> [178] "ISODataQuality"                          
#> [179] "ISODataQualityAbstractElement"           
#> [180] "ISODataQualityScope"                     
#> [181] "ISODataSet"                              
#> [182] "ISODatatype"                             
#> [183] "ISODate"                                 
#> [184] "ISODateType"                             
#> [185] "ISODefinitionReference"                  
#> [186] "ISODefinitionSource"                     
#> [187] "ISODigitalTransferOptions"               
#> [188] "ISODimension"                            
#> [189] "ISODimensionNameType"                    
#> [190] "ISODistance"                             
#> [191] "ISODistribution"                         
#> [192] "ISODistributionUnits"                    
#> [193] "ISODistributor"                          
#> [194] "ISODomainConsistency"                    
#> [195] "ISOElementSequence"                      
#> [196] "ISOEvaluationMethodType"                 
#> [197] "ISOExtendedElementInformation"           
#> [198] "ISOExtent"                               
#> [199] "ISOFeatureAssociation"                   
#> [200] "ISOFeatureAttribute"                     
#> [201] "ISOFeatureCatalogue"                     
#> [202] "ISOFeatureCatalogueDescription"          
#> [203] "ISOFeatureOperation"                     
#> [204] "ISOFeatureType"                          
#> [205] "ISOFeatureTypeInfo"                      
#> [206] "ISOFileName"                             
#> [207] "ISOFormat"                               
#> [208] "ISOFormatConsistency"                    
#> [209] "ISOFreeText"                             
#> [210] "ISOGeographicBoundingBox"                
#> [211] "ISOGeographicDescription"                
#> [212] "ISOGeographicExtent"                     
#> [213] "ISOGeometricObjectType"                  
#> [214] "ISOGeometricObjects"                     
#> [215] "ISOGeorectified"                         
#> [216] "ISOGeoreferenceable"                     
#> [217] "ISOGridSpatialRepresentation"            
#> [218] "ISOGriddedDataPositionalAccuracy"        
#> [219] "ISOIdentification"                       
#> [220] "ISOIdentifier"                           
#> [221] "ISOImageDescription"                     
#> [222] "ISOImageryAbstractGeolocationInformation"
#> [223] "ISOImageryAcquisitionInformation"        
#> [224] "ISOImageryAlgorithm"                     
#> [225] "ISOImageryBand"                          
#> [226] "ISOImageryBandDefinition"                
#> [227] "ISOImageryContext"                       
#> [228] "ISOImageryCoverageDescription"           
#> [229] "ISOImageryCoverageResult"                
#> [230] "ISOImageryEnvironmentalRecord"           
#> [231] "ISOImageryEvent"                         
#> [232] "ISOImageryGCP"                           
#> [233] "ISOImageryGCPCollection"                 
#> [234] "ISOImageryGeometryType"                  
#> [235] "ISOImageryGeorectified"                  
#> [236] "ISOImageryGeoreferenceable"              
#> [237] "ISOImageryImageDescription"              
#> [238] "ISOImageryInstrument"                    
#> [239] "ISOImageryMetadata"                      
#> [240] "ISOImageryNominalResolution"             
#> [241] "ISOImageryObjective"                     
#> [242] "ISOImageryObjectiveType"                 
#> [243] "ISOImageryOperation"                     
#> [244] "ISOImageryOperationType"                 
#> [245] "ISOImageryPlan"                          
#> [246] "ISOImageryPlatform"                      
#> [247] "ISOImageryPlatformPass"                  
#> [248] "ISOImageryPolarisationOrientation"       
#> [249] "ISOImageryPriority"                      
#> [250] "ISOImageryProcessStep"                   
#> [251] "ISOImageryProcessStepReport"             
#> [252] "ISOImageryProcessing"                    
#> [253] "ISOImageryRangeElementDescription"       
#> [254] "ISOImageryRequestedDate"                 
#> [255] "ISOImageryRequirement"                   
#> [256] "ISOImagerySensorType"                    
#> [257] "ISOImagerySequence"                      
#> [258] "ISOImagerySource"                        
#> [259] "ISOImageryTransferFunctionType"          
#> [260] "ISOImageryTrigger"                       
#> [261] "ISOImageryUsability"                     
#> [262] "ISOImagingCondition"                     
#> [263] "ISOIndividual"                           
#> [264] "ISOInheritanceRelation"                  
#> [265] "ISOInitiative"                           
#> [266] "ISOInitiativeType"                       
#> [267] "ISOInstrumentationEventType"             
#> [268] "ISOKeywordClass"                         
#> [269] "ISOKeywordType"                          
#> [270] "ISOKeywords"                             
#> [271] "ISOLanguage"                             
#> [272] "ISOLegalConstraints"                     
#> [273] "ISOLength"                               
#> [274] "ISOLineage"                              
#> [275] "ISOListedValue"                          
#> [276] "ISOLocalName"                            
#> [277] "ISOLocale"                               
#> [278] "ISOLocaleContainer"                      
#> [279] "ISOLocalisedCharacterString"             
#> [280] "ISOMDFeatureCatalogue"                   
#> [281] "ISOMLCodeDefinition"                     
#> [282] "ISOMLCodeListDictionary"                 
#> [283] "ISOMaintenanceFrequency"                 
#> [284] "ISOMaintenanceInformation"               
#> [285] "ISOMeasure"                              
#> [286] "ISOMedium"                               
#> [287] "ISOMediumFormat"                         
#> [288] "ISOMediumName"                           
#> [289] "ISOMemberName"                           
#> [290] "ISOMetaIdentifier"                       
#> [291] "ISOMetadata"                             
#> [292] "ISOMetadataExtensionInformation"         
#> [293] "ISOMetadataScope"                        
#> [294] "ISOMimeFileType"                         
#> [295] "ISOMultiplicity"                         
#> [296] "ISOMultiplicityRange"                    
#> [297] "ISONonQuantitativeAttributeAccuracy"     
#> [298] "ISOObligation"                           
#> [299] "ISOOnLineFunction"                       
#> [300] "ISOOnlineResource"                       
#> [301] "ISOOperationChainMetadata"               
#> [302] "ISOOperationMetadata"                    
#> [303] "ISOOrganisation"                         
#> [304] "ISOOtherAggregate"                       
#> [305] "ISOParameter"                            
#> [306] "ISOParameterDirection"                   
#> [307] "ISOPeriodDuration"                       
#> [308] "ISOPixelOrientation"                     
#> [309] "ISOPlatform"                             
#> [310] "ISOPortrayalCatalogueReference"          
#> [311] "ISOPresentationForm"                     
#> [312] "ISOProcessStep"                          
#> [313] "ISOProductionSeries"                     
#> [314] "ISOProgress"                             
#> [315] "ISOPropertyType"                         
#> [316] "ISOQuantitativeAttributeAccuracy"        
#> [317] "ISOQuantitativeResult"                   
#> [318] "ISORangeDimension"                       
#> [319] "ISORecord"                               
#> [320] "ISORecordType"                           
#> [321] "ISOReferenceIdentifier"                  
#> [322] "ISOReferenceSystem"                      
#> [323] "ISOReferenceSystemType"                  
#> [324] "ISORelativeInternalPositionalAccuracy"   
#> [325] "ISOReleasability"                        
#> [326] "ISORepresentativeFraction"               
#> [327] "ISOResolution"                           
#> [328] "ISOResponsibility"                       
#> [329] "ISOResponsibleParty"                     
#> [330] "ISORestriction"                          
#> [331] "ISORole"                                 
#> [332] "ISORoleType"                             
#> [333] "ISOSRVServiceIdentification"             
#> [334] "ISOSampleDimension"                      
#> [335] "ISOScale"                                
#> [336] "ISOScope"                                
#> [337] "ISOScopeCode"                            
#> [338] "ISOScopeDescription"                     
#> [339] "ISOScopedName"                           
#> [340] "ISOSecurityConstraints"                  
#> [341] "ISOSensor"                               
#> [342] "ISOSeries"                               
#> [343] "ISOServiceIdentification"                
#> [344] "ISOSource"                               
#> [345] "ISOSpatialRepresentation"                
#> [346] "ISOSpatialRepresentationType"            
#> [347] "ISOSpatialTemporalExtent"                
#> [348] "ISOStandardOrderProcess"                 
#> [349] "ISOStatus"                               
#> [350] "ISOStereoMate"                           
#> [351] "ISOTelephone"                            
#> [352] "ISOTelephoneType"                        
#> [353] "ISOTemporalConsistency"                  
#> [354] "ISOTemporalExtent"                       
#> [355] "ISOTemporalValidity"                     
#> [356] "ISOThematicClassificationCorrectness"    
#> [357] "ISOTopicCategory"                        
#> [358] "ISOTopologicalConsistency"               
#> [359] "ISOTopologyLevel"                        
#> [360] "ISOTypeName"                             
#> [361] "ISOURI"                                  
#> [362] "ISOURL"                                  
#> [363] "ISOUnlimitedInteger"                     
#> [364] "ISOUomIdentifier"                        
#> [365] "ISOUsage"                                
#> [366] "ISOVectorSpatialRepresentation"          
#> [367] "ISOVerticalExtent"                       
#> [368] "ISoAbstractFeatureCatalogue"             
#> [369] "SWEAbstractDataComponent"                
#> [370] "SWEAbstractEncoding"                     
#> [371] "SWEAbstractObject"                       
#> [372] "SWEAbstractSWE"                          
#> [373] "SWEAbstractSWEIdentifiable"              
#> [374] "SWEAbstractSimpleComponent"              
#> [375] "SWECategory"                             
#> [376] "SWECategoryRange"                        
#> [377] "SWECount"                                
#> [378] "SWECountRange"                           
#> [379] "SWEDataRecord"                           
#> [380] "SWEElement"                              
#> [381] "SWENilValues"                            
#> [382] "SWEQuantity"                             
#> [383] "SWEQuantityRange"                        
#> [384] "SWEText"                                 
#> [385] "SWETextEncoding"                         
#> [386] "SWETime"                                 
#> [387] "SWETimeRange"                            
#> [388] "SWEXMLEncoding"