getISOClasses

get the list of ISO classes, ie classes extending ISOAbstractObject super class, including classes eventually defined outside geometa. In case the latter is on the search path, the list of ISO classes will be cached for optimized used by geometa encoder/decoder.

getISOClasses()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOClasses()
#>   [1] "GMLAbstractCRS"                          
#>   [2] "GMLAbstractCoordinateOperation"          
#>   [3] "GMLAbstractCoordinateSystem"             
#>   [4] "GMLAbstractCoverage"                     
#>   [5] "GMLAbstractCurve"                        
#>   [6] "GMLAbstractDiscreteCoverage"             
#>   [7] "GMLAbstractFeature"                      
#>   [8] "GMLAbstractGML"                          
#>   [9] "GMLAbstractGeneralConversion"            
#>  [10] "GMLAbstractGeneralDerivedCRS"            
#>  [11] "GMLAbstractGeneralOperationParameter"    
#>  [12] "GMLAbstractGeneralParameterValue"        
#>  [13] "GMLAbstractGeometricAggregate"           
#>  [14] "GMLAbstractGeometricPrimitive"           
#>  [15] "GMLAbstractGeometry"                     
#>  [16] "GMLAbstractImplicitGeometry"             
#>  [17] "GMLAbstractObject"                       
#>  [18] "GMLAbstractReferenceableGrid"            
#>  [19] "GMLAbstractRing"                         
#>  [20] "GMLAbstractSingleCRS"                    
#>  [21] "GMLAbstractSingleOperation"              
#>  [22] "GMLAbstractSurface"                      
#>  [23] "GMLAbstractTimeGeometricPrimitive"       
#>  [24] "GMLAbstractTimeObject"                   
#>  [25] "GMLAbstractTimePrimitive"                
#>  [26] "GMLAffineCS"                             
#>  [27] "GMLBaseUnit"                             
#>  [28] "GMLCOVAbstractCoverage"                  
#>  [29] "GMLCOVExtension"                         
#>  [30] "GMLCartesianCS"                          
#>  [31] "GMLCodeType"                             
#>  [32] "GMLCompoundCRS"                          
#>  [33] "GMLConventionalUnit"                     
#>  [34] "GMLConversion"                           
#>  [35] "GMLCoordinateSystemAxis"                 
#>  [36] "GMLCylindricalCS"                        
#>  [37] "GMLDefinition"                           
#>  [38] "GMLDerivedCRS"                           
#>  [39] "GMLDerivedUnit"                          
#>  [40] "GMLElement"                              
#>  [41] "GMLEllipsoidalCS"                        
#>  [42] "GMLEnvelope"                             
#>  [43] "GMLEnvelopeWithTimePeriod"               
#>  [44] "GMLGeneralGridAxis"                      
#>  [45] "GMLGeodeticCRS"                          
#>  [46] "GMLGrid"                                 
#>  [47] "GMLGridCoverage"                         
#>  [48] "GMLGridEnvelope"                         
#>  [49] "GMLGridFunction"                         
#>  [50] "GMLLineString"                           
#>  [51] "GMLLinearCS"                             
#>  [52] "GMLLinearRing"                           
#>  [53] "GMLMultiCurve"                           
#>  [54] "GMLMultiCurveCoverage"                   
#>  [55] "GMLMultiPoint"                           
#>  [56] "GMLMultiPointCoverage"                   
#>  [57] "GMLMultiSolidCoverage"                   
#>  [58] "GMLMultiSurface"                         
#>  [59] "GMLMultiSurfaceCoverage"                 
#>  [60] "GMLObliqueCartesianCS"                   
#>  [61] "GMLOperationMethod"                      
#>  [62] "GMLOperationParameter"                   
#>  [63] "GMLOperationParameterGroup"              
#>  [64] "GMLParameterValue"                       
#>  [65] "GMLParameterValueGroup"                  
#>  [66] "GMLPoint"                                
#>  [67] "GMLPolarCS"                              
#>  [68] "GMLPolygon"                              
#>  [69] "GMLProjectedCRS"                         
#>  [70] "GMLRectifiedGrid"                        
#>  [71] "GMLRectifiedGridCoverage"                
#>  [72] "GMLReferenceableGridByArray"             
#>  [73] "GMLReferenceableGridByTransformation"    
#>  [74] "GMLReferenceableGridByVectors"           
#>  [75] "GMLSphericalCS"                          
#>  [76] "GMLTemporalCRS"                          
#>  [77] "GMLTemporalCS"                           
#>  [78] "GMLTimeCS"                               
#>  [79] "GMLTimeInstant"                          
#>  [80] "GMLTimePeriod"                           
#>  [81] "GMLUnitDefinition"                       
#>  [82] "GMLUserDefinedCS"                        
#>  [83] "GMLVerticalCRS"                          
#>  [84] "GMLVerticalCS"                           
#>  [85] "ISOAbsoluteExternalPositionalAccuracy"   
#>  [86] "ISOAbstractAcquisitionInformation"       
#>  [87] "ISOAbstractAggregate"                    
#>  [88] "ISOAbstractApplicationSchemaInformation" 
#>  [89] "ISOAbstractCarrierOfCharacteristics"     
#>  [90] "ISOAbstractCatalogue"                    
#>  [91] "ISOAbstractCitation"                     
#>  [92] "ISOAbstractCompleteness"                 
#>  [93] "ISOAbstractConstraints"                  
#>  [94] "ISOAbstractContentInformation"           
#>  [95] "ISOAbstractDataEvaluation"               
#>  [96] "ISOAbstractDataQuality"                  
#>  [97] "ISOAbstractDistribution"                 
#>  [98] "ISOAbstractExtent"                       
#>  [99] "ISOAbstractFeatureType"                  
#> [100] "ISOAbstractFormat"                       
#> [101] "ISOAbstractGenericName"                  
#> [102] "ISOAbstractLineageInformation"           
#> [103] "ISOAbstractLogicalConsistency"           
#> [104] "ISOAbstractMDContentInformation"         
#> [105] "ISOAbstractMaintenanceInformation"       
#> [106] "ISOAbstractMetadata"                     
#> [107] "ISOAbstractMetadataExtension"            
#> [108] "ISOAbstractMetaquality"                  
#> [109] "ISOAbstractOnlineResource"               
#> [110] "ISOAbstractParameter"                    
#> [111] "ISOAbstractParty"                        
#> [112] "ISOAbstractPlatform"                     
#> [113] "ISOAbstractPortrayalCatalogueInformation"
#> [114] "ISOAbstractPositionalAccuracy"           
#> [115] "ISOAbstractPropertyType"                 
#> [116] "ISOAbstractQualityElement"               
#> [117] "ISOAbstractRSReferenceSystem"            
#> [118] "ISOAbstractReferenceSystem"              
#> [119] "ISOAbstractResourceDescription"          
#> [120] "ISOAbstractResponsibility"               
#> [121] "ISOAbstractResult"                       
#> [122] "ISOAbstractSpatialRepresentation"        
#> [123] "ISOAbstractSpatialResolution"            
#> [124] "ISOAbstractStandardOrderProcess"         
#> [125] "ISOAbstractTemporalAccuracy"             
#> [126] "ISOAbstractTemporalQuality"              
#> [127] "ISOAbstractThematicAccuracy"             
#> [128] "ISOAbstractTypedDate"                    
#> [129] "ISOAccuracyOfATimeMeasurement"           
#> [130] "ISOAddress"                              
#> [131] "ISOAggregateInformation"                 
#> [132] "ISOAnchor"                               
#> [133] "ISOAngle"                                
#> [134] "ISOApplicationSchemaInformation"         
#> [135] "ISOAssociatedResource"                   
#> [136] "ISOAssociation"                          
#> [137] "ISOAssociationRole"                      
#> [138] "ISOAssociationType"                      
#> [139] "ISOAttributeGroup"                       
#> [140] "ISOBand"                                 
#> [141] "ISOBaseBoolean"                          
#> [142] "ISOBaseCharacterString"                  
#> [143] "ISOBaseDate"                             
#> [144] "ISOBaseDateTime"                         
#> [145] "ISOBaseDecimal"                          
#> [146] "ISOBaseInteger"                          
#> [147] "ISOBaseReal"                             
#> [148] "ISOBinary"                               
#> [149] "ISOBinding"                              
#> [150] "ISOBoundAssociationRole"                 
#> [151] "ISOBoundFeatureAttribute"                
#> [152] "ISOBoundingPolygon"                      
#> [153] "ISOBrowseGraphic"                        
#> [154] "ISOCarrierOfCharacteristics"             
#> [155] "ISOCellGeometry"                         
#> [156] "ISOCharacterSet"                         
#> [157] "ISOCitation"                             
#> [158] "ISOCitationSeries"                       
#> [159] "ISOClassification"                       
#> [160] "ISOCodeDefinition"                       
#> [161] "ISOCodeListDictionary"                   
#> [162] "ISOCodeListValue"                        
#> [163] "ISOCodelist"                             
#> [164] "ISOCodelistCatalogue"                    
#> [165] "ISOCodelistValue"                        
#> [166] "ISOCompletenessCommission"               
#> [167] "ISOCompletenessOmission"                 
#> [168] "ISOConceptualConsistency"                
#> [169] "ISOConfidence"                           
#> [170] "ISOConformanceResult"                    
#> [171] "ISOConstraint"                           
#> [172] "ISOConstraints"                          
#> [173] "ISOContact"                              
#> [174] "ISOCountry"                              
#> [175] "ISOCoupledResource"                      
#> [176] "ISOCouplingType"                         
#> [177] "ISOCoverageContentType"                  
#> [178] "ISOCoverageDescription"                  
#> [179] "ISODCPList"                              
#> [180] "ISODataFile"                             
#> [181] "ISODataIdentification"                   
#> [182] "ISODataIdentification19115_3"            
#> [183] "ISODataIdentification19139"              
#> [184] "ISODataQuality"                          
#> [185] "ISODataQualityAbstractElement"           
#> [186] "ISODataQualityScope"                     
#> [187] "ISODataSet"                              
#> [188] "ISODatatype"                             
#> [189] "ISODate"                                 
#> [190] "ISODateType"                             
#> [191] "ISODefinitionReference"                  
#> [192] "ISODefinitionSource"                     
#> [193] "ISODigitalTransferOptions"               
#> [194] "ISODimension"                            
#> [195] "ISODimensionNameType"                    
#> [196] "ISODistance"                             
#> [197] "ISODistribution"                         
#> [198] "ISODistributionUnits"                    
#> [199] "ISODistributor"                          
#> [200] "ISODomainConsistency"                    
#> [201] "ISOElementSequence"                      
#> [202] "ISOEvaluationMethodType"                 
#> [203] "ISOExtendedElementInformation"           
#> [204] "ISOExtent"                               
#> [205] "ISOFeatureAssociation"                   
#> [206] "ISOFeatureAttribute"                     
#> [207] "ISOFeatureCatalogue"                     
#> [208] "ISOFeatureCatalogueDescription"          
#> [209] "ISOFeatureOperation"                     
#> [210] "ISOFeatureType"                          
#> [211] "ISOFeatureTypeInfo"                      
#> [212] "ISOFileName"                             
#> [213] "ISOFormat"                               
#> [214] "ISOFormatConsistency"                    
#> [215] "ISOFreeText"                             
#> [216] "ISOGeographicBoundingBox"                
#> [217] "ISOGeographicDescription"                
#> [218] "ISOGeographicExtent"                     
#> [219] "ISOGeometricObjectType"                  
#> [220] "ISOGeometricObjects"                     
#> [221] "ISOGeorectified"                         
#> [222] "ISOGeoreferenceable"                     
#> [223] "ISOGridSpatialRepresentation"            
#> [224] "ISOGriddedDataPositionalAccuracy"        
#> [225] "ISOHomogeneity"                          
#> [226] "ISOIdentification"                       
#> [227] "ISOIdentification19115_3"                
#> [228] "ISOIdentification19139"                  
#> [229] "ISOImageDescription"                     
#> [230] "ISOImageryAbstractGeolocationInformation"
#> [231] "ISOImageryAcquisitionInformation"        
#> [232] "ISOImageryAlgorithm"                     
#> [233] "ISOImageryBand"                          
#> [234] "ISOImageryBandDefinition"                
#> [235] "ISOImageryContext"                       
#> [236] "ISOImageryCoverageDescription"           
#> [237] "ISOImageryCoverageResult"                
#> [238] "ISOImageryEnvironmentalRecord"           
#> [239] "ISOImageryEvent"                         
#> [240] "ISOImageryGCP"                           
#> [241] "ISOImageryGCPCollection"                 
#> [242] "ISOImageryGeometryType"                  
#> [243] "ISOImageryGeorectified"                  
#> [244] "ISOImageryGeoreferenceable"              
#> [245] "ISOImageryImageDescription"              
#> [246] "ISOImageryInstrument"                    
#> [247] "ISOImageryMetadata"                      
#> [248] "ISOImageryNominalResolution"             
#> [249] "ISOImageryObjective"                     
#> [250] "ISOImageryObjectiveType"                 
#> [251] "ISOImageryOperation"                     
#> [252] "ISOImageryOperationType"                 
#> [253] "ISOImageryPlan"                          
#> [254] "ISOImageryPlatform"                      
#> [255] "ISOImageryPlatformPass"                  
#> [256] "ISOImageryPolarisationOrientation"       
#> [257] "ISOImageryPriority"                      
#> [258] "ISOImageryProcessStep"                   
#> [259] "ISOImageryProcessStepReport"             
#> [260] "ISOImageryProcessing"                    
#> [261] "ISOImageryRangeElementDescription"       
#> [262] "ISOImageryRequestedDate"                 
#> [263] "ISOImageryRequirement"                   
#> [264] "ISOImageryRevision"                      
#> [265] "ISOImagerySensor"                        
#> [266] "ISOImagerySensorType"                    
#> [267] "ISOImagerySequence"                      
#> [268] "ISOImagerySource"                        
#> [269] "ISOImageryTransferFunctionType"          
#> [270] "ISOImageryTrigger"                       
#> [271] "ISOImageryUsability"                     
#> [272] "ISOImagingCondition"                     
#> [273] "ISOIndividual"                           
#> [274] "ISOInheritanceRelation"                  
#> [275] "ISOInitiative"                           
#> [276] "ISOInitiativeType"                       
#> [277] "ISOInstrumentationEvent"                 
#> [278] "ISOInstrumentationEventList"             
#> [279] "ISOInstrumentationEventType"             
#> [280] "ISOKeywordClass"                         
#> [281] "ISOKeywordType"                          
#> [282] "ISOKeywords"                             
#> [283] "ISOLanguage"                             
#> [284] "ISOLegalConstraints"                     
#> [285] "ISOLength"                               
#> [286] "ISOLineage"                              
#> [287] "ISOListedValue"                          
#> [288] "ISOLocalName"                            
#> [289] "ISOLocale"                               
#> [290] "ISOLocaleContainer"                      
#> [291] "ISOLocalisedCharacterString"             
#> [292] "ISOMDFeatureCatalogue"                   
#> [293] "ISOMLCodeDefinition"                     
#> [294] "ISOMLCodeListDictionary"                 
#> [295] "ISOMaintenanceFrequency"                 
#> [296] "ISOMaintenanceInformation"               
#> [297] "ISOMeasure"                              
#> [298] "ISOMedium"                               
#> [299] "ISOMediumFormat"                         
#> [300] "ISOMediumName"                           
#> [301] "ISOMemberName"                           
#> [302] "ISOMetaIdentifier"                       
#> [303] "ISOMetadata"                             
#> [304] "ISOMetadataExtensionInformation"         
#> [305] "ISOMetadataScope"                        
#> [306] "ISOMimeFileType"                         
#> [307] "ISOMultiplicity"                         
#> [308] "ISOMultiplicityRange"                    
#> [309] "ISONonQuantitativeAttributeAccuracy"     
#> [310] "ISOObligation"                           
#> [311] "ISOOnLineFunction"                       
#> [312] "ISOOnlineResource"                       
#> [313] "ISOOperationChainMetadata"               
#> [314] "ISOOperationMetadata"                    
#> [315] "ISOOrganisation"                         
#> [316] "ISOOtherAggregate"                       
#> [317] "ISOParameterDirection"                   
#> [318] "ISOPeriodDuration"                       
#> [319] "ISOPixelOrientation"                     
#> [320] "ISOPlatform"                             
#> [321] "ISOPortrayalCatalogueReference"          
#> [322] "ISOPresentationForm"                     
#> [323] "ISOProcessParameter"                     
#> [324] "ISOProcessStep"                          
#> [325] "ISOProductionSeries"                     
#> [326] "ISOProgress"                             
#> [327] "ISOPropertyType"                         
#> [328] "ISOQuantitativeAttributeAccuracy"        
#> [329] "ISOQuantitativeResult"                   
#> [330] "ISORangeDimension"                       
#> [331] "ISORecord"                               
#> [332] "ISORecordType"                           
#> [333] "ISOReferenceIdentifier"                  
#> [334] "ISOReferenceSystem"                      
#> [335] "ISOReferenceSystemType"                  
#> [336] "ISORelativeInternalPositionalAccuracy"   
#> [337] "ISOReleasability"                        
#> [338] "ISORepresentativeFraction"               
#> [339] "ISORepresentativity"                     
#> [340] "ISOResolution"                           
#> [341] "ISOResponsibility"                       
#> [342] "ISOResponsibleParty"                     
#> [343] "ISORestriction"                          
#> [344] "ISORole"                                 
#> [345] "ISORoleType"                             
#> [346] "ISOSRVParameter"                         
#> [347] "ISOSRVParameterDirection"                
#> [348] "ISOSRVServiceIdentification"             
#> [349] "ISOSampleDimension"                      
#> [350] "ISOScale"                                
#> [351] "ISOScope"                                
#> [352] "ISOScopeCode"                            
#> [353] "ISOScopeDescription"                     
#> [354] "ISOScopedName"                           
#> [355] "ISOSecurityConstraints"                  
#> [356] "ISOSensor"                               
#> [357] "ISOSeries"                               
#> [358] "ISOServiceIdentification"                
#> [359] "ISOSource"                               
#> [360] "ISOSpatialRepresentation"                
#> [361] "ISOSpatialRepresentationType"            
#> [362] "ISOSpatialTemporalExtent"                
#> [363] "ISOStandardOrderProcess"                 
#> [364] "ISOStatus"                               
#> [365] "ISOStereoMate"                           
#> [366] "ISOTelephone"                            
#> [367] "ISOTelephoneType"                        
#> [368] "ISOTemporalConsistency"                  
#> [369] "ISOTemporalExtent"                       
#> [370] "ISOTemporalValidity"                     
#> [371] "ISOThematicClassificationCorrectness"    
#> [372] "ISOTopicCategory"                        
#> [373] "ISOTopologicalConsistency"               
#> [374] "ISOTopologyLevel"                        
#> [375] "ISOTypeName"                             
#> [376] "ISOURI"                                  
#> [377] "ISOURL"                                  
#> [378] "ISOUnlimitedInteger"                     
#> [379] "ISOUomIdentifier"                        
#> [380] "ISOUsage"                                
#> [381] "ISOVectorSpatialRepresentation"          
#> [382] "ISOVerticalExtent"                       
#> [383] "ISoAbstractFeatureCatalogue"             
#> [384] "SWEAbstractDataComponent"                
#> [385] "SWEAbstractEncoding"                     
#> [386] "SWEAbstractObject"                       
#> [387] "SWEAbstractSWE"                          
#> [388] "SWEAbstractSWEIdentifiable"              
#> [389] "SWEAbstractSimpleComponent"              
#> [390] "SWECategory"                             
#> [391] "SWECategoryRange"                        
#> [392] "SWECount"                                
#> [393] "SWECountRange"                           
#> [394] "SWEDataRecord"                           
#> [395] "SWEElement"                              
#> [396] "SWENilValues"                            
#> [397] "SWEQuantity"                             
#> [398] "SWEQuantityRange"                        
#> [399] "SWEText"                                 
#> [400] "SWETextEncoding"                         
#> [401] "SWETime"                                 
#> [402] "SWETimeRange"                            
#> [403] "SWEXMLEncoding"