getISOClasses

get the list of ISO classes, ie classes extending ISOAbstractObject super class, including classes eventually defined outside geometa. In case the latter is on the search path, the list of ISO classes will be cached for optimized used by geometa encoder/decoder.

getISOClasses()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOClasses()
#>   [1] "GMLAbstractCRS"                          
#>   [2] "GMLAbstractCoordinateOperation"          
#>   [3] "GMLAbstractCoordinateSystem"             
#>   [4] "GMLAbstractCoverage"                     
#>   [5] "GMLAbstractCurve"                        
#>   [6] "GMLAbstractDiscreteCoverage"             
#>   [7] "GMLAbstractFeature"                      
#>   [8] "GMLAbstractGML"                          
#>   [9] "GMLAbstractGeneralConversion"            
#>  [10] "GMLAbstractGeneralDerivedCRS"            
#>  [11] "GMLAbstractGeneralOperationParameter"    
#>  [12] "GMLAbstractGeneralParameterValue"        
#>  [13] "GMLAbstractGeometricAggregate"           
#>  [14] "GMLAbstractGeometricPrimitive"           
#>  [15] "GMLAbstractGeometry"                     
#>  [16] "GMLAbstractImplicitGeometry"             
#>  [17] "GMLAbstractObject"                       
#>  [18] "GMLAbstractReferenceableGrid"            
#>  [19] "GMLAbstractRing"                         
#>  [20] "GMLAbstractSingleCRS"                    
#>  [21] "GMLAbstractSingleOperation"              
#>  [22] "GMLAbstractSurface"                      
#>  [23] "GMLAbstractTimeGeometricPrimitive"       
#>  [24] "GMLAbstractTimeObject"                   
#>  [25] "GMLAbstractTimePrimitive"                
#>  [26] "GMLAffineCS"                             
#>  [27] "GMLBaseUnit"                             
#>  [28] "GMLCOVAbstractCoverage"                  
#>  [29] "GMLCOVExtension"                         
#>  [30] "GMLCartesianCS"                          
#>  [31] "GMLCodeType"                             
#>  [32] "GMLCompoundCRS"                          
#>  [33] "GMLConventionalUnit"                     
#>  [34] "GMLConversion"                           
#>  [35] "GMLCoordinateSystemAxis"                 
#>  [36] "GMLCylindricalCS"                        
#>  [37] "GMLDefinition"                           
#>  [38] "GMLDerivedCRS"                           
#>  [39] "GMLDerivedUnit"                          
#>  [40] "GMLElement"                              
#>  [41] "GMLEllipsoidalCS"                        
#>  [42] "GMLEnvelope"                             
#>  [43] "GMLEnvelopeWithTimePeriod"               
#>  [44] "GMLGeneralGridAxis"                      
#>  [45] "GMLGeodeticCRS"                          
#>  [46] "GMLGrid"                                 
#>  [47] "GMLGridCoverage"                         
#>  [48] "GMLGridEnvelope"                         
#>  [49] "GMLGridFunction"                         
#>  [50] "GMLLineString"                           
#>  [51] "GMLLinearCS"                             
#>  [52] "GMLLinearRing"                           
#>  [53] "GMLMultiCurve"                           
#>  [54] "GMLMultiCurveCoverage"                   
#>  [55] "GMLMultiPoint"                           
#>  [56] "GMLMultiPointCoverage"                   
#>  [57] "GMLMultiSolidCoverage"                   
#>  [58] "GMLMultiSurface"                         
#>  [59] "GMLMultiSurfaceCoverage"                 
#>  [60] "GMLObliqueCartesianCS"                   
#>  [61] "GMLOperationMethod"                      
#>  [62] "GMLOperationParameter"                   
#>  [63] "GMLOperationParameterGroup"              
#>  [64] "GMLParameterValue"                       
#>  [65] "GMLParameterValueGroup"                  
#>  [66] "GMLPoint"                                
#>  [67] "GMLPolarCS"                              
#>  [68] "GMLPolygon"                              
#>  [69] "GMLProjectedCRS"                         
#>  [70] "GMLRectifiedGrid"                        
#>  [71] "GMLRectifiedGridCoverage"                
#>  [72] "GMLReferenceableGridByArray"             
#>  [73] "GMLReferenceableGridByTransformation"    
#>  [74] "GMLReferenceableGridByVectors"           
#>  [75] "GMLSphericalCS"                          
#>  [76] "GMLTemporalCRS"                          
#>  [77] "GMLTemporalCS"                           
#>  [78] "GMLTimeCS"                               
#>  [79] "GMLTimeInstant"                          
#>  [80] "GMLTimePeriod"                           
#>  [81] "GMLUnitDefinition"                       
#>  [82] "GMLUserDefinedCS"                        
#>  [83] "GMLVerticalCRS"                          
#>  [84] "GMLVerticalCS"                           
#>  [85] "ISOAbsoluteExternalPositionalAccuracy"   
#>  [86] "ISOAbstractAcquisitionInformation"       
#>  [87] "ISOAbstractAggregate"                    
#>  [88] "ISOAbstractApplicationSchemaInformation" 
#>  [89] "ISOAbstractCarrierOfCharacteristics"     
#>  [90] "ISOAbstractCatalogue"                    
#>  [91] "ISOAbstractCitation"                     
#>  [92] "ISOAbstractCompleteness"                 
#>  [93] "ISOAbstractConstraints"                  
#>  [94] "ISOAbstractContentInformation"           
#>  [95] "ISOAbstractDataEvaluation"               
#>  [96] "ISOAbstractDataQuality"                  
#>  [97] "ISOAbstractDistribution"                 
#>  [98] "ISOAbstractExtent"                       
#>  [99] "ISOAbstractFeatureCatalogue"             
#> [100] "ISOAbstractFeatureType"                  
#> [101] "ISOAbstractFormat"                       
#> [102] "ISOAbstractGenericName"                  
#> [103] "ISOAbstractLineageInformation"           
#> [104] "ISOAbstractLogicalConsistency"           
#> [105] "ISOAbstractMDContentInformation"         
#> [106] "ISOAbstractMaintenanceInformation"       
#> [107] "ISOAbstractMetadata"                     
#> [108] "ISOAbstractMetadataExtension"            
#> [109] "ISOAbstractMetaquality"                  
#> [110] "ISOAbstractOnlineResource"               
#> [111] "ISOAbstractParameter"                    
#> [112] "ISOAbstractParty"                        
#> [113] "ISOAbstractPlatform"                     
#> [114] "ISOAbstractPortrayalCatalogueInformation"
#> [115] "ISOAbstractPositionalAccuracy"           
#> [116] "ISOAbstractPropertyType"                 
#> [117] "ISOAbstractQualityElement"               
#> [118] "ISOAbstractRSReferenceSystem"            
#> [119] "ISOAbstractReferenceSystem"              
#> [120] "ISOAbstractResourceDescription"          
#> [121] "ISOAbstractResponsibility"               
#> [122] "ISOAbstractResult"                       
#> [123] "ISOAbstractSpatialRepresentation"        
#> [124] "ISOAbstractSpatialResolution"            
#> [125] "ISOAbstractStandardOrderProcess"         
#> [126] "ISOAbstractTemporalAccuracy"             
#> [127] "ISOAbstractTemporalQuality"              
#> [128] "ISOAbstractThematicAccuracy"             
#> [129] "ISOAbstractTypedDate"                    
#> [130] "ISOAccuracyOfATimeMeasurement"           
#> [131] "ISOAddress"                              
#> [132] "ISOAggregateInformation"                 
#> [133] "ISOAggregationDerivation"                
#> [134] "ISOAnchor"                               
#> [135] "ISOAngle"                                
#> [136] "ISOApplicationSchemaInformation"         
#> [137] "ISOAssociatedResource"                   
#> [138] "ISOAssociation"                          
#> [139] "ISOAssociationRole"                      
#> [140] "ISOAssociationType"                      
#> [141] "ISOAttributeGroup"                       
#> [142] "ISOBand"                                 
#> [143] "ISOBaseBoolean"                          
#> [144] "ISOBaseCharacterString"                  
#> [145] "ISOBaseDate"                             
#> [146] "ISOBaseDateTime"                         
#> [147] "ISOBaseDecimal"                          
#> [148] "ISOBaseInteger"                          
#> [149] "ISOBaseReal"                             
#> [150] "ISOBinary"                               
#> [151] "ISOBinding"                              
#> [152] "ISOBoundAssociationRole"                 
#> [153] "ISOBoundFeatureAttribute"                
#> [154] "ISOBoundingPolygon"                      
#> [155] "ISOBrowseGraphic"                        
#> [156] "ISOCTCodelistValue"                      
#> [157] "ISOCarrierOfCharacteristics"             
#> [158] "ISOCellGeometry"                         
#> [159] "ISOCharacterSet"                         
#> [160] "ISOCitation"                             
#> [161] "ISOCitationSeries"                       
#> [162] "ISOClassification"                       
#> [163] "ISOCodeDefinition"                       
#> [164] "ISOCodeListDictionary"                   
#> [165] "ISOCodeListValue"                        
#> [166] "ISOCodelist"                             
#> [167] "ISOCodelistCatalogue"                    
#> [168] "ISOCompletenessCommission"               
#> [169] "ISOCompletenessOmission"                 
#> [170] "ISOConceptualConsistency"                
#> [171] "ISOConfidence"                           
#> [172] "ISOConformanceResult"                    
#> [173] "ISOConstraint"                           
#> [174] "ISOConstraints"                          
#> [175] "ISOContact"                              
#> [176] "ISOCountry"                              
#> [177] "ISOCoupledResource"                      
#> [178] "ISOCouplingType"                         
#> [179] "ISOCoverageContentType"                  
#> [180] "ISOCoverageDescription"                  
#> [181] "ISODCPList"                              
#> [182] "ISODataFile"                             
#> [183] "ISODataIdentification"                   
#> [184] "ISODataIdentification19115_3"            
#> [185] "ISODataIdentification19139"              
#> [186] "ISODataInspection"                       
#> [187] "ISODataQuality"                          
#> [188] "ISODataQualityAbstractElement"           
#> [189] "ISODataQualityScope"                     
#> [190] "ISODataSet"                              
#> [191] "ISODatatype"                             
#> [192] "ISODate"                                 
#> [193] "ISODateType"                             
#> [194] "ISODefinitionReference"                  
#> [195] "ISODefinitionSource"                     
#> [196] "ISODescriptiveResult"                    
#> [197] "ISODigitalTransferOptions"               
#> [198] "ISODimension"                            
#> [199] "ISODimensionNameType"                    
#> [200] "ISODistance"                             
#> [201] "ISODistribution"                         
#> [202] "ISODistributionUnits"                    
#> [203] "ISODistributor"                          
#> [204] "ISODomainConsistency"                    
#> [205] "ISOElementSequence"                      
#> [206] "ISOEvaluationMethod"                     
#> [207] "ISOEvaluationMethodType"                 
#> [208] "ISOExtendedElementInformation"           
#> [209] "ISOExtent"                               
#> [210] "ISOFeatureAssociation"                   
#> [211] "ISOFeatureAssociation19115_3"            
#> [212] "ISOFeatureAssociation19139"              
#> [213] "ISOFeatureAttribute"                     
#> [214] "ISOFeatureCatalogue"                     
#> [215] "ISOFeatureCatalogueDescription"          
#> [216] "ISOFeatureOperation"                     
#> [217] "ISOFeatureType"                          
#> [218] "ISOFeatureType19115_3"                   
#> [219] "ISOFeatureType19139"                     
#> [220] "ISOFeatureTypeInfo"                      
#> [221] "ISOFileName"                             
#> [222] "ISOFormat"                               
#> [223] "ISOFormatConsistency"                    
#> [224] "ISOFreeText"                             
#> [225] "ISOFullInspection"                       
#> [226] "ISOGeographicBoundingBox"                
#> [227] "ISOGeographicDescription"                
#> [228] "ISOGeographicExtent"                     
#> [229] "ISOGeometricObjectType"                  
#> [230] "ISOGeometricObjects"                     
#> [231] "ISOGeorectified"                         
#> [232] "ISOGeoreferenceable"                     
#> [233] "ISOGridSpatialRepresentation"            
#> [234] "ISOGriddedDataPositionalAccuracy"        
#> [235] "ISOHomogeneity"                          
#> [236] "ISOIdentification"                       
#> [237] "ISOIdentification19115_3"                
#> [238] "ISOIdentification19139"                  
#> [239] "ISOImageDescription"                     
#> [240] "ISOImageryAbstractGeolocationInformation"
#> [241] "ISOImageryAcquisitionInformation"        
#> [242] "ISOImageryAlgorithm"                     
#> [243] "ISOImageryBand"                          
#> [244] "ISOImageryBandDefinition"                
#> [245] "ISOImageryContext"                       
#> [246] "ISOImageryCoverageDescription"           
#> [247] "ISOImageryCoverageResult"                
#> [248] "ISOImageryEnvironmentalRecord"           
#> [249] "ISOImageryEvent"                         
#> [250] "ISOImageryGCP"                           
#> [251] "ISOImageryGCPCollection"                 
#> [252] "ISOImageryGeometryType"                  
#> [253] "ISOImageryGeorectified"                  
#> [254] "ISOImageryGeoreferenceable"              
#> [255] "ISOImageryImageDescription"              
#> [256] "ISOImageryInstrument"                    
#> [257] "ISOImageryMetadata"                      
#> [258] "ISOImageryNominalResolution"             
#> [259] "ISOImageryObjective"                     
#> [260] "ISOImageryObjectiveType"                 
#> [261] "ISOImageryOperation"                     
#> [262] "ISOImageryOperationType"                 
#> [263] "ISOImageryPlan"                          
#> [264] "ISOImageryPlatform"                      
#> [265] "ISOImageryPlatformPass"                  
#> [266] "ISOImageryPolarisationOrientation"       
#> [267] "ISOImageryPriority"                      
#> [268] "ISOImageryProcessStep"                   
#> [269] "ISOImageryProcessStepReport"             
#> [270] "ISOImageryProcessing"                    
#> [271] "ISOImageryRangeElementDescription"       
#> [272] "ISOImageryRequestedDate"                 
#> [273] "ISOImageryRequirement"                   
#> [274] "ISOImageryRevision"                      
#> [275] "ISOImagerySensor"                        
#> [276] "ISOImagerySensorType"                    
#> [277] "ISOImagerySequence"                      
#> [278] "ISOImagerySource"                        
#> [279] "ISOImageryTransferFunctionType"          
#> [280] "ISOImageryTrigger"                       
#> [281] "ISOImageryUsability"                     
#> [282] "ISOImagingCondition"                     
#> [283] "ISOIndirectEvaluation"                   
#> [284] "ISOIndividual"                           
#> [285] "ISOInheritanceRelation"                  
#> [286] "ISOInitiative"                           
#> [287] "ISOInitiativeType"                       
#> [288] "ISOInstrumentationEvent"                 
#> [289] "ISOInstrumentationEventList"             
#> [290] "ISOInstrumentationEventType"             
#> [291] "ISOKeywordClass"                         
#> [292] "ISOKeywordType"                          
#> [293] "ISOKeywords"                             
#> [294] "ISOLanguage"                             
#> [295] "ISOLegalConstraints"                     
#> [296] "ISOLength"                               
#> [297] "ISOLineage"                              
#> [298] "ISOListedValue"                          
#> [299] "ISOLocalName"                            
#> [300] "ISOLocale"                               
#> [301] "ISOLocaleContainer"                      
#> [302] "ISOLocalisedCharacterString"             
#> [303] "ISOMDFeatureCatalogue"                   
#> [304] "ISOMLCodeDefinition"                     
#> [305] "ISOMLCodeListDictionary"                 
#> [306] "ISOMaintenanceFrequency"                 
#> [307] "ISOMaintenanceInformation"               
#> [308] "ISOMeasure"                              
#> [309] "ISOMeasureReference"                     
#> [310] "ISOMedium"                               
#> [311] "ISOMediumFormat"                         
#> [312] "ISOMediumName"                           
#> [313] "ISOMemberName"                           
#> [314] "ISOMetaIdentifier"                       
#> [315] "ISOMetadata"                             
#> [316] "ISOMetadataExtensionInformation"         
#> [317] "ISOMetadataScope"                        
#> [318] "ISOMimeFileType"                         
#> [319] "ISOMultiplicity"                         
#> [320] "ISOMultiplicityRange"                    
#> [321] "ISONonQuantitativeAttributeAccuracy"     
#> [322] "ISONonQuantitativeAttributeCorrectness"  
#> [323] "ISOObligation"                           
#> [324] "ISOOnLineFunction"                       
#> [325] "ISOOnlineResource"                       
#> [326] "ISOOperationChainMetadata"               
#> [327] "ISOOperationMetadata"                    
#> [328] "ISOOrganisation"                         
#> [329] "ISOOtherAggregate"                       
#> [330] "ISOParameterDirection"                   
#> [331] "ISOPeriodDuration"                       
#> [332] "ISOPixelOrientation"                     
#> [333] "ISOPlatform"                             
#> [334] "ISOPortrayalCatalogueReference"          
#> [335] "ISOPresentationForm"                     
#> [336] "ISOProcessParameter"                     
#> [337] "ISOProcessStep"                          
#> [338] "ISOProductionSeries"                     
#> [339] "ISOProgress"                             
#> [340] "ISOPropertyType"                         
#> [341] "ISOQualityResultFile"                    
#> [342] "ISOQuantitativeAttributeAccuracy"        
#> [343] "ISOQuantitativeResult"                   
#> [344] "ISORangeDimension"                       
#> [345] "ISORecord"                               
#> [346] "ISORecordType"                           
#> [347] "ISOReferenceIdentifier"                  
#> [348] "ISOReferenceSystem"                      
#> [349] "ISOReferenceSystemType"                  
#> [350] "ISORelativeInternalPositionalAccuracy"   
#> [351] "ISOReleasability"                        
#> [352] "ISORepresentativeFraction"               
#> [353] "ISORepresentativity"                     
#> [354] "ISOResolution"                           
#> [355] "ISOResponsibility"                       
#> [356] "ISOResponsibleParty"                     
#> [357] "ISORestriction"                          
#> [358] "ISORole"                                 
#> [359] "ISORoleType"                             
#> [360] "ISOSRVParameter"                         
#> [361] "ISOSRVParameterDirection"                
#> [362] "ISOSRVServiceIdentification"             
#> [363] "ISOSRVServiceIdentification19115_3"      
#> [364] "ISOSRVServiceIdentification19139"        
#> [365] "ISOSampleBasedInspection"                
#> [366] "ISOSampleDimension"                      
#> [367] "ISOScale"                                
#> [368] "ISOScope"                                
#> [369] "ISOScopeCode"                            
#> [370] "ISOScopeDescription"                     
#> [371] "ISOScopedName"                           
#> [372] "ISOSecurityConstraints"                  
#> [373] "ISOSensor"                               
#> [374] "ISOSeries"                               
#> [375] "ISOServiceIdentification"                
#> [376] "ISOServiceIdentification19115_3"         
#> [377] "ISOServiceIdentification19139"           
#> [378] "ISOSource"                               
#> [379] "ISOSpatialRepresentation"                
#> [380] "ISOSpatialRepresentationType"            
#> [381] "ISOSpatialTemporalExtent"                
#> [382] "ISOStandaloneQualityReportInformation"   
#> [383] "ISOStandardOrderProcess"                 
#> [384] "ISOStatus"                               
#> [385] "ISOStereoMate"                           
#> [386] "ISOTelephone"                            
#> [387] "ISOTelephoneType"                        
#> [388] "ISOTemporalConsistency"                  
#> [389] "ISOTemporalExtent"                       
#> [390] "ISOTemporalValidity"                     
#> [391] "ISOThematicClassificationCorrectness"    
#> [392] "ISOTopicCategory"                        
#> [393] "ISOTopologicalConsistency"               
#> [394] "ISOTopologyLevel"                        
#> [395] "ISOTypeName"                             
#> [396] "ISOURI"                                  
#> [397] "ISOURL"                                  
#> [398] "ISOUnlimitedInteger"                     
#> [399] "ISOUomIdentifier"                        
#> [400] "ISOUsabilityElement"                     
#> [401] "ISOUsage"                                
#> [402] "ISOVectorSpatialRepresentation"          
#> [403] "ISOVerticalExtent"                       
#> [404] "SWEAbstractDataComponent"                
#> [405] "SWEAbstractEncoding"                     
#> [406] "SWEAbstractObject"                       
#> [407] "SWEAbstractSWE"                          
#> [408] "SWEAbstractSWEIdentifiable"              
#> [409] "SWEAbstractSimpleComponent"              
#> [410] "SWECategory"                             
#> [411] "SWECategoryRange"                        
#> [412] "SWECount"                                
#> [413] "SWECountRange"                           
#> [414] "SWEDataRecord"                           
#> [415] "SWEElement"                              
#> [416] "SWENilValues"                            
#> [417] "SWEQuantity"                             
#> [418] "SWEQuantityRange"                        
#> [419] "SWEText"                                 
#> [420] "SWETextEncoding"                         
#> [421] "SWETime"                                 
#> [422] "SWETimeRange"                            
#> [423] "SWEXMLEncoding"