getISOCodelists

getISOCodelists allows to get the list of ISO codelists registered in geometa, their description and XML definition. The object returned is of class "data.frame"

getISOCodelists(version)

Arguments

version

the metadata standard version. If NULL (default), all codelists will be listed

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOCodelists()
#>                            classname  codeSpace
#> 1                 ISOAssociationType       <NA>
#> 2                    ISOCellGeometry       <NA>
#> 3                    ISOCharacterSet       <NA>
#> 4                  ISOClassification       <NA>
#> 5                         ISOCountry ISO 3166-2
#> 6                    ISOCouplingType       <NA>
#> 7             ISOCoverageContentType       <NA>
#> 8                         ISODCPList       <NA>
#> 9                        ISODatatype       <NA>
#> 10                       ISODateType       <NA>
#> 11              ISODimensionNameType       <NA>
#> 12              ISODistributionUnits       <NA>
#> 13           ISOEvaluationMethodType       <NA>
#> 14            ISOGeometricObjectType       <NA>
#> 15          ISOImageryBandDefinition       <NA>
#> 16                 ISOImageryContext       <NA>
#> 17            ISOImageryGeometryType       <NA>
#> 18           ISOImageryObjectiveType       <NA>
#> 19           ISOImageryOperationType       <NA>
#> 20 ISOImageryPolarisationOrientation       <NA>
#> 21                ISOImageryPriority       <NA>
#> 22              ISOImagerySensorType       <NA>
#> 23                ISOImagerySequence       <NA>
#> 24    ISOImageryTransferFunctionType       <NA>
#> 25                 ISOImageryTrigger       <NA>
#> 26               ISOImagingCondition       <NA>
#> 27                 ISOInitiativeType       <NA>
#> 28       ISOInstrumentationEventType       <NA>
#> 29                    ISOKeywordType       <NA>
#> 30                       ISOLanguage  ISO 639-2
#> 31           ISOMaintenanceFrequency       <NA>
#> 32                   ISOMediumFormat       <NA>
#> 33                     ISOMediumName       <NA>
#> 34                     ISOObligation       <NA>
#> 35                 ISOOnLineFunction       <NA>
#> 36             ISOParameterDirection       <NA>
#> 37               ISOPixelOrientation       <NA>
#> 38               ISOPresentationForm       <NA>
#> 39                       ISOProgress       <NA>
#> 40            ISOReferenceSystemType       <NA>
#> 41                    ISORestriction       <NA>
#> 42                           ISORole       <NA>
#> 43                       ISORoleType       <NA>
#> 44          ISOSRVParameterDirection       <NA>
#> 45                      ISOScopeCode       <NA>
#> 46      ISOSpatialRepresentationType       <NA>
#> 47                  ISOTelephoneType       <NA>
#> 48                  ISOTopicCategory       <NA>
#> 49                  ISOTopologyLevel       <NA>
#> 50                ISOAssociationType       <NA>
#> 51                   ISOCellGeometry       <NA>
#> 52                   ISOCharacterSet       <NA>
#> 53                 ISOClassification       <NA>
#> 54                        ISOCountry       <NA>
#> 55                   ISOCouplingType       <NA>
#> 56            ISOCoverageContentType       <NA>
#> 57                        ISODCPList       <NA>
#> 58                       ISODatatype       <NA>
#> 59                       ISODateType       <NA>
#> 60              ISODimensionNameType       <NA>
#> 61              ISODistributionUnits       <NA>
#> 62           ISOEvaluationMethodType       <NA>
#> 63            ISOGeometricObjectType       <NA>
#> 64          ISOImageryBandDefinition       <NA>
#> 65                 ISOImageryContext       <NA>
#> 66            ISOImageryGeometryType       <NA>
#> 67           ISOImageryObjectiveType       <NA>
#> 68           ISOImageryOperationType       <NA>
#> 69 ISOImageryPolarisationOrientation       <NA>
#> 70                ISOImageryPriority       <NA>
#> 71              ISOImagerySensorType       <NA>
#> 72                ISOImagerySequence       <NA>
#> 73    ISOImageryTransferFunctionType       <NA>
#> 74                 ISOImageryTrigger       <NA>
#> 75               ISOImagingCondition       <NA>
#> 76                 ISOInitiativeType       <NA>
#> 77       ISOInstrumentationEventType       <NA>
#> 78                    ISOKeywordType       <NA>
#> 79                       ISOLanguage  ISO 639-2
#> 80           ISOMaintenanceFrequency       <NA>
#> 81                   ISOMediumFormat       <NA>
#> 82                     ISOMediumName       <NA>
#> 83                     ISOObligation       <NA>
#> 84                 ISOOnLineFunction       <NA>
#> 85             ISOParameterDirection       <NA>
#> 86               ISOPixelOrientation       <NA>
#> 87               ISOPresentationForm       <NA>
#> 88                       ISOProgress       <NA>
#> 89            ISOReferenceSystemType       <NA>
#> 90                    ISORestriction       <NA>
#> 91                           ISORole       <NA>
#> 92                       ISORoleType       <NA>
#> 93          ISOSRVParameterDirection       <NA>
#> 94                      ISOScopeCode       <NA>
#> 95      ISOSpatialRepresentationType       <NA>
#> 96                  ISOTelephoneType       <NA>
#> 97                  ISOTopicCategory       <NA>
#> 98                  ISOTopologyLevel       <NA>
#>                                                                                                                                                                                                                                                                                                                                   description
#> 1                                                                                                                                                                                                                                                                                          justification for the correlation of two resources
#> 2                                                                                                                                                                                                                                                                             code indicating the geometry represented by the grid cell value
#> 3                                                                                                                                                                                                                                                                                  name of the character coding standard used in the resource
#> 4                                                                                                                                                                                                                                                                                           name of the handling restrictions on the resource
#> 5                                                                                                                                                                                                                                                                                                         Country : ISO 3166-2 (2 characters)
#> 6                                                                                                                                                                                                                                                                            type of coupling between service and associated data (if exists)
#> 7                                                                                                                                                                                                                                                                                        specific type of information represented in the cell
#> 8                                                                                                                                                                                                                                                                                class of information to which the referencing entity applies
#> 9                                                                                                                                                                                                                                                                                                               datatype of element or entity
#> 10                                                                                                                                                                                                                                                                                              identification of when a given event occurred
#> 11                                                                                                                                                                                                                                                                                                                      name of the dimension
#> 12                                                                                                                                                                                                                                                                                                                                       <NA>
#> 13                                                                                                                                                                                                                                                                           type or method for evaluating an identified data quality measure
#> 14                                                                                                                                                                                                                    name of point or vector objects used to locate zero-, one-, two-, or three-dimensional spatial locations in the dataset
#> 15                                                                                                                                                                                                                                                  designation of criterion for defining maximum and minimum wavelengths for a spectral band
#> 16                                                                                                                                                                                                                                                            designation of criterion for defining the context of the scanning process event
#> 17                                                                                                                                                                                                                                                                                                    geometric description of the collection
#> 18                                                                                                                                                                                                                                                                                               temporal persistence of collection objective
#> 19                                                                                                                                                                                                                                               code indicating whether the data contained in this packet is real, simulated, or synthesized
#> 20                                                                                                                                                                                                                                                                                       Polarisation of the antenna relative to the waveform
#> 21                                                                                                                                                                                                                                                                                                                 ordered list of priorities
#> 22                                                                                                                                                                                                                                                                                                                                       <NA>
#> 23                                                                                                                                                                                                                                                                                                            temporal relation of activation
#> 24                                                                                                                                                                                                                                                            transform function to be used when scaling a physical value for a given element
#> 25                                                                                                                                                                                                                                                                                                                    mechanism of activation
#> 26                                                                                                                                                                                                                                                                                 code which indicates conditions which may affect the image
#> 27                                                                                                                                                                                                                                                                                type of aggregation activity in which resources are related
#> 28                                                                                                                                                                                                                                                                                                                                       <NA>
#> 29                                                                                                                                                                                                                                                                                                     methods used to group similar keywords
#> 30                                                                                                                                                                                                                                                                                                        Language : ISO 639-2 (3 characters)
#> 31                                                                                                                                                                                                                                           frequency with which modifications and deletions are made to the data after it is first produced
#> 32                                                                                                                                                                                                                                                                                                         method used to write to the medium
#> 33                                                                                                                                                                                                                                                                                                                         name of the medium
#> 34                                                                                                                                                                                                                                                                                                        obligation of the element or entity
#> 35                                                                                                                                                                                                                                                                                                         function performed by the resource
#> 36                                                                                                                                                                                                                                                            identificaties the parameters as an input to the process, or an output, or both
#> 37                                                                                                                                                                                                                                                                          point in a pixel corresponding to the Earth location of the pixel
#> 38                                                                                                                                                                                                                                                                                                     mode in which the data are represented
#> 39                                                                                                                                                                                                                                                                                                                     status of the resource
#> 40                                                                                                                                                                                                                                                                                                      defines type of reference system used
#> 41                                                                                                                                                                                                                                                                                    limitation(s) placed upon the access or use of the data
#> 42                                                                                                                                                                                                                                                                                                function performed by the responsible party
#> 43                                                                                                                                                                                                                                                                                                 code list for the classification of roles 
#> 44                                                                                                                                                                                                                                                                               class of information to which the referencing entity applies
#> 45                                                                                                                                                                                                                                        Extension of MD_ScopeCode for the needs of GMX application schemas and in the context of a transfer
#> 46                                                                                                                                                                                                                                                                            method used to represent geographic information in the resource
#> 47                                                                                                                                                                                                                                                                                                                          type of telephone
#> 48 high-level geographic data thematic classification to assist in the grouping and search of available geographic data sets. Can be used to group keywords as well. Listed examples are not exhaustive.  NOTE: It is understood there are overlaps between general categories and the user is encouraged to select the one most appropriate.
#> 49                                                                                                                                                                                                                                                                                          degree of complexity of the spatial relationships
#> 50                                                                                                                                                                                                                                                                                         justification for the correlation of two resources
#> 51                                                                                                                                                                                                                                                                            code indicating the geometry represented by the grid cell value
#> 52                                                                                                                                                                                                                                                                                 name of the character coding standard used in the resource
#> 53                                                                                                                                                                                                                                                                                          name of the handling restrictions on the resource
#> 54                                                                                                                                                                                                                                                                                                                                       <NA>
#> 55                                                                                                                                                                                                                                                                           type of coupling between service and associated data (if exists)
#> 56                                                                                                                                                                                                                                                                                       specific type of information represented in the cell
#> 57                                                                                                                                                                                                                                                                               class of information to which the referencing entity applies
#> 58                                                                                                                                                                                                                                                                                                              datatype of element or entity
#> 59                                                                                                                                                                                                                                                                                              identification of when a given event occurred
#> 60                                                                                                                                                                                                                                                                                                                      name of the dimension
#> 61                                                                                                                                                                                                                                                                                                                                       <NA>
#> 62                                                                                                                                                                                                                                                                           type or method for evaluating an identified data quality measure
#> 63                                                                                                                                                                                                                    name of point or vector objects used to locate zero-, one-, two-, or three-dimensional spatial locations in the dataset
#> 64                                                                                                                                                                                                                                                  designation of criterion for defining maximum and minimum wavelengths for a spectral band
#> 65                                                                                                                                                                                                                                                            designation of criterion for defining the context of the scanning process event
#> 66                                                                                                                                                                                                                                                                                                    geometric description of the collection
#> 67                                                                                                                                                                                                                                                                                               temporal persistence of collection objective
#> 68                                                                                                                                                                                                                                               code indicating whether the data contained in this packet is real, simulated, or synthesized
#> 69                                                                                                                                                                                                                                                                                       Polarisation of the antenna relative to the waveform
#> 70                                                                                                                                                                                                                                                                                                                 ordered list of priorities
#> 71                                                                                                                                                                                                                                                                                                                                       <NA>
#> 72                                                                                                                                                                                                                                                                                                            temporal relation of activation
#> 73                                                                                                                                                                                                                                                            transform function to be used when scaling a physical value for a given element
#> 74                                                                                                                                                                                                                                                                                                                    mechanism of activation
#> 75                                                                                                                                                                                                                                                                                 code which indicates conditions which may affect the image
#> 76                                                                                                                                                                                                                                                                                type of aggregation activity in which resources are related
#> 77                                                                                                                                                                                                                                                                                                                                       <NA>
#> 78                                                                                                                                                                                                                                                                                                     methods used to group similar keywords
#> 79                                                                                                                                                                                                                                                                                                        Language : ISO 639-2 (3 characters)
#> 80                                                                                                                                                                                                                                           frequency with which modifications and deletions are made to the data after it is first produced
#> 81                                                                                                                                                                                                                                                                                                         method used to write to the medium
#> 82                                                                                                                                                                                                                                                                                                                         name of the medium
#> 83                                                                                                                                                                                                                                                                                                        obligation of the element or entity
#> 84                                                                                                                                                                                                                                                                                                         function performed by the resource
#> 85                                                                                                                                                                                                                                                            identificaties the parameters as an input to the process, or an output, or both
#> 86                                                                                                                                                                                                                                                                          point in a pixel corresponding to the Earth location of the pixel
#> 87                                                                                                                                                                                                                                                                                                     mode in which the data are represented
#> 88                                                                                                                                                                                                                                                                                                                     status of the resource
#> 89                                                                                                                                                                                                                                                                                                      defines type of reference system used
#> 90                                                                                                                                                                                                                                                                                    limitation(s) placed upon the access or use of the data
#> 91                                                                                                                                                                                                                                                                                                function performed by the responsible party
#> 92                                                                                                                                                                                                                                                                                                 code list for the classification of roles 
#> 93                                                                                                                                                                                                                                                                               class of information to which the referencing entity applies
#> 94                                                                                                                                                                                                                                        Extension of MD_ScopeCode for the needs of GMX application schemas and in the context of a transfer
#> 95                                                                                                                                                                                                                                                                            method used to represent geographic information in the resource
#> 96                                                                                                                                                                                                                                                                                                                          type of telephone
#> 97 high-level geographic data thematic classification to assist in the grouping and search of available geographic data sets. Can be used to group keywords as well. Listed examples are not exhaustive.  NOTE: It is understood there are overlaps between general categories and the user is encouraged to select the one most appropriate.
#> 98                                                                                                                                                                                                                                                                                          degree of complexity of the spatial relationships
#>                                     xmlNamespace
#> 1               http://www.isotc211.org/2005/gmd
#> 2               http://www.isotc211.org/2005/gmd
#> 3               http://www.isotc211.org/2005/gmd
#> 4               http://www.isotc211.org/2005/gmd
#> 5               http://www.isotc211.org/2005/gmd
#> 6               http://www.isotc211.org/2005/srv
#> 7               http://www.isotc211.org/2005/gmd
#> 8               http://www.isotc211.org/2005/srv
#> 9               http://www.isotc211.org/2005/gmd
#> 10              http://www.isotc211.org/2005/gmd
#> 11              http://www.isotc211.org/2005/gmd
#> 12              http://www.isotc211.org/2005/gmd
#> 13              http://www.isotc211.org/2005/gmd
#> 14              http://www.isotc211.org/2005/gmd
#> 15 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 16 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 17 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 18 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 19 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 20 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 21 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 22 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 23 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 24 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 25 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 26              http://www.isotc211.org/2005/gmd
#> 27              http://www.isotc211.org/2005/gmd
#> 28 http://standards.iso.org/iso/19115/-2/gmi/1.0
#> 29              http://www.isotc211.org/2005/gmd
#> 30              http://www.isotc211.org/2005/gmd
#> 31              http://www.isotc211.org/2005/gmd
#> 32              http://www.isotc211.org/2005/gmd
#> 33              http://www.isotc211.org/2005/gmd
#> 34              http://www.isotc211.org/2005/gmd
#> 35              http://www.isotc211.org/2005/gmd
#> 36              http://www.isotc211.org/2005/srv
#> 37              http://www.isotc211.org/2005/gmd
#> 38              http://www.isotc211.org/2005/gmd
#> 39              http://www.isotc211.org/2005/gmd
#> 40 http://standards.iso.org/iso/19115/-3/mrs/1.0
#> 41              http://www.isotc211.org/2005/gmd
#> 42              http://www.isotc211.org/2005/gmd
#> 43              http://www.isotc211.org/2005/gfc
#> 44              http://www.isotc211.org/2005/srv
#> 45              http://www.isotc211.org/2005/gmd
#> 46              http://www.isotc211.org/2005/gmd
#> 47 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 48              http://www.isotc211.org/2005/gmd
#> 49              http://www.isotc211.org/2005/gmd
#> 50 http://standards.iso.org/iso/19115/-3/mri/1.0
#> 51 http://standards.iso.org/iso/19115/-3/msr/2.0
#> 52 http://standards.iso.org/iso/19115/-3/lan/1.0
#> 53 http://standards.iso.org/iso/19115/-3/mco/1.0
#> 54 http://standards.iso.org/iso/19115/-3/lan/1.0
#> 55 http://standards.iso.org/iso/19115/-3/srv/2.0
#> 56 http://standards.iso.org/iso/19115/-3/mrc/2.0
#> 57 http://standards.iso.org/iso/19115/-3/srv/2.0
#> 58              http://www.isotc211.org/2005/gmd
#> 59 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 60 http://standards.iso.org/iso/19115/-3/msr/2.0
#> 61              http://www.isotc211.org/2005/gmd
#> 62 http://standards.iso.org/iso/19157/-2/mdq/1.0
#> 63 http://standards.iso.org/iso/19115/-3/msr/2.0
#> 64 http://standards.iso.org/iso/19115/-3/mrc/2.0
#> 65 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 66 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 67 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 68 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 69 http://standards.iso.org/iso/19115/-3/mrc/2.0
#> 70 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 71 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 72 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 73 http://standards.iso.org/iso/19115/-3/mrc/2.0
#> 74 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 75 http://standards.iso.org/iso/19115/-3/mrc/2.0
#> 76 http://standards.iso.org/iso/19115/-3/mri/1.0
#> 77 http://standards.iso.org/iso/19115/-3/mac/2.0
#> 78 http://standards.iso.org/iso/19115/-3/mri/1.0
#> 79 http://standards.iso.org/iso/19115/-3/lan/1.0
#> 80 http://standards.iso.org/iso/19115/-3/mmi/1.0
#> 81 http://standards.iso.org/iso/19115/-3/mrd/1.0
#> 82              http://www.isotc211.org/2005/gmd
#> 83              http://www.isotc211.org/2005/gmd
#> 84 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 85 http://standards.iso.org/iso/19115/-3/mrl/2.0
#> 86 http://standards.iso.org/iso/19115/-3/msr/2.0
#> 87 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 88 http://standards.iso.org/iso/19115/-3/mcc/1.0
#> 89 http://standards.iso.org/iso/19115/-3/mrs/1.0
#> 90 http://standards.iso.org/iso/19115/-3/mco/1.0
#> 91 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 92    http://standards.iso.org/iso/19110/gfc/1.1
#> 93 http://standards.iso.org/iso/19115/-3/srv/2.0
#> 94 http://standards.iso.org/iso/19115/-3/mcc/1.0
#> 95 http://standards.iso.org/iso/19115/-3/mcc/1.0
#> 96 http://standards.iso.org/iso/19115/-3/cit/2.0
#> 97 http://standards.iso.org/iso/19115/-3/mri/1.0
#> 98 http://standards.iso.org/iso/19115/-3/msr/2.0
#>                          xmlElement
#> 1            DS_AssociationTypeCode
#> 2               MD_CellGeometryCode
#> 3               MD_CharacterSetCode
#> 4             MD_ClassificationCode
#> 5                           Country
#> 6                   SV_CouplingType
#> 7        MD_CoverageContentTypeCode
#> 8                           DCPList
#> 9                   MD_DatatypeCode
#> 10                  CI_DateTypeCode
#> 11         MD_DimensionNameTypeCode
#> 12             MD_DistributionUnits
#> 13      DQ_EvaluationMethodTypeCode
#> 14       MD_GeometricObjectTypeCode
#> 15                MI_BandDefinition
#> 16                   MI_ContextCode
#> 17              MI_GeometryTypeCode
#> 18             MI_ObjectiveTypeCode
#> 19             MI_OperationTypeCode
#> 20   MI_PolarisationOrientationCode
#> 21                  MI_PriorityCode
#> 22                MI_SensorTypeCode
#> 23                  MI_SequenceCode
#> 24      MI_TransferFunctionTypeCode
#> 25                   MI_TriggerCode
#> 26          MD_ImagingConditionCode
#> 27            DS_InitiativeTypeCode
#> 28  MI_InstrumentationEventTypeCode
#> 29               MD_KeywordTypeCode
#> 30                     LanguageCode
#> 31      MD_MaintenanceFrequencyCode
#> 32              MD_MediumFormatCode
#> 33                MD_MediumNameCode
#> 34                MD_ObligationCode
#> 35            CI_OnLineFunctionCode
#> 36            LE_ParameterDirection
#> 37          MD_PixelOrientationCode
#> 38          CI_PresentationFormCode
#> 39                  MD_ProgressCode
#> 40       MD_ReferenceSystemTypeCode
#> 41               MD_RestrictionCode
#> 42                      CI_RoleCode
#> 43                      FC_RoleType
#> 44            SV_ParameterDirection
#> 45                     MX_ScopeCode
#> 46 MD_SpatialRepresentationTypeCode
#> 47             CI_TelephoneTypeCode
#> 48             MD_TopicCategoryCode
#> 49             MD_TopologyLevelCode
#> 50           DS_AssociationTypeCode
#> 51              MD_CellGeometryCode
#> 52              MD_CharacterSetCode
#> 53            MD_ClassificationCode
#> 54                      CountryCode
#> 55                  SV_CouplingType
#> 56       MD_CoverageContentTypeCode
#> 57                          DCPList
#> 58                  MD_DatatypeCode
#> 59                  CI_DateTypeCode
#> 60         MD_DimensionNameTypeCode
#> 61             MD_DistributionUnits
#> 62      DQ_EvaluationMethodTypeCode
#> 63       MD_GeometricObjectTypeCode
#> 64                MI_BandDefinition
#> 65                   MI_ContextCode
#> 66              MI_GeometryTypeCode
#> 67             MI_ObjectiveTypeCode
#> 68             MI_OperationTypeCode
#> 69   MI_PolarisationOrientationCode
#> 70                  MI_PriorityCode
#> 71                MI_SensorTypeCode
#> 72                  MI_SequenceCode
#> 73      MI_TransferFunctionTypeCode
#> 74                   MI_TriggerCode
#> 75          MD_ImagingConditionCode
#> 76            DS_InitiativeTypeCode
#> 77  MI_InstrumentationEventTypeCode
#> 78               MD_KeywordTypeCode
#> 79                     LanguageCode
#> 80      MD_MaintenanceFrequencyCode
#> 81              MD_MediumFormatCode
#> 82                MD_MediumNameCode
#> 83                MD_ObligationCode
#> 84            CI_OnLineFunctionCode
#> 85            LE_ParameterDirection
#> 86          MD_PixelOrientationCode
#> 87          CI_PresentationFormCode
#> 88                  MD_ProgressCode
#> 89       MD_ReferenceSystemTypeCode
#> 90               MD_RestrictionCode
#> 91                      CI_RoleCode
#> 92                      FC_RoleType
#> 93            SV_ParameterDirection
#> 94                     MX_ScopeCode
#> 95 MD_SpatialRepresentationTypeCode
#> 96             CI_TelephoneTypeCode
#> 97             MD_TopicCategoryCode
#> 98             MD_TopologyLevelCode