getISOCodelists

getISOCodelists allows to get the list of ISO codelists registered in geometa, their description and XML definition. The object returned is of class "data.frame"

getISOCodelists()

Author

Emmanuel Blondel, emmanuel.blondel1@gmail.com

Examples

            
  getISOCodelists()
#>                            classname        codeSpace
#> 1                 ISOAssociationType   ISOTC211/19115
#> 2                    ISOCellGeometry   ISOTC211/19115
#> 3                    ISOCharacterSet   ISOTC211/19115
#> 4                  ISOClassification   ISOTC211/19115
#> 5                         ISOCountry       ISO 3166-2
#> 6                    ISOCouplingType   ISOTC211/19119
#> 7             ISOCoverageContentType   ISOTC211/19115
#> 8                         ISODCPList   ISOTC211/19119
#> 9                        ISODatatype   ISOTC211/19115
#> 10                       ISODateType   ISOTC211/19115
#> 11              ISODimensionNameType   ISOTC211/19115
#> 12              ISODistributionUnits             <NA>
#> 13           ISOEvaluationMethodType   ISOTC211/19115
#> 14            ISOGeometricObjectType   ISOTC211/19115
#> 15                 ISOHierarchyLevel   ISOTC211/19115
#> 16          ISOImageryBandDefinition ISOTC211/19115-2
#> 17                 ISOImageryContext ISOTC211/19115-2
#> 18            ISOImageryGeometryType ISOTC211/19115-2
#> 19           ISOImageryObjectiveType ISOTC211/19115-2
#> 20           ISOImageryOperationType ISOTC211/19115-2
#> 21 ISOImageryPolarisationOrientation ISOTC211/19115-2
#> 22                ISOImageryPriority ISOTC211/19115-2
#> 23                ISOImagerySequence ISOTC211/19115-2
#> 24    ISOImageryTransferFunctionType ISOTC211/19115-2
#> 25                 ISOImageryTrigger ISOTC211/19115-2
#> 26               ISOImagingCondition   ISOTC211/19115
#> 27                 ISOInitiativeType   ISOTC211/19115
#> 28                    ISOKeywordType   ISOTC211/19115
#> 29                       ISOLanguage        ISO 639-2
#> 30           ISOMaintenanceFrequency   ISOTC211/19115
#> 31                   ISOMediumFormat   ISOTC211/19115
#> 32                     ISOMediumName   ISOTC211/19115
#> 33                     ISOObligation   ISOTC211/19115
#> 34                 ISOOnLineFunction   ISOTC211/19115
#> 35             ISOParameterDirection   ISOTC211/19119
#> 36               ISOPixelOrientation   ISOTC211/19115
#> 37               ISOPresentationForm   ISOTC211/19115
#> 38                    ISORestriction   ISOTC211/19115
#> 39                           ISORole   ISOTC211/19115
#> 40                       ISORoleType   ISOTC211/19110
#> 41      ISOSpatialRepresentationType   ISOTC211/19115
#> 42                         ISOStatus   ISOTC211/19115
#> 43                  ISOTopicCategory   ISOTC211/19115
#> 44                  ISOTopologyLevel   ISOTC211/19115
#>                                                                                                                                                                                              description
#> 1                                                                                                                                                      justification for the correlation of two datasets
#> 2                                                                                                                                                     code indicating whether grid data is point or area
#> 3                                                                                                                                             name of the character coding standard used in the resource
#> 4                                                                                                                                                       name of the handling restrictions on the dataset
#> 5                                                                                                                                                                    Country : ISO 3166-2 (2 characters)
#> 6                                                                                                                                                                indication of service instance coupling
#> 7                                                                                                                                                   specific type of information represented in the cell
#> 8                                                                                                                            distributed computing platforms on which the operation has been implemented
#> 9                                                                                                                                                                          datatype of element or entity
#> 10                                                                                                                                                         identification of when a given event occurred
#> 11                                                                                                                                                                                 name of the dimension
#> 12                                                                                                                                                                                                  <NA>
#> 13                                                                                                                                      type or method for evaluating an identified data quality measure
#> 14                                                                               name of point or vector objects used to locate zero-, one-, two-, or three-dimensional spatial locations in the dataset
#> 15                                                                                                   Extension of MD_ScopeCode for the needs of GMX application schemas and in the context of a transfer
#> 16                                                                                                             designation of criterion for defining maximum and minimum wavelengths for a spectral band
#> 17                                                                                                                       designation of criterion for defining the context of the scanning process event
#> 18                                                                                                                                                               geometric description of the collection
#> 19                                                                                                                                                          temporal persistence of collection objective
#> 20                                                                                                          code indicating whether the data contained in this packet is real, simulated, or synthesized
#> 21                                                                                                                                                  Polarisation of the antenna relative to the waveform
#> 22                                                                                                                                                                            ordered list of priorities
#> 23                                                                                                                                                                       temporal relation of activation
#> 24                                                                                                                       transform function to be used when scaling a physical value for a given element
#> 25                                                                                                                                                                               mechanism of activation
#> 26                                                                                                                                            code which indicates conditions which may affect the image
#> 27                                                                                                                                            type of aggregation activity in which datasets are related
#> 28                                                                                                                                                                methods used to group similar keywords
#> 29                                                                                                                                                                   Language : ISO 639-2 (3 characters)
#> 30                                                                                                      frequency with which modifications and deletions are made to the data after it is first produced
#> 31                                                                                                                                                                    method used to write to the medium
#> 32                                                                                                                                                                                    name of the medium
#> 33                                                                                                                                                                   obligation of the element or entity
#> 34                                                                                                                                                                    function performed by the resource
#> 35                                                                                                                             indication if the parameter is an input to the service, an output or both
#> 36                                                                                                                                     point in a pixel corresponding to the Earth location of the pixel
#> 37                                                                                                                                                                 mode in which the data is represented
#> 38                                                                                                                                               limitation(s) placed upon the access or use of the data
#> 39                                                                                                                                                           function performed by the responsible party
#> 40                                                                                                                                                            code list for the classification of roles 
#> 41                                                                                                                                        method used to represent geographic information in the dataset
#> 42                                                                                                                                                         status of the dataset or progress of a review
#> 43 high-level geographic data thematic classification to assist in the grouping and search of available geographic data sets. Can be used to group keywords as well. Listed examples are not exhaustive.
#> 44                                                                                                                                                     degree of complexity of the spatial relationships
#>                        xmlNamespace                       xmlElement
#> 1  http://www.isotc211.org/2005/gmd           DS_AssociationTypeCode
#> 2  http://www.isotc211.org/2005/gmd              MD_CellGeometryCode
#> 3  http://www.isotc211.org/2005/gmd              MD_CharacterSetCode
#> 4  http://www.isotc211.org/2005/gmd            MD_ClassificationCode
#> 5  http://www.isotc211.org/2005/gmd                          Country
#> 6  http://www.isotc211.org/2005/srv                  SV_CouplingType
#> 7  http://www.isotc211.org/2005/gmd       MD_CoverageContentTypeCode
#> 8  http://www.isotc211.org/2005/srv                          DCPList
#> 9  http://www.isotc211.org/2005/gmd                  MD_DatatypeCode
#> 10 http://www.isotc211.org/2005/gmd                  CI_DateTypeCode
#> 11 http://www.isotc211.org/2005/gmd         MD_DimensionNameTypeCode
#> 12 http://www.isotc211.org/2005/gmd             MD_DistributionUnits
#> 13 http://www.isotc211.org/2005/gmd      DQ_EvaluationMethodTypeCode
#> 14 http://www.isotc211.org/2005/gmd       MD_GeometricObjectTypeCode
#> 15 http://www.isotc211.org/2005/gmd                     MX_ScopeCode
#> 16 http://www.isotc211.org/2005/gmi                MI_BandDefinition
#> 17 http://www.isotc211.org/2005/gmi                   MI_ContextCode
#> 18 http://www.isotc211.org/2005/gmi              MI_GeometryTypeCode
#> 19 http://www.isotc211.org/2005/gmi             MI_ObjectiveTypeCode
#> 20 http://www.isotc211.org/2005/gmi             MI_OperationTypeCode
#> 21 http://www.isotc211.org/2005/gmi   MI_PolarisationOrientationCode
#> 22 http://www.isotc211.org/2005/gmi                  MI_PriorityCode
#> 23 http://www.isotc211.org/2005/gmi                  MI_SequenceCode
#> 24 http://www.isotc211.org/2005/gmi      MI_TransferFunctionTypeCode
#> 25 http://www.isotc211.org/2005/gmi                   MI_TriggerCode
#> 26 http://www.isotc211.org/2005/gmd          MD_ImagingConditionCode
#> 27 http://www.isotc211.org/2005/gmd            DS_InitiativeTypeCode
#> 28 http://www.isotc211.org/2005/gmd               MD_KeywordTypeCode
#> 29 http://www.isotc211.org/2005/gmd                     LanguageCode
#> 30 http://www.isotc211.org/2005/gmd      MD_MaintenanceFrequencyCode
#> 31 http://www.isotc211.org/2005/gmd              MD_MediumFormatCode
#> 32 http://www.isotc211.org/2005/gmd                MD_MediumNameCode
#> 33 http://www.isotc211.org/2005/gmd                MD_ObligationCode
#> 34 http://www.isotc211.org/2005/gmd            CI_OnLineFunctionCode
#> 35 http://www.isotc211.org/2005/srv            SV_ParameterDirection
#> 36 http://www.isotc211.org/2005/gmd          MD_PixelOrientationCode
#> 37 http://www.isotc211.org/2005/gmd          CI_PresentationFormCode
#> 38 http://www.isotc211.org/2005/gmd               MD_RestrictionCode
#> 39 http://www.isotc211.org/2005/gmd                      CI_RoleCode
#> 40 http://www.isotc211.org/2005/gfc                      FC_RoleType
#> 41 http://www.isotc211.org/2005/gmd MD_SpatialRepresentationTypeCode
#> 42 http://www.isotc211.org/2005/gmd                  MD_ProgressCode
#> 43 http://www.isotc211.org/2005/gmd             MD_TopicCategoryCode
#> 44 http://www.isotc211.org/2005/gmd             MD_TopologyLevelCode